### 2024-08-22 #7136 @editorialbot
[PRE REVIEW]: DVMDOSTEM: a terrestrial ecosystem model designed to represent arctic, boreal and permafrost ecosystem dynamics
**Submitting author:** @tobeycarman (<a href="http://orcid.org/0000-0003-4617-4674">Tobey Carman</a>)
**Repository:** https://github.com/uaf-arctic-eco-modeling/dvm-dos-tem
**Branch with paper.md** (empty if default branch): joss-paper-2
**Version:** v0.8.1
