### 2024-08-23 #7136 @tobeycarman
@kthyng Thanks! re: the above issues:
 - license, looks OK
 - install instructions, see [QuickStart](https://github.com/uaf-arctic-eco-modeling/dvm-dos-tem#quickstart) in the README, (requires some comfort with Docker); much more info in the User Guide, see especially the [Dev Info section](https://uaf-arctic-eco-modeling.github.io/dvm-dos-tem/software_development_info.html) and the [Basic Model Setup and Run](https://uaf-arctic-eco-modeling.github.io/dvm-dos-tem/examples_and_tutorials/basic_model_run.html)
 - Documentation - never complete, but see the [User Guide](https://uaf-arctic-eco-modeling.github.io/dvm-dos-tem/index.html)
 - Tests of some sort - never complete, but see the section in the User Guide about [testing and deployment ](https://uaf-arctic-eco-modeling.github.io/dvm-dos-tem/software_development_info.html#testing-and-deployment)
 - Check paper - we are over the recommended word count, but feel it is justified given the size of the project. I was able to compile the paper but was not able to get the ORCID icons to work for all the authors.
 - Substantial scholarly effort - see 30 year history.
 - Clear research application - see "statement of need" in paper

One note - I just discovered an issue with Git LFS which we have been using to manage testing data. I will try to address this ASAP, but if you or any other editors run into issues with cloning or pulling from the repo, please let me know.
