### 2024-08-26 c5b3787
get parameter ranges function
