### 2024-08-26 8f55ab7
Fix degSecNeeded missing dz, scale using VWC, remove comments
