+++ src/Stefan.cpp
@@ -32,11 +32,8 @@ void Stefan::updateFronts(const double & tdrv, const double &timestep) {
   // driving force
   int freezing1; //the freezing/thawing force based on the driving temperature
   double tdrv1 = tdrv;
-  //20180820 While comparing the code to the papers describing these
-  // processes, it seemed like tdrv1 needed to be converted from
-  // degrees C to degrees K. However, we tested that change, and
-  // it appears that not converting it produces better results.
-  double dse = fabs(tdrv1 * timestep); // the extra degree second >>> try make this Ks
+  // extra degree second - Yi et al. 2009
+  double dse = fabs(tdrv1 * timestep); // the extra degree second
   double sumresabv  =0. ; // sum of resistence for above layers;
 
   if(tdrv1>0.0) {
@@ -179,11 +176,9 @@ void Stefan::processNewFrontSoilLayerDown(const int &freezing,
     dz *= fmax(0., 1.0-sl->frozenfrac); //assuming frozen/unfrozen
                                         //  soil segments not mixed
     volwat = fmax(0.0, sl->getVolLiq())*sl->dz;
-    // >>> volwat = sl->getUnfVolLiq() * sl->dz;
   } else {
     dz *= sl->frozenfrac; //assuming frozen/unfrozen soil segments not mixed
     volwat = fmax(0.0, sl->getVolIce())*sl->dz;
-    // >>> volwat = (1 - sl->getUnfVolLiq()) * sl->dz;
   }
 
   if (dz<=0.0001*sl->dz) { //this will avoid 'front' exactly falling on the
@@ -192,7 +187,7 @@ void Stefan::processNewFrontSoilLayerDown(const int &freezing,
     dz = 0.0001*sl->dz;
   }
 
-  dsn = getDegSecNeeded(dz, volwat, tkfront, sumrescum);// >>> try doubling/halving dsn here
+  dsn = getDegSecNeeded(dz, volwat, tkfront, sumrescum);
   if(sl->tem < 0){ //if layer below 0, include energy needed to bring layer to 0 before thawing
     dsn += abs(sl->tem)*timestep;
   }
@@ -359,11 +354,9 @@ void Stefan::processNewFrontSoilLayerUp(const int &freezing,
     dz *= fmax(0., 1.0-sl->frozenfrac); //assuming frozen/unfrozen
                                         //  soil segments not mixed
     volwat = fmax(0.0, sl->getVolLiq())*sl->dz;
-    // >>>volwat = sl->getUnfVolLiq() * sl->dz;
   } else {
     dz *= sl->frozenfrac; //assuming frozen/unfrozen soil segments not mixed
     volwat = fmax(0.0, sl->getVolIce())*sl->dz;
-    // >>>volwat = (1 - sl->getUnfVolLiq()) * sl->dz;
   }
 
   if (dz<=0.0001*sl->dz) { //this will avoid 'front' exactly falling on the
@@ -537,8 +530,8 @@ double Stefan::getDegSecNeeded(const double & dz, const double & volwat,
   double needed=0.;
   double effvolwat = volwat;
   double lhfv = LHFUS * 1000;//Converting units
-  needed = lhfv * effvolwat * (sumresabv + 0.5 * dz/tk);
-  return needed; // >>> brute forcings here potentially
+  needed = lhfv * effvolwat * dz * (sumresabv + 0.5 * dz/tk); // need dz outside of parentheses too
+  return needed;
 };
 
 //calculate partial depth based on extra degree seconds
@@ -729,12 +722,12 @@ void Stefan::updateLayerFrozenState(Layer * toplayer, const int freezing1) {
       } else { // no front at all
         if (currl->frozen==0 && currl->isSoil){ //suggests that this was a front layer but fronts have been swept out
[…]
