### 2024-08-27 aa8a011
Adding sum_pftpart functionality to plot eq relationships
