+++ mads_calibration/SA_post_hoc_analysis.py
@@ -1085,7 +1085,7 @@ def n_top_runs(results, targets, params, r2lim, N=None):
 
   return best_params, best_results
 
-def plot_equilibrium_relationships(path='', plot_metrics=False, save=False, saveprefix=''):
+def plot_equilibrium_relationships(path='', sum_pftpart=False, save=False, saveprefix=''):
   '''
   Plots equilibrium timeseries for target variables in output directory
   
@@ -1093,9 +1093,9 @@ def plot_equilibrium_relationships(path='', plot_metrics=False, save=False, save
   ==========
   path : str
     specifies path to sensitivity sample run directory
-  plot_metrics : bool
-    plot slope, coefficient of variation and pvalue in 30 yr 
-    increments, note this takes a long time
+  sumpftpart : bool
+    flag whether to sum results and targets across compartments
+    to compare pft totals
   save : bool
     saves figure if True
   saveprefix : string
@@ -1107,11 +1107,7 @@ def plot_equilibrium_relationships(path='', plot_metrics=False, save=False, save
   .. image:: /images/SA_post_hoc_analysis/eq_rel_plot.png
   
   '''
-  if plot_metrics:
-    print("plot_metrics enabled, this may take several minutes")
-    print('|',end="")
-    progress = '='
-      
+     
   # defining list of strings for compartment reference
   comp_ref = ['Leaf', 'Stem', 'Root']
   
@@ -1136,39 +1132,22 @@ def plot_equilibrium_relationships(path='', plot_metrics=False, save=False, save
     # operations depending on variable, pft, compartment
     if len(targ_var_info[0]) == 1:
       fig, ax = plt.subplots()
-      if plot_metrics:
-        fig_slope, ax_slope = plt.subplots()
-        fig_cv, ax_cv = plt.subplots()
-        fig_conv, ax_conv = plt.subplots()
     if len(targ_var_info[0]) == 2:
       fig, ax = plt.subplots(1, len(targ_var_info))
-      if plot_metrics:
-        fig_slope, ax_slope = plt.subplots(1, len(targ_var_info))
-        fig_cv, ax_cv = plt.subplots(1, len(targ_var_info))
-        fig_conv, ax_conv = plt.subplots(1, len(targ_var_info))
-
     if len(targ_var_info[0]) == 3:
-      fig, ax = plt.subplots(3, len(np.unique(np.asarray(targ_var_info)[:,1])))
-      if plot_metrics:
-        fig_slope, ax_slope = plt.subplots(3, len(np.unique(np.asarray(targ_var_info)[:,1])))
-        fig_cv, ax_cv = plt.subplots(3, len(np.unique(np.asarray(targ_var_info)[:,1])))
-        fig_conv, ax_conv = plt.subplots(3, len(np.unique(np.asarray(targ_var_info)[:,1])))
-
+      if sum_pftpart==False:
+        fig, ax = plt.subplots(3, len(np.unique(np.asarray(targ_var_info)[:,1])))
+      elif sum_pftpart:
+        fig, ax = plt.subplots(1, len(np.unique(np.asarray(targ_var_info)[:,1])))
+        
     # filtering for directories containing the name sample 
     samples = np.sort([name for name in os.listdir(path) if os.path.isdir(path+name) and "sample" in name])
     
     # looping through sample folders in directory:
     for n, sample in enumerate(samples):
 
-      if plot_metrics:
-        print(progress, end="")
-
       # reading output variable for each sample
[…]
