### 2024-08-28 c5b7b4b
Adding plot_eq_relationships for MADS output
