+++ mads_calibration/SA_post_hoc_analysis.py
@@ -1139,8 +1139,8 @@ def plot_equilibrium_relationships(path='', sum_pftpart=False, save=False, savep
         fig, ax = plt.subplots(3, len(np.unique(np.asarray(targ_var_info)[:,1])))
       elif sum_pftpart:
         fig, ax = plt.subplots(1, len(np.unique(np.asarray(targ_var_info)[:,1])))
-        
-    # filtering for directories containing the name sample 
+
+    # filtering for directories containing the name sample - SA
     samples = np.sort([name for name in os.listdir(path) if os.path.isdir(path+name) and "sample" in name])
     
     # looping through sample folders in directory:
@@ -1240,6 +1240,163 @@ def plot_equilibrium_relationships(path='', sum_pftpart=False, save=False, savep
           
   if save:
     plt.savefig(saveprefix + f"{targ}_eq_rel_plot.png", bbox_inches='tight')
+
+def plot_mads_relationships(targets, path='', sum_pftpart=False, save=False, saveprefix=''):
+  '''
+  Plots equilibrium timeseries for target variables in mads output directory
+  
+  Parameters
+  ==========
+  targets : pandas.DataFrame
+    target values (used in sensitivity analysis step) for indexing
+    and comparing to final mads optimization
+  path : str
+    specifies path to mads run directory
+  sumpftpart : bool
+    flag whether to sum results and targets across compartments
+    to compare pft totals
+  save : bool
+    saves figure if True
+  saveprefix : string
+    path to use if saving is enabled
+  
+  Returns
+    None
+  
+  .. image:: /images/SA_post_hoc_analysis/eq_rel_plot.png
+  
+  '''
+     
+  # defining list of strings for compartment reference
+  comp_ref = ['Leaf', 'Stem', 'Root']
+    
+  # splitting column names to provide variable, pft, and compartment if available
+  targ_info = [i.split('_') for i in targets.columns]
+  
+  # returning only unique variable names for file selection
+  targ_vars = np.unique([i.split('_')[0] for i in targets.columns])
+
+  # looping through each target variable
+  for targ in targ_vars:
+    
+    # filtering for said target variable
+    targ_filter = [targ in info for info in targ_info]
+
+    # returning filtered information on variable, pft, compartment
+    targ_var_info = [i for indx,i in enumerate(targ_info) if targ_filter[indx] == True]
+
+    # operations depending on variable, pft, compartment
+    if len(targ_var_info[0]) == 1:
+      fig, ax = plt.subplots()
+    if len(targ_var_info[0]) == 2:
+      fig, ax = plt.subplots(1, len(targ_var_info))
+    if len(targ_var_info[0]) == 3:
+      if sum_pftpart==False:
+        fig, ax = plt.subplots(3, len(np.unique(np.asarray(targ_var_info)[:,1])))
+      elif sum_pftpart:
+        fig, ax = plt.subplots(1, len(np.unique(np.asarray(targ_var_info)[:,1])))
+
+    # There should only be an output folder in this case 
+    # (adpated from plot_equilibrium_relationships)
+    samples = ['']
+    
+    # looping through sample folders in directory:
+    for n, sample in enumerate(samples):
+
+      # reading output variable for each sample
+      output = nc.Dataset(path+sample+f'output/{targ}_yearly_eq.nc').variables[targ][:].data
+
[…]
