### 2024-08-28 #7136 @kthyng
@tobeycarman Some questions/comments:
* are there installation instructions that do not rely on docker? While docker images may get users up and running, another part of JOSS software is that users would want to contribute to your software, and they will need to be able to develop which will usually mean build the software not using docker images.
* Reviewers will most likely comment on empty sections in your comments and ask you to fill these in. Now is a good time to do this.
* It looks like you did a lot of updates in 2022 — publishing with JOSS is a great reason to update to present day!
* Indeed this is a large code base, but I disagree that that justifies an overly long paper. Please shorten it and aim for 1000 words.
