### 2024-08-29 #7136 @tobeycarman
@kthyng,
 - re: Docker, we don't actively maintain install instructions without Docker. We found that using Docker images and containers to standardize the development environments dramatically improved the development process.
 - re: Empty sections in the comments, I am not sure what you are referring to?
 - re: Word count, I will look into reducing it. We would gladly accept feedback on which sections reviewers find to be less relevant.
