### 2024-08-29 281f825
Strip several words, fix a few typos.
