### 2024-08-30 fffd9b3
Implements pause indicator file for the DA process
This change introduces a basic text file that will
be created in the same directory as the DA statefile
while the model is paused for data assimilation to run.
It contains the current pause stage, year, and month,
although that data is not the purpose of the file.
