### 2024-09-03 718eb1a
Fix saving individual figures and sizing
