### 2024-09-04 f135c35
update CPCRW fire exp figs
