### 2024-09-04 8c312af
Starting copy of wildfire seismic line disturbance
