+++ src/SeismicLine.cpp
@@ -1,0 +1,684 @@
+/*
+ * Modified from WildFire.cpp
+ *
+ * This is for wildfire occurrence and C/N pool dynamics due to fire
+ *   (1) Fire occurrence either by input or FRI derived
+ *   (2) Fire severity either by ALFRESCO input, or determined by
+ *         'size','season' and 'landscape position'
+ *
+ *   (3) C/N pools updated here through operation on 'bd' -
+ *         NOTE this 'bd' is for all PFTs and soil.
+ *         this is to say, individual PFT's C/N pools updating need to
+ *         do elsewhere
+ *
+ *   (4) ROOTFRAC is updated here through operation on 'ed' - also for ALL PFTS.
+ *
+ *   (5) soil layer structure is NOT changed here, but will re-do in
+ *         Ground::adjustSoilAfterBurn()
+ *
+ *   (6) NOTE: assuming ONE root pool, and its index is 'NUM_PFT_PART'
+ *         (i.e. the last).
+ *
+ */
+
+#include <string>
+#include <sstream>
+
+#include "../include/WildFire.h"
+
+#include "../include/TEMUtilityFunctions.h"
+#include "../include/TEMLogger.h"
+
+extern src::severity_logger< severity_level > glg;
+
+WildFire::WildFire() {}
+
+WildFire::~WildFire() {}
+
+WildFire::WildFire(const std::string& fri_fname,
+                   const std::string& exp_fname, const double cell_slope,
+                   const double cell_aspect, const double cell_elevation,
+                   const int y, const int x) {
+
+  #pragma omp critical(load_input)
+  {
+    BOOST_LOG_SEV(glg, info) << "Setting up FRI fire data...";
+    this->fri = temutil::get_scalar<int>(fri_fname, "fri", y, x);
+    this->fri_severity = temutil::get_scalar<int>(fri_fname, "fri_severity", y, x);
+    this->fri_jday_of_burn = temutil::get_scalar<int>(fri_fname, "fri_jday_of_burn", y, x);
+    this->fri_area_of_burn = temutil::get_scalar<int>(fri_fname, "fri_area_of_burn", y, x);
+  }//End critical(fri)
+
+  #pragma omp critical(load_input)
+  {
+    BOOST_LOG_SEV(glg, info) << "Setting up explicit fire data...";
+    this->exp_burn_mask = temutil::get_timeseries<int>(exp_fname, "exp_burn_mask", y, x);
+    this->exp_fire_severity = temutil::get_timeseries<int>(exp_fname, "exp_fire_severity", y, x);
+    this->exp_jday_of_burn = temutil::get_timeseries<int>(exp_fname, "exp_jday_of_burn", y, x);
+    this->exp_area_of_burn = temutil::get_timeseries<int64_t>(exp_fname, "exp_area_of_burn", y, x);
+  }//End critical(exp_fir) 
+
+  this->slope = cell_slope;
+  this->asp = cell_aspect;
+  this->elev = cell_elevation;
+
+  BOOST_LOG_SEV(glg, debug) << "Done making WildFire object.";
+  BOOST_LOG_SEV(glg, debug) << this->report_fire_inputs();
+
+}
+
+void WildFire::load_projected_explicit_data(const std::string& exp_fname, int y, int x) {
+    BOOST_LOG_SEV(glg, info) << "Setting up explicit fire data...";
+    this->exp_burn_mask = temutil::get_timeseries<int>(exp_fname, "exp_burn_mask", y, x);
[…]
