### 2024-09-05 435a9fc
Implements holding for AVLN
