### 2024-09-05 20af849
adding pft params to get_max_param_ranges
