### 2024-09-05 9e4c8a5
Adding relative error to calc_metrics
