### 2024-09-05 4f3434e
Adding information to doc string calc_metrics
