### 2024-09-06 478331d
Fixing calc_metrics instances
