+++ calibration/calibration_targets.py
@@ -28,6 +28,7 @@ calibration_targets = {
     'CarbonMineralSum': {'units': 'g/m2', 'desc': '', 'comment': 'minec', 'ref': '', 'ncname': 'MINEC'},
     'OrganicNitrogenSum': {'units': '', 'desc': '', 'comment': 'soln', 'ref': '', 'ncname': 'ORGN'},
     'AvailableNitrogenSum': {'units': '', 'desc': '', 'comment': 'avln', 'ref': '', 'ncname': 'AVLN'},
+    'CH4All': {'units': '', 'desc': '', 'comment': 'ch4all', 'ref': '', 'ncname': 'CH4EFFLUXTOT'},
   },
 
   ## WARNING: JUNK, PLACEHOLDER VALUES! USE AT YOUR OWN RISK!
@@ -54,6 +55,7 @@ calibration_targets = {
     'CarbonMineralSum':          0.00,    #  minec
     'OrganicNitrogenSum':        0.00,    #  soln
     'AvailableNitrogenSum':      0.00,    #  avln
+    'CH4All':      0.00,    #  ch4all
   },
   ## CMT01 - Black Spruce Forest, calibration for Murphy Dome climate.
   "black spruce forest": {
@@ -80,6 +82,7 @@ calibration_targets = {
     'CarbonMineralSum':      41665.00,    #  minec
     'OrganicNitrogenSum':     2145.87,    #  soln
     'AvailableNitrogenSum':      0.76,    #  avln
+    'CH4All':      0.00,    #  ch4all
   },
   ## CMT02 - White Spruce Forest, calibration for Murphy Dome climate.
   "white spruce forest": {
@@ -106,6 +109,7 @@ calibration_targets = {
     'CarbonMineralSum':      11005.00,    #  minec
     'OrganicNitrogenSum':      699.81,    #  soln
     'AvailableNitrogenSum':      1.69,    #  avln
+    'CH4All':      0.00,    #  ch4all
   },
   ## CMT03 - Deciduous Forest, calibration for Murphy Dome climate.
   "deciduous forest": {
@@ -132,6 +136,7 @@ calibration_targets = {
     'CarbonMineralSum':      25915.00,    #  minec
     'OrganicNitrogenSum':     1513.10,    #  soln
     'AvailableNitrogenSum':       3.5,    #  avln
+    'CH4All':      0.00,    #  ch4all
   },
   ## CMT04 - Shrub Tundra, calibration for Toolik climate.
   "shrub tundra": {
@@ -158,6 +163,7 @@ calibration_targets = {
     'CarbonMineralSum':      37022.00,    #  minec
     'OrganicNitrogenSum':     1843.00,    #  soln
     'AvailableNitrogenSum':      3.93,    #  avln
+    'CH4All':      0.00,    #  ch4all
   },
   ## CMT05 Tussock Tundra (updated 2/23/2016, JDC: GPPain, NPPain, NPPall, Nuptake, Veg from Shaver & Chapin 1991; 
   ## Assume Toolik C:N same as Council; then Veg N = Shaver & Chapin Veg C * (Council N / Council C); MossDeathC from CTucker data;
@@ -188,6 +194,7 @@ calibration_targets = {
     'CarbonMineralSum':      43404.00,    #  minec
     'OrganicNitrogenSum':     2206.00,    #  soln
     'AvailableNitrogenSum':      8.958,   #  avln
+    'CH4All':      0.00,    #  ch4all
   },
   ## CMT06 - WETSEDGE TUNDRA - CALIBRATION WITH toolik climate (also barrow climate)  Lichen gpp was 0.375, npp 0.187, feather 8.4 sphg 2.9, 1.45
   "wet sedge tundra": {
[…]
