+++ parameters/cmt_calparbgc.txt
@@ -29,8 +29,12 @@
 0.00      // kdcsomach4:
 0.00      // kdcsomprch4:
 0.00      // kdcsomcrch4:
-0.0     // ch4ebulrate: ebullition rate, per hour
-0.0     // ch4transportrate: plant-mediated transport rate, per hour
+0.00      // ch4ebulrate: ebullition rate, per hour
+0.00      // ch4transportrate: plant-mediated transport rate, per hour
+4.00      // prodq10ch4:
+2.20      // oxidq10ch4:
+5.00      // oxidkmch4:
+20.0      // oxidvmaxch4:
 //===========================================================
 // CMT01 // Boreal Black Spruce // 6/29/20 boreal black spruce with Murphy Dome climate
 //EverTree  DecidShrub   DecidTree     Moss         PFT4       PFT5      PFT6     PFT7       PFT8     PFT9       // pftnames:
@@ -59,6 +63,10 @@
 0.00      // kdcsomcrch4:
 0.0     // ch4ebulrate: ebullition rate, per hour
 0.0     // ch4transportrate: plant-mediated transport rate, per hour
+4.00      // prodq10ch4:
+2.20      // oxidq10ch4:
+5.00      // oxidkmch4:
+20.0      // oxidvmaxch4:
 //===========================================================
 // CMT02 // Boreal White Spruce Forest // 1/22/21 boreal white spruce with Bonanza Creek climate
 //EverTree  DecidShrub   DecidTree     Moss         PFT4       PFT5      PFT6     PFT7       PFT8     PFT9       // pftnames:
@@ -87,6 +95,10 @@
 0.00      // kdcsomcrch4:
 0.0     // ch4ebulrate: ebullition rate, per hour
 0.0     // ch4transportrate: plant-mediated transport rate, per hour
+4.00      // prodq10ch4:
+2.20      // oxidq10ch4:
+5.00      // oxidkmch4:
+20.0      // oxidvmaxch4:
 //===========================================================
 // CMT03 // Boreal Deciduous Forest // 7/7/20 boreal deciduous forest with Murphy Dome climate
 //EverTree  DecidShrub   DecidTree      Moss        PFT4     PFT5    PFT6    PFT7     PFT8     PFT9    // pftnames:
@@ -115,6 +127,10 @@
 0.00      // kdcsomcrch4:
 0.0     // ch4ebulrate: ebullition rate, per hour
 0.0     // ch4transportrate: plant-mediated transport rate, per hour
+4.00      // prodq10ch4:
+2.20      // oxidq10ch4:
+5.00      // oxidkmch4:
+20.0      // oxidvmaxch4:
 //===========================================================
 // CMT04 // Shrub Tundra // Calibrated for Toolik area. J.Clein updated, July 2019
 //Salix      Betula       Decid       EGreen       Sedges       Forbs       Grasses     Lichens      Feather    PFT9     // pftnames:
@@ -143,6 +159,10 @@
 0.00      // kdcsomcrch4:
 0.0     // ch4ebulrate: ebullition rate, per hour
 0.0     // ch4transportrate: plant-mediated transport rate, per hour
+4.00      // prodq10ch4:
+2.20      // oxidq10ch4:
+5.00      // oxidkmch4:
+20.0      // oxidvmaxch4:
 //===========================================================
 // CMT05 // Tussock Tundra // Updated by E.E. calibration, April, May 2017 //Reupdated July 2019 by Joy//Joy June 2022
 // Betula      Decid      EGreen      Sedges       Forbs      Lichens    Feather     Sphag       PFT8         PFT9    // pftnames:
@@ -171,6 +191,10 @@
[…]
