### 2024-09-09 c5023fd
Adds control flag for CH4
