### 2024-09-10 5d16123
Reverting some changes and adding comments Currently, dz in stefan has been modified to account for the whole layer and not just the part below/above the front following Yi et al. 2009
