### 2024-09-10 a5fa4f4
Implement output accumulation for LAI
