### 2024-09-10 89cc621
Implement output accumulation for LWCLAYER
