+++ include/OutputHolder.h
@@ -54,6 +54,7 @@ public:
   std::vector<std::array<double, MAX_SOI_LAY>> layerdepth_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> layerdz_for_output;
   std::vector<std::array<int, MAX_SOI_LAY>> layertype_for_output;
+  std::vector<std::array<double, MAX_SOI_LAY>> lwclayer_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> tlayer_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> vwclayer_for_output;
 
