### 2024-09-10 a551ffc
Implement output accumulation for RG
