+++ src/Runner.cpp
@@ -4095,8 +4095,9 @@ void Runner::output_netCDF(std::map<std::string, OutputSpec> &netcdf_outputs, in
     {
       //PFT and compartment
       if(curr_spec.pft && curr_spec.compartment){
+        std::array<std::array<double, NUM_PFT>, NUM_PFT_PART> m_rg{};
+        std::array<std::array<double, NUM_PFT>, NUM_PFT_PART> y_rg{};
 
-        double m_rg[NUM_PFT_PART][NUM_PFT], y_rg[NUM_PFT_PART][NUM_PFT];
         for(int ip=0; ip<NUM_PFT; ip++){
           for(int ipp=0; ipp<NUM_PFT_PART; ipp++){
             m_rg[ipp][ip] = cohort.bd[ip].m_v2a.rg[ipp];
@@ -4105,37 +4106,56 @@ void Runner::output_netCDF(std::map<std::string, OutputSpec> &netcdf_outputs, in
         }
         //monthly
         if(curr_spec.monthly){
-          output_nc_5dim(&curr_spec, file_stage_suffix, &m_rg[0][0], NUM_PFT_PART, NUM_PFT, month_timestep, 1);
+          outhold.rg_for_output.push_back(m_rg);
+
+          if(output_this_timestep){
+            output_nc_5dim(&curr_spec, file_stage_suffix, &outhold.rg_for_output[0][0][0], NUM_PFT_PART, NUM_PFT, month_start_idx, months_to_output);
+            outhold.rg_for_output.clear();
+          }
         }
         //yearly
         else if(curr_spec.yearly){
-          output_nc_5dim(&curr_spec, file_stage_suffix, &y_rg[0][0], NUM_PFT_PART, NUM_PFT, year, 1);
+          outhold.rg_for_output.push_back(y_rg);
+
+          if(output_this_timestep){
+            output_nc_5dim(&curr_spec, file_stage_suffix, &outhold.rg_for_output[0][0][0], NUM_PFT_PART, NUM_PFT, year_start_idx, years_to_output);
+            outhold.rg_for_output.clear();
+          }
         }
       }
       //PFT only (4 dimensions)
       else if(curr_spec.pft && !curr_spec.compartment){
-
-        double m_rg[NUM_PFT], y_rg[NUM_PFT];
+        std::array<double, NUM_PFT> m_rg{};
+        std::array<double, NUM_PFT> y_rg{};
 
         for(int ip=0; ip<NUM_PFT; ip++){
-            m_rg[ip] = cohort.bd[ip].m_v2a.rgall;
-            y_rg[ip] = cohort.bd[ip].y_v2a.rgall;
+          m_rg[ip] = cohort.bd[ip].m_v2a.rgall;
+          y_rg[ip] = cohort.bd[ip].y_v2a.rgall;
         }
 
         //monthly
         if(curr_spec.monthly){
-          output_nc_4dim(&curr_spec, file_stage_suffix, &m_rg[0], NUM_PFT, month_timestep, 1);
+          outhold.rg_pft_for_output.push_back(m_rg);
+
+          if(output_this_timestep){
+            output_nc_4dim(&curr_spec, file_stage_suffix, &outhold.rg_pft_for_output[0][0], NUM_PFT, month_start_idx, months_to_output);
+            outhold.rg_pft_for_output.clear();
+          }
         }
         //yearly
         else if(curr_spec.yearly){
-          output_nc_4dim(&curr_spec, file_stage_suffix, &y_rg[0], NUM_PFT, year, 1);
+          outhold.rg_pft_for_output.push_back(y_rg);
+
+          if(output_this_timestep){
+            output_nc_4dim(&curr_spec, file_stage_suffix, &outhold.rg_pft_for_output[0][0], NUM_PFT, year_start_idx, years_to_output);
[…]
