### 2024-09-10 7e8ea18
Implement output accumulation for SOMA
