+++ include/OutputHolder.h
@@ -51,6 +51,9 @@ public:
   std::vector<double> rh_tot_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> rh_for_output;
 
+  std::vector<double> soma_tot_for_output;
+  std::vector<std::array<double, MAX_SOI_LAY>> soma_for_output;
+
   std::vector<std::array<double, MAX_SOI_LAY>> layerdepth_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> layerdz_for_output;
   std::vector<std::array<int, MAX_SOI_LAY>> layertype_for_output;
