### 2024-09-10 af6cc5f
Implement output accumulation for SOMCR
