+++ include/OutputHolder.h
@@ -54,6 +54,9 @@ public:
   std::vector<double> soma_tot_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> soma_for_output;
 
+  std::vector<double> somcr_for_output;
+  std::vector<std::array<double, MAX_SOI_LAY>> somcr_tot_for_output;
+
   std::vector<std::array<double, MAX_SOI_LAY>> layerdepth_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> layerdz_for_output;
   std::vector<std::array<int, MAX_SOI_LAY>> layertype_for_output;
