### 2024-09-10 573fc81
Implement output accumulation for SOMPR
