+++ include/OutputHolder.h
@@ -57,6 +57,9 @@ public:
   std::vector<double> somcr_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> somcr_tot_for_output;
 
+  std::vector<double> sompr_for_output;
+  std::vector<std::array<double, MAX_SOI_LAY>> sompr_tot_for_output;
+
   std::vector<std::array<double, MAX_SOI_LAY>> layerdepth_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> layerdz_for_output;
   std::vector<std::array<int, MAX_SOI_LAY>> layertype_for_output;
