### 2024-09-10 2214b64
Implement output accumulation for SOMRAWC
