+++ include/OutputHolder.h
@@ -54,11 +54,14 @@ public:
   std::vector<double> soma_tot_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> soma_for_output;
 
-  std::vector<double> somcr_for_output;
-  std::vector<std::array<double, MAX_SOI_LAY>> somcr_tot_for_output;
+  std::vector<double> somcr_tot_for_output;
+  std::vector<std::array<double, MAX_SOI_LAY>> somcr_for_output;
 
-  std::vector<double> sompr_for_output;
-  std::vector<std::array<double, MAX_SOI_LAY>> sompr_tot_for_output;
+  std::vector<double> sompr_tot_for_output;
+  std::vector<std::array<double, MAX_SOI_LAY>> sompr_for_output;
+
+  std::vector<double> somrawc_tot_for_output;
+  std::vector<std::array<double, MAX_SOI_LAY>> somrawc_for_output;
 
   std::vector<std::array<double, MAX_SOI_LAY>> layerdepth_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> layerdz_for_output;
