+++ scripts/tests/doctests/doctests_MadsTEMDriver.rst
@@ -52,8 +52,52 @@
 
 >>> d.setup_outputs(d.target_names)
 
->>> d.observed_vec()
+Grab the targets in two differnt formats (flat and labeled) and make sure the 
+values line up as expected.
+
+>>> import pandas as pd
+
+>>> flat_targets = d.observed_vec(format='flat')
+>>> df_targets = pd.DataFrame(d.observed_vec(format='labeled'))
+
+Check on the first block of data, which is a PFT target, but not by compartment.
+First print out the tables so we can see the expected shapes.
+>>> df_targets.loc[df_targets['ctname'] == 'GPPAllIgnoringNitrogen']
+   cmtnum                  ctname  pft  observed cmprt
+0       6  GPPAllIgnoringNitrogen    0    11.833   NaN
+1       6  GPPAllIgnoringNitrogen    1   197.867   NaN
+2       6  GPPAllIgnoringNitrogen    2    42.987   NaN
+3       6  GPPAllIgnoringNitrogen    3    10.667   NaN
+4       6  GPPAllIgnoringNitrogen    4     3.375   NaN
+5       6  GPPAllIgnoringNitrogen    5    16.000   NaN
+6       6  GPPAllIgnoringNitrogen    6     6.000   NaN
+>>> flat_targets[0:len(d.pftnums)]
 [11.833, 197.867, 42.987, 10.667, 3.375, 16.0, 6.0]
+
+Then actually check the data so that we are confident in the output order for
+the flat list.
+>>> a = flat_targets[0:len(d.pftnums)]
+>>> b = df_targets.loc[df_targets['ctname'] == 'GPPAllIgnoringNitrogen']['observed']
+>>> all(a == b)
+True
+
+Then check on some compartment data. First print out the tables of data so we 
+can see the expected shapes.
+
+>>> flat_targets[len(d.pftnums):len(d.pftnums)+3]
+[2.0, 4.0, 0.297]
+>>> df_targets.loc[ (df_targets['ctname']=='VegCarbon') & (df_targets['pft']==0) ]
+   cmtnum     ctname  pft  observed cmprt
+7       6  VegCarbon    0     2.000  Leaf
+8       6  VegCarbon    0     4.000  Stem
+9       6  VegCarbon    0     0.297  Root
+
+Then check the data so that we are confident that we understand the ordering.
+
+>>> a = flat_targets[len(d.pftnums):len(d.pftnums)+3]
+>>> b = df_targets.loc[ (df_targets['ctname']=='VegCarbon') & (df_targets['pft']==0) ]['observed']
+>>> all(a == b)
+True
 
 >>> d.params_vec()
 [22.8, 250.6, 65.0, 38.5, 7.8, 21.0, 36.3]
@@ -77,9 +121,27 @@ This makes sense because we haven't run the model yet so there are no outputs.
 
 >>> final_data = d.gather_model_outputs()
 >>> import pandas as pd
+>>> df_finaldata = pd.DataFrame(final_data)
+>>> df_finaldata.loc[(df_finaldata['ctname']=='VegCarbon') & (df_finaldata['cmprt']=='Leaf')]
+      cmt     ctname      value  truth  pft cmprt
+7   CMT06  VegCarbon   2.138998   2.00    0  Leaf
+10  CMT06  VegCarbon  42.925257  37.10    1  Leaf
+12  CMT06  VegCarbon   0.156739   8.06    2  Leaf
+14  CMT06  VegCarbon   2.602119   2.00    3  Leaf
+16  CMT06  VegCarbon   2.250932   2.00    4  Leaf
+17  CMT06  VegCarbon  22.572059  22.00    5  Leaf
[…]
