### 2024-09-10 f77d4c1
Trim a few words from software dev section.
