### 2024-09-10 fda49a5
Trim more words from text.
