### 2024-09-11 1dbec21
Implements holding for RM
