### 2024-09-11 5229bcc
Implements holding for CMTNUM
