### 2024-09-11 29f7ea6
Implements holding for MOSSDEATHC, MOSSDEATHN
