+++ src/Runner.cpp
@@ -3181,11 +3181,21 @@ void Runner::output_netCDF(std::map<std::string, OutputSpec> &netcdf_outputs, in
     {
       //monthly
       if(curr_spec.monthly){
-        output_nc_3dim(&curr_spec, file_stage_suffix, &cohort.bdall->m_v2soi.mossdeathc, 1, month_timestep, 1);
+        outhold.mossdeathc_for_output.push_back(cohort.bdall->m_v2soi.mossdeathc);
+
+        if(output_this_timestep){
+          output_nc_3dim(&curr_spec, file_stage_suffix, &outhold.mossdeathc_for_output[0], 1, month_start_idx, months_to_output);
+          outhold.mossdeathc_for_output.clear();
+        }
       }
       //yearly
       else if(curr_spec.yearly){
-        output_nc_3dim(&curr_spec, file_stage_suffix, &cohort.bdall->y_v2soi.mossdeathc, 1, year, 1);
+        outhold.mossdeathc_for_output.push_back(cohort.bdall->y_v2soi.mossdeathc);
+
+        if(output_this_timestep){
+          output_nc_3dim(&curr_spec, file_stage_suffix, &outhold.mossdeathc_for_output[0], 1, year_start_idx, years_to_output);
+          outhold.mossdeathc_for_output.clear();
+        }
       }
     }//end critical(outputMOSSDEATHC)
   }//end MOSSDEATHC
@@ -3202,11 +3212,21 @@ void Runner::output_netCDF(std::map<std::string, OutputSpec> &netcdf_outputs, in
     {
       //monthly
       if(curr_spec.monthly){
-        output_nc_3dim(&curr_spec, file_stage_suffix, &cohort.bdall->m_v2soi.mossdeathn, 1, month_timestep, 1);
+        outhold.mossdeathn_for_output.push_back(cohort.bdall->m_v2soi.mossdeathn);
+
+        if(output_this_timestep){
+          output_nc_3dim(&curr_spec, file_stage_suffix, &outhold.mossdeathn_for_output[0], 1, month_start_idx, months_to_output);
+          outhold.mossdeathn_for_output.clear();
+        }
       }
       //yearly
       else if(curr_spec.yearly){
-        output_nc_3dim(&curr_spec, file_stage_suffix, &cohort.bdall->y_v2soi.mossdeathn, 1, year, 1);
+        outhold.mossdeathn_for_output.push_back(cohort.bdall->y_v2soi.mossdeathn);
+
+        if(output_this_timestep){
+          output_nc_3dim(&curr_spec, file_stage_suffix, &outhold.mossdeathn_for_output[0], 1, year_start_idx, years_to_output);
+          outhold.mossdeathn_for_output.clear();
+        }
       }
     }//end critical(outputMOSSDEATHN)
   }//end MOSSDEATHN
