### 2024-09-11 #7136 @kthyng
@tobeycarman

> Docker, we don't actively maintain install instructions without Docker. We found that using Docker images and containers to standardize the development environments dramatically improved the development process.

I won't block this from review based on this but this may come up again.

> Empty sections in the comments, I am not sure what you are referring to?

I see I wrote comment but I meant your docs. As an example, lots of [this page](https://uaf-arctic-eco-modeling.github.io/dvm-dos-tem/model_overview.html#soil-layers) says "write this".

> Word count, I will look into reducing it. We would gladly accept feedback on which sections reviewers find to be less relevant.

Please, follow this: https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain. Information that is not in the paper can be in your docs. You probably want sections Summary, Statement of Need, maybe one more section, and Acknowledgments, maybe a figure or two. The goal is to explain the software at a high level to a general audience initially, then give just enough information at the next level for someone to know if they want more information, comparing with the relevant other packages.
