### 2024-09-11 63bfdb5
Simplifies ice redistribution post layer splitting
This should resolve an issue with ice being assigned
negative values, or values greater than the original
ice quantity being split.
