### 2024-09-12 e911c9a
More editing and trimming of text courtesy of M.J. Lara.
