+++ joss-paper/paper.md
@@ -147,9 +147,7 @@ bibliography: joss-paper.bib
 
 The impacts of climate change on natural ecosystems are the result of complex
 physical and ecological processes operating and interacting at a variety of
-spatio-temporal scales. For this reason, process-based ecosystem models are
-efficient tools to formalize and extrapolate our current understanding of
-ecosystem dynamics and predict local, regional and global climate impacts.
+spatio-temporal scales, that can be represented in process-based ecosystem models.
 
 `DVMDOSTEM` is an advanced process-based terrestrial ecosystem model (TEM)
 designed to study ecosystem responses to climate changes and disturbances. It
@@ -166,7 +164,7 @@ in permafrost, vegetation, and carbon (C) and nitrogen (N) dynamics.
 
 Arctic and boreal regions underlain by permafrost store nearly half of the
 world’s soil organic C - approximately 1,440-1,600 Pg [@Hugelius2014;
-@Schuur2022]. These regions are warming roughly two to four times faster than
+@Schuur2022]. These regions are warming four times faster than
 the rest of the globe, driving widespread and rapid permafrost thaw
 [@Rantanen2022; @Smith2022]. As permafrost thaws, soil organic C becomes
 available for decomposition and release as greenhouse gasses (GHGs) to the
@@ -174,31 +172,29 @@ atmosphere. Climate-driven permafrost thaw and the associated release of GHGs
 can influence the global climate system, a phenomenon called the permafrost
 carbon-climate feedback or PCCF [@Koven2011; @Schuur2015]. The PCCF has been
 identified as one of the largest sources of uncertainty in future climate
-projections and therefore needs to be accurately represented in process-based
-ecosystem models that inform global earth system models [@Schadel2024].
-
+projections and therefore needs to be accurately represented in global earth 
+system models [@Schadel2024].
 
 # Model Design
 
-`DVMDOSTEM` is a process-based ecosystem model designed to simulate the key
+`DVMDOSTEM` is designed to simulate the key
 biophysical and biogeochemical processes between the soil, the vegetation and
 the atmosphere. The evolution and refinement of `DVMDOSTEM` have been shaped by
 extensive research programs and applications both in permafrost and
 non-permafrost regions [@Genet2013; @Genet2018; @Jafarov2013; @Yi2010;
 @Yi2009; @Euskirchen2022; @Briones2024]. The model is spatially explicit and
 represents ecosystem response to climate and disturbances at seasonal (i.e.
-monthly) up to centennial scales. The snow and soil columns are split into a
+monthly) to centennial scales. The snow and soil columns are split into a
 dynamic number of layers to represent their impact on thermal and hydrological
 dynamics and the consequences for soil C and N dynamics. Vegetation composition
 is modeled using community types (CMTs), each of which consists of multiple
 plant functional types (PFTs - groups of species sharing similar ecological
 traits). This structure allows the model to represent the effect of competition
-for light, water and nutrients on vegetation composition [@Euskirchen2009]. The
-model also represents the ecosystem N cycle to evaluate the role of nutrient
-limitations, characteristic of permafrost regions, on ecosystem dynamics, with
-coupling between the C and N cycles [@McGuire1992; @Euskirchen2009]. Finally,
+for light, water and nutrients on vegetation composition [@Euskirchen2009], as well
+as the role of nutrient limitation on permafrost ecosystem dynamics, with
+coupling between C and N cycles [@McGuire1992; @Euskirchen2009]. Finally,
 the model represents the effects of wildfire in order to evaluate the role of
-climate-driven fire intensification on ecosystem structure and functions
+climate-driven fire intensification on ecosystem structure and function
 [@Yi2010; @Genet2013]. The structure of `DVMDOSTEM` is represented visually in
 \autoref{fig:modeloverview}.
 
[…]
