+++ mads_calibration/SA_post_hoc_analysis.py
@@ -42,7 +42,7 @@ import matplotlib.pyplot as plt
 from matplotlib.lines import Line2D
 
 # This stuff is diamond box in SA (orange half)
-def plot_boxplot(results, targets, save=False, saveprefix=''):
+def plot_boxplot(results, targets, check_filter=None, save=False, saveprefix=''):
   '''
   Plots a box and whiskers for each column in ``results``. Plots a dot for
   each target value.
@@ -61,6 +61,10 @@ def plot_boxplot(results, targets, save=False, saveprefix=''):
   targets : pandas.DataFrame
     One column for each target (truth, or observation) value. One row.
 
+  check_filter : pandas.Series
+    Series with boolean values for additional tests following
+    SA run, e.g. equilibrium_check, eq_check['result']
+
   save : bool
     Assumes False so plot will not be saved. If set to true it will plot
     in current directory unless saveprefix is specified
@@ -73,13 +77,35 @@ def plot_boxplot(results, targets, save=False, saveprefix=''):
   None
   '''
   plt.close('all')
-  fig, ax = plt.subplots(nrows=1, ncols=1,figsize=(6,6))
-  results.boxplot(ax=ax, rot=45)
-  ax.scatter(range(1,len(targets.columns)+1), targets, color='red', zorder=1000)
+
+  x = np.linspace(1, len(results.columns), len(results.columns))
+
+  if check_filter is not None:
+    c = 'grey'
+    plt.boxplot(results, positions=x-0.15, widths=0.25,
+          boxprops=dict(color=c), capprops=dict(color=c), whiskerprops=dict(color=c), 
+          flierprops=dict(color=c, markeredgecolor=c), medianprops=dict(color=c)
+          )
+    c = 'green'
+    plt.boxplot(results[check_filter==True], positions=x+0.15, widths=0.25,
+          boxprops=dict(color=c), capprops=dict(color=c), whiskerprops=dict(color=c), 
+          flierprops=dict(color=c, markeredgecolor=c), medianprops=dict(color=c)
+          )
+    plt.xticks(x, results.columns, rotation=90)
+  else:
+    c = 'grey'
+    plt.boxplot(results, positions=x, widths=0.25,
+          boxprops=dict(color=c), capprops=dict(color=c), whiskerprops=dict(color=c), 
+          flierprops=dict(color=c, markeredgecolor=c), medianprops=dict(color=c)
+          )
+    plt.xticks(x, results.columns, rotation=90)
+  
+  plt.scatter(x, targets, color='red', zorder=1000)
+
   if save:
     plt.savefig(saveprefix + "results_boxplot.png", bbox_inches='tight')
 
-def plot_spaghetti(results, targets, save=False, saveprefix=''):
+def plot_spaghetti(results, targets, check_filter=None, save=False, saveprefix=''):
   '''
   Plots one line for each sample (row) in ``results``. Plots targets as dots.
   X axis of plot are for different columns in ``results``. Makes 2 plots, the 
@@ -99,6 +125,10 @@ def plot_spaghetti(results, targets, save=False, saveprefix=''):
   targets : pandas.DataFrame
     Single row, one column for each target (truth, or observation) value.
 
+  check_filter : pandas.Series
+    Series with boolean values for additional tests following
+    SA run, e.g. equilibrium_check, eq_check['result']
+
   save : bool
     Assumes False so plot will not be saved. If set to true it will plot
     in current directory unless saveprefix is specified
@@ -115,20 +145,17 @@ def plot_spaghetti(results, targets, save=False, saveprefix=''):
 
   for i, sample in results.iterrows():
     ax1.plot(sample, color='gray', alpha=0.1)
+    if check_filter is not None:
[…]
