### 2024-09-18 76b5daf
Removes hardcoded testing cell coordinates
