+++ src/Cohort.cpp
@@ -871,7 +871,7 @@ void Cohort::updateMonthly_DIMveg(std::string stage, const int & yearind, const 
 
   //Pause here for LAI data assimilation if specified
   if(this->dacontroller_ptr){
-    timestep_id curr_step(0, 0, stage, yearind, currmind);
+    timestep_id curr_step(this->x, this->y, stage, yearind, currmind);
 
     this->dacontroller_ptr->run_DA_LAI(curr_step);
   }
