### 2024-09-18 f355803
Switches LAI for DA to by-pft
