+++ src/TEMUtilityFunctions.cpp
@@ -689,6 +689,45 @@ namespace temutil {
     return data;
   }
 
+  std::array<double, NUM_PFT> read_pft_var_timestep(
+                                    const std::string &filename,
+                                    const std::string &varname,
+                                    const int y, const int x){
+
+    BOOST_LOG_SEV(glg, fatal) << "Opening dataset: " << filename;
+
+    int ncid;
+    temutil::nc( nc_open(filename.c_str(), NC_NOWRITE, &ncid) );
+
+    int varV;
+    temutil::nc( nc_inq_varid(ncid, varname.c_str(), &varV) );
+
+    BOOST_LOG_SEV(glg, note) << "Getting value for pixel(y,x): ("<< y <<","<< x <<").";
+    int yD, xD;
+    size_t yD_len, xD_len;
+
+    // specify start indices for each dimension (y, x)
+    size_t start[4];
+    start[0] = 0; // from beginning of time
+    start[1] = 0; //pft
+    start[2] = y;
+    start[3] = x;
+
+    // specify counts for each dimension
+    size_t count[5];
+    count[0] = 1;     // One timestep
+    count[1] = NUM_PFT;
+    count[2] = 1;     // one location
+    count[3] = 1;     // one location
+
+    std::array<double, NUM_PFT> data;
+    temutil::nc( nc_get_vara_double(ncid, varV, start, count, &data[0]) );
+
+    temutil::nc( nc_close(ncid) );
+
+    return data;
+  }
+
   std::array<std::array<double, NUM_PFT>, NUM_PFT_PART> read_veg_var_timestep(
                                           const std::string &filename,
                                           const std::string &varname,
