+++ src/Stefan.cpp
@@ -27,12 +27,15 @@ void Stefan::updateFronts(const double & tdrv, const double &timestep) {
   Layer * toplayer = ground->toplayer;
   double tkres; //thermal conductivity for calculating resistence
   double tkfront; //thermal conductivity for calculating part front depth
-  double tkunf, tkfrz, tkmix;
+  double tkunf, tkfrz;
   // top-down propogation of front
   // driving force
   int freezing1; //the freezing/thawing force based on the driving temperature
   double tdrv1 = tdrv;
-  // extra degree second - Yi et al. 2009
+  //20180820 While comparing the code to the papers describing these
+  // processes, it seemed like tdrv1 needed to be converted from
+  // degrees C to degrees K. However, we tested that change, and
+  // it appears that not converting it produces better results.
   double dse = fabs(tdrv1 * timestep); // the extra degree second
   double sumresabv  =0. ; // sum of resistence for above layers;
 
@@ -63,14 +66,13 @@ void Stefan::updateFronts(const double & tdrv, const double &timestep) {
 
     tkunf = currl->getUnfThermCond();
     tkfrz = currl->getFrzThermCond();
-    tkmix = currl->getMixThermCond();
 
     if(tdrv1<0.0) {
-      tkres   = tkmix;
-      tkfront = tkmix;
+      tkres   = tkfrz;
+      tkfront = tkunf;
     } else {
-      tkres   = tkmix;
-      tkfront = tkmix;
+      tkres   = tkunf;
+      tkfront = tkfrz;
     }
 
     sumresabv += currl->dz/tkres;
@@ -175,10 +177,10 @@ void Stefan::processNewFrontSoilLayerDown(const int &freezing,
   if (freezing==1) {
     dz *= fmax(0., 1.0-sl->frozenfrac); //assuming frozen/unfrozen
                                         //  soil segments not mixed
-    volwat = fmax(0.0, fmax(sl->getVolLiq(),sl->getUnfVolLiq())) * dz;
-  } else { // >>> BM look at dos-tem, it looks like sumrescum should be calculated here also
+    volwat = fmax(0.0, sl->getVolLiq())*sl->dz;
+  } else {
     dz *= sl->frozenfrac; //assuming frozen/unfrozen soil segments not mixed
-    volwat = fmax(0.0, sl->getVolIce()) * dz;
+    volwat = fmax(0.0, sl->getVolIce())*sl->dz;
   }
 
   if (dz<=0.0001*sl->dz) { //this will avoid 'front' exactly falling on the
@@ -199,8 +201,8 @@ void Stefan::processNewFrontSoilLayerDown(const int &freezing,
     dse -= dsn;
   } else {
     double partdz=0.;
-    partdz = getPartialDepth(volwat / dz, tkfront, sumrescum, dse); // may not be consistent
-    // partdz = dse/dsn*dz;
+    //partdz=getPartialDepth(volwat, tkfront, sumrescum, dse); //may not be consistent
+    partdz = dse/dsn*dz;
     // find the existing front(s) within the layer
     int fntnum = ground->frontsz.size();
     vector<int> fntsintype;
@@ -353,10 +355,10 @@ void Stefan::processNewFrontSoilLayerUp(const int &freezing,
   if (freezing==1) {
     dz *= fmax(0., 1.0-sl->frozenfrac); //assuming frozen/unfrozen
                                         //  soil segments not mixed
-    volwat = fmax(0.0, fmax(sl->getVolLiq(),sl->getUnfVolLiq())) * dz;
[…]
