+++ mads_calibration/sa-demo-config.yaml
@@ -8,10 +8,10 @@ seed_path: /work/parameters
 observations: /work/calibration
 
 # The number of samples to draw over the parameter space
-N_samples: 500
+N_samples: 10
 
 # The location where all the individual sample folders will be setup.
-work_dir: /data/workflows/ch4_test_range
+work_dir: /data/workflows/ch4_test_new_params
 
 # The following keys (params, pftnums, and percent_diffs) should be lists of
 # identical length. I.e. for each parameter we need to specify the PFT and a
@@ -21,6 +21,7 @@ work_dir: /data/workflows/ch4_test_range
 
 # Which parameters are being tested/analyzed
 params: [kdcrawc, kdcsoma, kdcsomcr, kdcsompr,
+         rhq10, s2dfraction,
          kdcrawcch4, kdcsomach4, kdcsomcrch4, kdcsomprch4,
          ch4ebulrate, ch4transportrate, 
          prodq10ch4, oxidq10ch4,
@@ -28,6 +29,7 @@ params: [kdcrawc, kdcsoma, kdcsomcr, kdcsompr,
 
 # Which PFTs are being tested/analyzed
 pftnums: [None, None, None, None,
+          None, None,
           None, None, None, None, 
           None, None,
           None, None,
@@ -40,11 +42,17 @@ pftnums: [None, None, None, None,
 # with percent_diffs. In other words only specify one of percent_diffs or 
 # p_bounds. For each parameter specify the min and max. Must use square 
 # brackets for yaml to parse it correctly.
-p_bounds: [[0.00035, 0.45],[0.00001, 0.4],[0.00001, 0.09],[0.0000001, 0.001],
-           [0.00000001,0.002],[0.00000001,0.00007],[0.0000001, 0.00025],[0.0003, 0.000375],
+p_bounds: [[0.003, 0.06],[0.05, 0.4],[0.000001, 0.09],[0.000000001, 0.000001],
+           [2.3, 10.3],[0.001, 1.0],
+           [0.0005,0.0015],[0.0001,0.01],[0.00000001, 0.0001],[0.000002, 0.000005],
            [0.01,1.0], [0.18,0.21],
            [10,16],[1.4,2.1],
            [1,5],[5,50]]
+           #p_bounds: [[0.000035, 0.9],[0.000001, 0.4],[0.000001, 0.09],[0.00000001, 0.0001],
+           #           [0.0000001,0.001],[0.0000001,0.001],[0.00000001, 0.0001],[0.0000000001, 0.00001],
+           #           [0.01,1.0], [0.18,0.21],
+           #           [10,16],[1.4,2.1],
+           #           [1,5],[5,50]
 
 # The path to the driving input data that will be used.
 site: /data/input-catalog/cru-ts40_ar5_rcp85_ncar-ccsm4_bonanzacreeklter_10x10
