### 2024-09-23 19b3645
Corrections to follow dos-tem Stefan logic
