### 2024-09-23 de2e897
Scales liquid redistribution by ratio of unfrozen sections
Also adds some comments and cleanup details.
