+++ src/RestartData.cpp
@@ -117,6 +117,8 @@ MPI_Datatype RestartData::register_mpi_datatype() {
     MAX_SOI_LAY, // double soma[MAX_SOI_LAY];
     MAX_SOI_LAY, // double sompr[MAX_SOI_LAY];
     MAX_SOI_LAY, // double somcr[MAX_SOI_LAY];
+
+    MAX_SOI_LAY, // double ch4[MAX_SOI_LAY]
     
     1, // double wdebrisn;
     MAX_SOI_LAY, // double orgn[MAX_SOI_LAY];
@@ -185,6 +187,7 @@ MPI_Datatype RestartData::register_mpi_datatype() {
     MPI_DOUBLE, // double soma[MAX_SOI_LAY];
     MPI_DOUBLE, // double sompr[MAX_SOI_LAY];
     MPI_DOUBLE, // double somcr[MAX_SOI_LAY];
+    MPI_DOUBLE, // double ch4[MAX_SOI_LAY];
     MPI_DOUBLE, // double wdebrisn;
     MPI_DOUBLE, // double orgn[MAX_SOI_LAY];
     MPI_DOUBLE, // double avln[MAX_SOI_LAY];
@@ -250,6 +253,7 @@ MPI_Datatype RestartData::register_mpi_datatype() {
     offsetof(RestartData, soma),
     offsetof(RestartData, sompr),
     offsetof(RestartData, somcr),
+    offsetof(RestartData, ch4),
     offsetof(RestartData, wdebrisn),
     offsetof(RestartData, orgn),
     offsetof(RestartData, avln),
@@ -373,6 +377,7 @@ void RestartData::reinitValue() {
     soma[il]  = MISSING_D;
     sompr[il] = MISSING_D;
     somcr[il] = MISSING_D;
+    ch4[il] = MISSING_D;
     orgn[il] = MISSING_D;
     avln[il] = MISSING_D;
 
@@ -528,6 +533,7 @@ void RestartData::verify_logical_values(){
     check_bounds("soma", soma[ii]);
     check_bounds("sompr", sompr[ii]);
     check_bounds("somcr", somcr[ii]);
+    check_bounds("ch4", ch4[ii]);
     check_bounds("orgn", orgn[ii]);
     check_bounds("avln", avln[ii]);
     for(int jj=0; jj<12; jj++){
@@ -820,6 +826,8 @@ void RestartData::read_px_soil_vars(const std::string& fname, const int rowidx, 
   temutil::nc( nc_get_vara_double(ncid, cv, start, count, &sompr[0]) );
   temutil::nc( nc_inq_varid(ncid, "somcr", &cv) );
   temutil::nc( nc_get_vara_double(ncid, cv, start, count, &somcr[0]) );
+  temutil::nc( nc_inq_varid(ncid, "ch4", &cv) );
+  temutil::nc( nc_get_vara_double(ncid, cv, start, count, &ch4[0]) );
   temutil::nc( nc_inq_varid(ncid, "orgn", &cv) );
   temutil::nc( nc_get_vara_double(ncid, cv, start, count, &orgn[0]) );
   temutil::nc( nc_inq_varid(ncid, "avln", &cv) );
@@ -972,15 +980,6 @@ void RestartData::create_empty_file(const std::string& fname,
   BOOST_LOG_SEV(glg, debug) << "Creating dimensions...";
   temutil::nc( nc_def_dim(ncid, "Y", ysize, &yD) );
   temutil::nc( nc_def_dim(ncid, "X", xsize, &xD) );
[…]
