### 2024-09-24 0a4b6c5
Stefan changes, bring parameters out and no runoff
