+++ include/MineralLayer.h
@@ -1,12 +1,13 @@
 #ifndef MINERALLAYER_H_
 #define MINERALLAYER_H_
 #include "SoilLayer.h"
-
+#include "CohortLookup.h"
+#include "cohortconst.h"
 
 class MineralLayer: public SoilLayer {
 public:
-
-  MineralLayer(const double & pdz, float psand, float psilt, float pclay
+  MineralLayer(const double &pdz, float psand, float psilt, float pclay,
+               const CohortLookup *chtlu
                /*int sttype */);
 
   ~MineralLayer();
@@ -21,7 +22,7 @@ public:
                          const double & partden);
 
 private:
-  void  updateProperty5Lookup();
+  void  updateProperty5Lookup(const CohortLookup *chtlu);
 
 };
 #endif /*MINERALLAYER_H_*/
