### 2024-09-26 28e6f87
added get_metric_matrix function and few other auxilary functions
