+++ mads_calibration/SA_post_hoc_analysis.py
@@ -42,6 +42,287 @@ import matplotlib.pyplot as plt
 from matplotlib.lines import Line2D
 
 # This stuff is diamond box in SA (orange half)
+from sklearn.metrics import r2_score,mean_squared_error,mean_absolute_error
+from sklearn.metrics import mean_absolute_percentage_error
+
+def param_and_targets_box_plots(x,y,y_true):
+    ''' plots the box plots the difererence between best parameters values and their mean 
+        and best modeled values and targets 
+        x: parameter dataframe (nxm)
+        y: model output dataframe (nxm))
+        y_true: target dataframe
+    '''
+    
+    fig, axs = plt.subplots(1, 2, figsize=(8, 4))
+
+    axs[0].boxplot(x-x.mean().values, labels=x.columns);
+    axs[0].set_ylabel('parameters: x-$\overline{x}$', fontsize=12)
+    axs[0].set_xticklabels(x.columns, rotation=45, fontsize=12)
+
+    axs[1].boxplot(y-y_true.values, labels=y.columns);
+    axs[1].set_ylabel('targets: y-$y_{obs}$', fontsize=12)
+    axs[1].set_xticklabels(y.columns, rotation=45, fontsize=12)
+
+    plt.tight_layout()
+
+def best_mean_match_spagetti(df_x,df_y,site_name=''):
+    ''' plots the spaghetti plot of modeled v.s. observed values 
+        df_x: parameter dataframe (nxm)
+        df_y: model output dataframe (nx(m+1)), last columns must be targets
+        site_name: string 
+    '''
+    nrange=range(len(df_y.columns))
+    df_xx, df_yy =  get_match_metric(df_x,df_y)
+
+    rmetric='r2rmse'
+    nelem=10
+    order=True
+    y=df_yy.sort_values(by=[rmetric],ascending=order)[:nelem]
+
+    yy=y.iloc[:,:-6]
+    yy.columns = list(nrange)
+    fig, ax = plt.subplots(figsize=(6, 3))
+    ax.plot(yy.mean().transpose(),alpha=0.5,color='black',linewidth=2.5)
+    ax.plot(df_y.iloc[-1,:],'o',color='red')
+    ax.fill_between(nrange, df_y.iloc[0:-1,:].min(), df_y.iloc[0:-1,:].max(), alpha=0.5,color='grey')
+    ax.set_xticklabels(df_y.columns, rotation=45, fontsize=12)
+    ax.set_title(f"{site_name}")
+    #ax.setxtick(df_y.columns,fontsize=12)
+    #ax.setylabel(ylabel,fontsize=14)
+    #ax.xlim([-0.1,6.1])
+
+    return
+
+def plot_site_metric_matrix(metric_matrix, error, colorb=True):
+    import seaborn as sns
+    """Plot a heatmap for the given metric matrix."""
+    [x_size,y_size]=metric_matrix.shape
+    fig, ax = plt.subplots(figsize=(2+x_size, 3))
+
+    cbar_label = 'RMSE score' if error == 'RMSE' else 'RE score'
+    cmap = "GnBu" if colorb else "coolwarm"
+    
+    sns.heatmap(metric_matrix, cmap=cmap, annot=True, fmt=".3f",
+                cbar_kws={'label': '', "orientation": 'vertical'},
+                annot_kws={"fontsize": 12}, ax=ax, cbar=colorb)
+    
+    ax.tick_params(axis='x', rotation=45)
+    ax.set_title(cbar_label)
+
+    return
+
+def get_metric_matrix(df_x,df_y,pft_names,error):
+    ''' plots the metric matrix error between modeled and observed values for all pfts 
+        df_x: parameter dataframe (nxm)
[…]
