### 2024-09-26 9b775b2
Renames output DWDRH->RHDWD
Updates pullrequest_test.py variable names for both
RHDWD and RHSOM.
