### 2024-09-26 7c6deb1
Fixing labeled and flat formatting in observed_vec() I have added a suggested fix for the outlabels bug in comments above the current break point, this does not fully work but may have some syntax that will be useful.
