### 2024-09-26 a93ab39
Adds output SOC - sum of individual carbon pools
