### 2024-09-27 efd85a2
Adds TEM_LAYERDZ and TEM_TLAYER to the DA statefile
These are reference variables and are not updated
from the DA process.
