+++ mads_calibration/AC-MADS-TEM.jl
@@ -273,10 +273,7 @@ println("targets=np.array(", targets, "),")
 
 # Generate a list of nicely formatted labels that can be used for plotting
 # These labels are for the output variables (aka calibration targets)
-
-### Suggestion for correcting outlabels issue:
 outlabels = []
-
 for x in dvmdostem.gather_model_outputs()
 
   if haskey(x, "pft") && haskey(x, "cmprt")
@@ -288,11 +285,10 @@ for x in dvmdostem.gather_model_outputs()
   else
     push!(outlabels, string(x["ctname"]))
   end
+
 end
 
 
-
-# outlabels=[string(x["ctname"],"_pft",x["pft"]) for x in dvmdostem.gather_model_outputs()]
 println("out_labels=", outlabels)
 println("")
 
