### 2024-09-30 1f958a0
Update doctest for MadsTEMDriver
