### 2024-09-30 bc897ff
add doctest for MadsTEMDriver
Adding test for when only soil params are specified
