### 2024-09-30 72ad11a
Add another test file for MadsTEMDriver.py
