+++ scripts/tests/doctests/doctests_MadsTEMDriver_alltypestargets.rst
@@ -1,0 +1,123 @@
+# Testing, developing and describing the interface for MadsTEMDriver.py
+=========================================================================
+
+This file is intended to test that the driver can work with all three types of
+targets specified (soil, PFT, PFT-compartment).
+
+>>> import yaml
+
+>>> my_yaml_string = """
+... work_dir: /tmp/test_CA
+... site: /work/testing-data/inputs/cru-ts40_ar5_rcp85_ncar-ccsm4_IMNAVIAT_CREEK_10x10
+... PXx: 0
+... PXy: 0
+... calib_mode: GPPAllIgnoringNitrogen
+... target_names: 
+... - GPPAllIgnoringNitrogen
+... - MossDeathC
+... - VegCarbon
+... cmtnum: 6
+... opt_run_setup: --pr-yrs 5 --eq-yrs 10 --sp-yrs 0 --tr-yrs 0 --sc-yrs 0
+... params:
+... - cmax
+... - cmax
+... - cmax
+... - cmax
+... - cmax
+... - cmax
+... - cmax
+... pftnums:
+... - 0
+... - 1
+... - 2
+... - 3
+... - 4
+... - 5
+... - 6
+... """
+
+>>> config_dict = yaml.load(my_yaml_string, Loader=yaml.FullLoader)
+
+>>> import drivers.MadsTEMDriver
+
+>>> d = drivers.MadsTEMDriver.MadsTEMDriver(config_dict)
+
+>>> d.set_seed_path('/work/parameters')
+
+>>> d.set_params_from_seed()
+
+>>> d.load_target_data(ref_target_path='/work/calibration/')
+
+>>> d.setup_outputs(d.target_names)
+
+Grab the targets in two differnt formats (flat and labeled) and make sure the
+values line up as expected. 
+
+>>> import pandas as pd
+
+>>> flat_targets = d.observed_vec(format='flat')
+>>> df_targets = pd.DataFrame(d.observed_vec(format='labeled'))
+
+Check on the first block of data, which is a PFT target, but not by compartment.
+First print out the tables so we can see the expected shapes.
+
+>>> df_targets.loc[df_targets['ctname'] == 'MossDeathC']
+   cmtnum      ctname  pft  observed cmprt
+7       6  MossDeathC  NaN     178.0   NaN
+
+Then actually check the data so that we are confident in the output order for
+the flat list.
+
+
+Then check on some compartment data. Here we need to make sure that only targets
+are included for PFTs, and compartments that are defined. So for example while
+the main targets data structure includes zero for undefined PFTs and
+compartments, these zero values should not be present in the `.observed_vec()`
+outputs. E.g.:
+
+>>> d.targets['VegCarbon']['Leaf']
+[2.0, 37.1, 8.06, 2.0, 2.0, 22.0, 23.0, 0.0, 0.0, 0.0]
+
+versus
+
+>>> df_targets.loc[(df_targets['ctname']=='VegCarbon') & (df_targets['pft']==4)]
+    cmtnum     ctname  pft  observed cmprt
+17       6  VegCarbon  4.0       2.0  Leaf
+
+
+>>> d.params_vec()
+[22.8, 250.6, 65.0, 38.5, 7.8, 21.0, 36.3]
+
+This makes sense because we haven't run the model yet so there are no outputs.
+
+>>> d.clean()
+
+>>> d.setup_run_dir()
+
+>>> d.run()
+
+>>> final_data = d.gather_model_outputs()
+>>> import pandas as pd
+>>> df_finaldata = pd.DataFrame(final_data)
+>>> df_finaldata
[…]
