### 2024-10-01 b6be3f1
Work on bugs with plot_optimization_fit
Problems with axes labeling.

Start by setting label rotation to 90 to accomodate longer labels, and
re-organize funciton to group commands for each axes together.
