### 2024-10-01 a1389ab
work on bug with plot_optimization_fit
Looks like this should have only handled axes 1 and 2 because they have
the same x scale, while axes 0 is different.
