### 2024-10-01 066a72f
Fixing plotting following mads
