+++ src/Stefan.cpp
@@ -27,7 +27,7 @@ void Stefan::updateFronts(const double & tdrv, const double &timestep) {
   Layer * toplayer = ground->toplayer;
   double tkres; //thermal conductivity for calculating resistence
   double tkfront; //thermal conductivity for calculating part front depth
-  double tkunf, tkfrz, tkmix;
+  double tkunf, tkfrz, tkmix, tksnw;
   // top-down propogation of front
   // driving force
   int freezing1; //the freezing/thawing force based on the driving temperature
@@ -35,16 +35,26 @@ void Stefan::updateFronts(const double & tdrv, const double &timestep) {
   double dse = fabs(tdrv1 * timestep); // the extra degree second
   double sumresabv  =0. ; // sum of resistence for above layers;
 
-  if(tdrv1>0.0) {
-    freezing1 = -1;
-  } else {
-    freezing1 =1;
+  // If top layer is soil compare to freezing temperature depression
+  // otherwise compare to 0.0 degrees
+  if (toplayer->isSoil){
+    if(tdrv1 >= toplayer->temp_dep) {
+      freezing1 = -1;
+    } else {
+      freezing1 =1;
+    }
+  } else{
+    if(tdrv1 > 0.0) {
+      freezing1 = -1;
+    } else {
+      freezing1 =1;
+    }
   }
 
   // find the new front
   double newfntz1 = 0.;
   Layer * currl=NULL;
-  if(ground->fstsoill->tem >0){ // if e.g. yesterday's tempupdater pushed the soil surface temp above zero
+  if(ground->fstsoill->tem >= ground->fstsoill->temp_dep){ // if e.g. yesterday's tempupdater pushed the soil surface temp above zero
     currl = ground->fstsoill; // skip any snow layers and force stefan to work in the soil
   } else {
     currl = toplayer; // otherwise, work in the snow layers (if any)
@@ -56,7 +66,6 @@ void Stefan::updateFronts(const double & tdrv, const double &timestep) {
         newfntz1 = currl->prevl->z+currl->prevl->dz; //this will weep out all
                                                      //fronts in the soil column
       }
-
       break; // for bedrock, break
     }
 
@@ -64,17 +73,30 @@ void Stefan::updateFronts(const double & tdrv, const double &timestep) {
     tkfrz = currl->getFrzThermCond();
     tkmix = currl->getMixThermCond();
 
-    if(tdrv1<0.0) {
-      tkres   = tkmix;
-      tkfront = tkmix;
+    // If top layer is soil compare to freezing temperature depression
+  // otherwise compare to 0.0 degrees
+  if (toplayer->isSoil){
+    if (tdrv1 < currl->temp_dep){
+      tkres = tkfrz;   // assume resistive tk above layer matches driving temperature
+      tkfront = tkmix; // assume tk of the layer containing the front is mixed
     } else {
-      tkres   = tkmix;
+      tkres = tkunf;
       tkfront = tkmix;
+    } 
+  } else {
+    if (tdrv1 < 0.0){                  
+      tkres = tkfrz;   // assume resistive tk above layer matches driving temperature
+      tkfront = tkunf; // assume tk of the layer containing the front is mixed
+    } else {
+      tkres = tkunf;
+      tkfront = tkfrz;
     }
+  }
 
+    // accumulating thermal resistance above the front layer
     sumresabv += currl->dz/tkres;
 
-    if (currl->frozen != freezing1){//if the layer has a different frozen status
+    if (currl->frozen != freezing1){ //if the layer has a different frozen status
       if(currl->isSnow) {
[…]
