+++ mads_calibration/calibration_files/calibrated/CMT55-Daring Lake Fen/cmt_bgcvegetation.txt
@@ -1,0 +1,37 @@
+//==========================================================
+// CMT55 // Arctic Fen // Daring Creek
+//    Sphagnum  EgreenShrub       PFT2         PFT3         PFT4         PFT5         PFT6         PFT7         PFT8         PFT9 // name: units // description // comment // refs
+      400.0000     400.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // kc: X //  //  // 
+       75.0000      75.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // ki: X //  //  // 
+       -5.0000      -8.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // tmin: X //  //  // 
+        5.5000       5.5000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // toptmin: X //  //  // 
+       20.0000      23.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // toptmax: X //  //  // 
+       35.0000      25.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // tmax: X //  //  // 
+        2.3570       2.3570       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // raq10a0: X //  //  // 
+    -5.308e-02   -5.308e-02       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // raq10a1: X //  //  // 
+     2.384e-03    2.384e-03       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // raq10a2: X //  //  // 
+    -4.110e-05   -4.110e-05       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // raq10a3: X //  //  // 
+        0.0042       0.0042       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // knuptake: X //  //  // 
+        1.0000       0.5002       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // cpartl:  // calculated //  // 
+        0.0000       0.0475       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // cpartw:  // calculated //  // 
+        0.0000       0.4523       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // cpartr:  // calculated //  // 
+       82.6280      35.9011       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // initcnevenl:  // calculated //  // 
+        0.0000      43.4074       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // initcnevenw:  // calculated //  // 
+        0.0000      51.9442       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // initcnevenr:  // calculated //  // 
+       82.6280      35.9011       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // cnevenbl:  // calculated //  // 
+        0.0000      43.4074       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // cnevenbw:  // calculated //  // 
[…]
