+++ mads_calibration/calibration_files/originals/CMT61-Scotty_Creek_Bog/cmt_dimvegetation.txt
@@ -1,0 +1,43 @@
+//==========================================================
+// CMT61 // Scotty Creek Bog // Values from Scotty Creek
+// Sphagnum   EgreenTree   Sedges    EricShrubs      Forbs       PFT5         PFT6         PFT7         PFT8         PFT9 // name: units // description // comment // refs
+1            1            1            1            1            0            0            0            0            0      // vegfrac:  vegetation coverage on land - 0 indicates that PFT doesn't exist 
+0            1            0            1            0            0            0            0            0            0      // ifwoody:  woody plants (1) or not (0) 
+0            0            0            0            0            0            0            0            0            0      // ifdeciwoody:  deciduous woodland (1) or not (0) 
+1            1            1            1            1            0            0            0            0            0      // ifperenial:  perenial (1) or not (0) 
+1            0            0            0            0            0            0            0            0            0      // nonvascular type:  Vascular (0), sphagnum (1), feathermoss (2), lichen (3) 
+0.0290       0.0700       0.0014       0.1200    1.005e-02       0.0000       0.0000       0.0000       0.0000       0.0000 // sla:  specific leaf area (m2/gC) 
+0.1500       0.5500       0.1500       0.3500       0.1500       0.0000       0.0000       0.0000       0.0000       0.0000 // klai:  extinction coeff for conversion from lai to fpc 
+0.0010       0.5000       0.0010       0.0010       0.0010       0.0000       0.0000       0.0000       0.0000       0.0000 // minleaf:  this and the following 4 parameters are for EET/PET controlled phenology - f(leaf) 
+0.7964       0.4289       0.4289       0.4289       0.4289       0.0000       0.0000       0.0000       0.0000       0.0000 // aleaf:  
+0.4663       0.3330       0.3330       0.3330       0.3330       0.0000       0.0000       0.0000       0.0000       0.0000 // bleaf:  
+-2.873e-02       0.3223       0.3223       0.3223       0.3223       0.0000       0.0000       0.0000       0.0000       0.0000 // cleaf:  
+1.6662       1.6662       1.6662       1.6662       1.6662       0.0000       0.0000       0.0000       0.0000       0.0000 // kfoliage:  this and the following 1 parameters are for non-woody plant's f(foliage) 
+-0.0690   -1.571e-03      -0.4800      -0.2470      -1.7210       0.0000       0.0000       0.0000       0.0000       0.0000 // cov:  
+15.2060      19.6200      15.2060      15.2060      15.2060       0.0000       0.0000       0.0000       0.0000       0.0000 // m1:  this and the following 3 parameters are for woody plant's f(foliage) 
+-0.3197     -0.3999      -0.3197      -0.3197      -0.3197       0.0000       0.0000       0.0000       0.0000       0.0000 // m2:  
+0.0401       0.1487       0.0401       0.0401       0.0401       0.0000       0.0000       0.0000       0.0000       0.0000 // m3:  
+0.0001       0.0001       0.0001       0.0001       0.0001       0.0000       0.0000       0.0000       0.0000       0.0000 // m4:  
+0.0000      52.7300      28.0000      38.0000      28.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // frprod[0]:  fine root allocation percentage vertically in soils (every 10 cm) 
+0.0000      14.4100      38.0000      28.0000      38.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // frprod[1]:  fine root allocation percentage vertically in soils (every 10 cm) 
[…]
