+++ mads_calibration/calibration_files/originals/CMT51-Trail Valley Creek Tussock Tundra/cmt_bgcsoil.txt
@@ -1,0 +1,20 @@
+//==========================================================
+// CMT51 // Tussock Tundra TVC // ##Values from Trail Valley###
+      2.0000  // rhq10:  //  //  //
+      0.0000  // moistmin:  //  //  //
+      0.5000  // moistopt:  //  //  //
+      1.0000  // moistmax:  //  //  //
+     43.5500  // lcclnc:  // litterfall C:N //  // Buckeridge etal 2010-mean
+      0.3720  // fsoma:  //  //  //
+      0.6110  // fsompr:  //  //  //
+      0.0170  // fsomcr:  //  //  //
+      0.2570  // som2co2:  //  //  //
+      0.0042  // kn2:  //  //  //
+     20.0000  // nmincnsoil:  // totalC/totalN //  // calculated
+      0.2000  // propftos:  //  //  //
+      0.0000  // fnloss:  // fraction N leaching (0 - 1) when drainage occurs //  // 
+   8000.0000  // initshlwc:  //  //  // from calib targets
+   9629.0000  // initdeepc:  //  //  // from calib targets
+   3.812e+04  // initminec:  //  //  // from calib targets
+   2788.0000  // initsoln:  //  //  // from calib targets
+      0.3920  // initavln:  //  //  // from calib targets
