+++ mads_calibration/calibration_files/originals/CMT66_Old_Jack_Pine/cmt_bgcsoil.txt
@@ -1,0 +1,22 @@
+// dmvdostem parameter file
+// soil microbial activity (BGC) parameters and default initial conditions
+//==========================================================
+// CMT66 // Mature Jack Pine // BOREAS
+      1.6000  // rhq10:  //  //  // 
+      0.0000  // moistmin:  //  //  // 
+      0.5000  // moistopt:  //  //  // 
+      1.0000  // moistmax:  //  //  // 
+    115.0000  // lcclnc:  //  //  // 
+      0.3717  // fsoma:  //  //  // 
+      0.6112  // fsompr:  //  //  // 
+      0.0171  // fsomcr:  //  //  // 
+      0.2571  // som2co2:  //  //  // 
+      0.0042  // kn2:  //  //  // 
+     27.0500  // nmincnsoil:  // C:N totalC/orgN // calculated // 
+      0.2000  // propftos:  //  //  // 
+      0.0000  // fnloss:  // fraction N leaching (0 - 1) when drainage occurs //  // 
+    516.0000  // initshlwc:  // same as targets //  // 
+    758.0000  // initdeepc:  // same as targets //  // 
+   2973.0000  // initminec:  // same as targets //  // 
+    157.0000  // initsoln:  // same as targets //  // 
+      5.4000  // initavln:  // same as targets //  // 

[…]
