+++ mads_calibration/calibration_files/calibrated/CMT50-Trail-Valley-Creek-Shrub-Tundra/cmt_dimvegetation.txt
@@ -1,0 +1,48 @@
+//==========================================================
+// CMT50 // Shrub Tundra TVC // ##Values from Trail Valley###
+//  dvmdostem parameters // v0.7.0-183-gc44a88a1
+//  cmtdescription // ""A long winded description. Spaces? Quotes? Special charachters?"""
+//  calibration site // ""Shrub Tundra TVC"""
+//  references file // refs.bib
+//  DecidShrub  EgreenShrub        Forbs        Sedge         Moss       Lichen         PFT6         PFT7         PFT8         PFT9 // name: units // description 
+        1.0000       1.0000       1.0000       1.0000       1.0000       1.0000       0.0000       0.0000       0.0000       0.0000 // vegfrac:  // vegetation coverage on land - 0 indicates that PFT doesn't exist  
+        1            1            0            0            0            0            0            0            0            0      // ifwoody:  // woody plants (1) or not (0)  
+        1            0            0            0            0            0            0            0            0            0      // ifdeciwoody:  // deciduous woodland (1) or not (0) 
+        1            1            1            1            1            1            0            0            0            0      // ifperenial:  // perennial (1) or not (0) 
+        0            0            0            0            0            3            0            0            0            0      // nonvascular type:  // vascular (0), sphagnum (1), feathermoss (2), lichen (3) 
+        0.0140       0.0103       0.0082       0.0060       0.0200       0.0060       0.0000       0.0000       0.0000       0.0000 // sla: m2/gC // specific leaf area  
+        0.2500       0.1133       0.1580       0.0537       0.2500       0.2500       0.0000       0.0000       0.0000       0.0000 // klai:  // extinction coeff for conversion from lai to fpc 
+        0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // minleaf:  // this and the following 3 parameters are for EET/PET controlled phenology - f(leaf) 
+        0.4289       0.7964       0.7964       0.7964       0.7964       0.7964       0.0000       0.0000       0.0000       0.0000 // aleaf:  
+        0.3330       0.4664       0.4664       0.4664       0.4664       0.4664       0.0000       0.0000       0.0000       0.0000 // bleaf: 
+        0.3223   -2.870e-02   -2.870e-02   -2.870e-02   -2.870e-02   -2.870e-02       0.0000       0.0000       0.0000       0.0000 // cleaf: 
+        1.6662       1.6662       1.6662       1.6662       1.6662       1.6662       0.0000       0.0000       0.0000       0.0000 // kfoliage:  // this and the following 1 parameter are for non-woody plant's f(foliage)
+       -0.0090      -0.4470      -0.3600      -0.5270      -0.0570      -0.2730       0.0000       0.0000       0.0000       0.0000 // cov:  
+       15.2060      15.2060      15.2060      15.2060      15.2060      15.2060       0.0000       0.0000       0.0000       0.0000 // m1:  // this and the following 3 parameters are for woody plant's f(foliage)  
+    -3.197e-01   -3.197e-01   -3.197e-01   -3.197e-01   -3.197e-01   -3.197e-01       0.0000       0.0000       0.0000       0.0000 // m2: 
+        0.0401       0.0401       0.0401       0.0401       0.0401       0.0401       0.0000       0.0000       0.0000       0.0000 // m3:  
+        0.0001       0.0001       0.0001       0.0001       0.0001       0.0001       0.0000       0.0000       0.0000       0.0000 // m4:  
[…]
