+++ mads_calibration/calibration_files/calibrated/CMT51-Trail-Valley-Creek-Tussock-Tundra/cmt_calparbgc.txt
@@ -1,0 +1,21 @@
+//==========================================================
+// CMT51 // Tussock Tundra TVC // ##Values from Trail Valley###
+//    Sedges     Sphagnum       Lichen         PFT3         PFT4         PFT5         PFT6         PFT7         PFT8         PFT9
+  436.780059   284.232584    84.321964     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // cmax: 
+   27.235234     3.013578     2.980000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // nmax: 
+    0.046051     0.045906     0.049602     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // cfall(0): 
+    0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // cfall(1): 
+    0.232827     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // cfall(2): 
+    0.000100     0.000100     0.001000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // nfall(0): 
+    0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // nfall(1): 
+    0.006953     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // nfall(2): 
+   -0.000081    -0.000081    -0.000081     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // kra: 
+   -0.756569    -0.577634    -0.311025     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // krb(0): 
+    0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // krb(1): 
+   -0.990516     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // krb(2): 
+    0.200000     0.200000     0.200000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // frg: 
+0.751000     // micbnup: 
+0.648000     // kdcrawc: 
+0.130000     // kdcsoma: 
+0.029000     // kdcsompr: 
+0.000001     // kdcsomcr: 
