+++ mads_calibration/calibration_files/originals/CMT50-Trail-Valley-Creek-Shrub-Tundra/cmt_bgcsoil.txt
@@ -1,0 +1,20 @@
+//===========================================================
+// CMT50 // Shrub Tundra TVC // Calibrated for Trail Valley Creek Joy and Valeria
+2.0               // rhq10: most values from 4
+0.0               // moistmin:
+0.5               // moistopt:
+1.0               // moistmax: 
+43.55             // lcclnc: Buckeridgeetal2010-mean
+0.372             // fsoma:
+0.611             // fsompr:
+0.017             // fsomcr:
+0.346             // som2co2:
+0.0042            // kn2:
+26.63             // nmincnsoil: calculated
+0.2               // propftos:
+0.0               // fnloss:  fraction N leaching (0 - 1) when drainage occurs
+4000              // initshlwc:
+7944              // initdeepc:
+38121             // initminec:
+1880              // initsoln:
+0.443             // initavln: 4 is 3.993

[…]
