### 2024-10-05 d42b911
Updated comments for Canadian community to match format of all other CMTs
