### 2024-10-07 97c3bab
Corrected the decimal place for the porosity values
