### 2024-10-07 e39c1c1
Correctted the porosity values to the correct decimal place
