### 2024-10-07 bba7fbb
Updated soil parameter value for kdcsomcr
