### 2024-10-08 3b66ff0
Adding inputs and plotting scripts for AM to test
