+++ demo-data/cru-ts40_ar5_rcp85_ncar-ccsm4_bonanzacreeklter_10x10/notes.txt
@@ -1,0 +1,4 @@
+2019-12-12T14:21:58.140828
+=====================
+Modified the vegetation file to set all CMT 08 pixels to CMT 07.
+Saved a copy of the original, just in case.
