### 2024-10-10 5d485cc
Actually implements total CH4POOL outputs
