### 2024-10-14 89a70e2
Starting with where layer_scheme left off
