### 2024-10-15 5d3bbe3
recalibrate black spruce to decrease post-fire recovery rate of moss
