+++ scripts/caribou-poker-exp/SWC_SA/SWC_SA_viz.py
@@ -24,7 +24,7 @@ cell_x_coord = 1
 #cell_x_coord = 0
 
 
-depthlist = [0.05, 0.1, 0.11, 0.2, 0.3]
+depthlist = [0.05, 0.051, 0.1, 0.11, 0.2, 0.3]
 
 def get_lwclayer_tlayer(depthlist, run_dir, var):
     
@@ -109,12 +109,8 @@ out_dir ='/data/workflows/US-Prr_SWC_SA/'
 
 
 run_dirs = [d for d in glob(out_dir+'*/', recursive = True) if 'sample' in d]
-run_dirs = ['/data/workflows/poker_flats/', '/data/workflows/poker_flats_unfrozen_water/', 
-            '/data/workflows/poker_flats_unfrozen_drain/', '/data/workflows/poker_flats_unfrozen_drain_new/']
-samples=['original', 'unfrozen', 'unfrozen_nodrain', 'unfrozen_nodrain_new']
-
-
-run_dirs
+#run_dirs = ['/data/workflows/poker_flats/', '/data/workflows/poker_flats_unfrozen_water/', '/data/workflows/poker_flats_unfrozen_drain_new/']
+samples=['original', 'unfr', 'unfr_runoff']
 
 
 lwc_layers = []
@@ -122,8 +118,9 @@ t_layers = []
 sample_dfs=[]
 for i, d in enumerate(run_dirs):
     if 'GPP_monthly_tr.nc' in os.listdir(os.path.join(d, 'output')):
-        #sample = int(d.split('/')[-2].split('_')[-1])
-        sample=samples[i]
+        print(i)
+        sample = int(d.split('/')[-2].split('_')[-1])
+        #sample=samples[i]
         
         TLAYER = ou.load_trsc_dataframe(var ='TLAYER', timeres='monthly', px_y=cell_y_coord, px_x=cell_x_coord, fileprefix=f'{d}output/')[0]
         LAYERDEPTH = ou.load_trsc_dataframe(var ='LAYERDEPTH', timeres='monthly', px_y=cell_y_coord, px_x=cell_x_coord, fileprefix=f'{d}output/')[0]
@@ -131,9 +128,11 @@ for i, d in enumerate(run_dirs):
         GPP = ou.load_trsc_dataframe(var ='GPP', timeres='monthly', px_y=cell_y_coord, px_x=cell_x_coord, fileprefix=f'{d}output/')[0][0]
         ALD = ou.load_trsc_dataframe(var ='ALD', timeres='yearly', px_y=cell_y_coord, px_x=cell_x_coord, fileprefix=f'{d}output/')[0][0]
         RH = ou.load_trsc_dataframe(var ='RH', timeres='monthly', px_y=cell_y_coord, px_x=cell_x_coord, fileprefix=f'{d}output/')[0][0]
+        #RG = ou.load_trsc_dataframe(var ='RG', timeres='monthly', px_y=cell_y_coord, px_x=cell_x_coord, fileprefix=f'{d}output/')[0][0]
+        #RM = ou.load_trsc_dataframe(var ='RM', timeres='monthly', px_y=cell_y_coord, px_x=cell_x_coord, fileprefix=f'{d}output/')[0][0]
         ALD.columns=['ALD']
         sample_df = pd.DataFrame({'date': GPP.index, 'sample': [sample]*len(GPP), #'LWCLAYER': LWCLAYER, 'TLAYER': TLAYER, 'LAYERDEPTH': LAYERDEPTH, 'LAYERDZ': LAYERDZ,
-                                  'GPP': GPP, 'RH':RH})
[…]
