+++ src/Soil_Bgc.cpp
@@ -589,7 +589,7 @@ void Soil_Bgc::deltac() {
     //HG: 01122023 - this condition allows for winter respiration
     //(Natali et al. 2019, Nature Climate Change)
     if (ed->m_sois.ts[il] <0.) {
-      bd->m_soid.rhmoist[il] = 0.1;
+      bd->m_soid.rhmoist[il] = 0.001;
     } else {
       // Yuan: vwc normalized by total pore - this will allow
       // respiration (methane/oxidation) implicitly

[…]
