### 2024-10-15 a2f7529
Adding s2d and d2m fraction as calpar parameters
