+++ parameters/cmt_calparbgc.txt
@@ -25,6 +25,8 @@
 0.0108046       // kdcsoma:
 0.0020805       // kdcsompr:
 0.0000108       // kdcsomcr:
+1.0             // s2dfraction:
+1.0             // d2mfraction:
 //===========================================================
 // CMT01 // Boreal Black Spruce // 6/29/20 boreal black spruce with Murphy Dome climate
 //EverTree  DecidShrub   DecidTree     Moss         PFT4       PFT5      PFT6     PFT7       PFT8     PFT9       // pftnames:
@@ -47,6 +49,8 @@
 0.54       // kdcsoma:     .58
 0.002       // kdcsompr:     .003
 0.00007   // kdcsomcr: 0.000025  0.00025 bad
+1.0             // s2dfraction:
+1.0             // d2mfraction:
 //===========================================================
 // CMT02 // Boreal White Spruce Forest // 1/22/21 boreal white spruce with Bonanza Creek climate
 //EverTree  DecidShrub   DecidTree     Moss         PFT4       PFT5      PFT6     PFT7       PFT8     PFT9       // pftnames:
@@ -69,6 +73,8 @@
 0.48       // kdcsoma:    .48
 0.0048       // kdcsompr:   .0028  
 0.000025   // kdcsomcr: 
+1.0             // s2dfraction:
+1.0             // d2mfraction:
 //===========================================================
 // CMT03 // Boreal Deciduous Forest // 7/7/20 boreal deciduous forest with Murphy Dome climate
 //EverTree  DecidShrub   DecidTree      Moss        PFT4     PFT5    PFT6    PFT7     PFT8     PFT9    // pftnames:
@@ -91,6 +97,8 @@
 0.58       // kdcsoma:   0.58
 0.003     // kdcsompr:   
 0.000025   // kdcsomcr:
+1.0             // s2dfraction:
+1.0             // d2mfraction:
 //===========================================================
 // CMT04 // Shrub Tundra // Calibrated for Toolik area. J.Clein updated, July 2019
 //Salix      Betula       Decid       EGreen       Sedges       Forbs       Grasses     Lichens      Feather    PFT9     // pftnames:
@@ -113,6 +121,8 @@
 0.93     // kdcsoma:
 0.40     // kdcsompr:
 0.00007  // kdcsomcr:
+1.0             // s2dfraction:
+1.0             // d2mfraction:
 //===========================================================
 // CMT05 // Tussock Tundra // Updated by E.E. calibration, April, May 2017 //Reupdated July 2019 by Joy//Joy June 2022
 // Betula      Decid      EGreen      Sedges       Forbs      Lichens    Feather     Sphag       PFT8         PFT9    // pftnames:
@@ -135,6 +145,8 @@
 0.03      // kdcsoma:  was 0.07
 0.015      // kdcsompr:   was 0.05
 0.000005     // kdcsomcr:  was 0.00005 
+1.0             // s2dfraction:
+1.0             // d2mfraction:
 //===========================================================
 // CMT06 // Wet Sedge Tundra // Calibrated for Toolik area.
 //Decid      Sedges       Grasses      Forbs        Lichens      Feather      Sphag        PFT7     PFT8      PFT9      // pftnames
@@ -157,6 +169,8 @@
 0.011       // kdcsoma:  was 0.12
 0.0025       // kdcsompr: was .03
 0.00025       // kdcsomcr:
+1.0             // s2dfraction:
+1.0             // d2mfraction:
 //===========================================================
 // CMT07 // Heath Tundra //
[…]
