### 2024-10-15 93709d1
Beginning re-write of scheme in Organic.cpp
