### 2024-10-17 bca8a23
Adds SOCFROZEN and SOCUNFROZEN outputs
