### 2024-10-18 a995ef8
CPCRW black spruce calibration
