+++ mads_calibration/run_mads_sensitivity_bona-bs.py
@@ -50,11 +50,12 @@ driver.design_experiment(sample_size, driver.cmtnum,
 #              [0.0001,0.001], [0.001, 0.002], [0.00001, 0.0005]]
 
 #PFT0
-new_bounds = [[15, 30], [-6.0, -3.5], [-8.0, -5.0], [-8.0, -5.0], [0.02, 0.08], [0.001, 0.003], [0.001, 0.005],
-              [0.0001,0.001], [0.0001, 0.001], [0.0001, 0.001]]
+#new_bounds = [[15, 30], [-6.0, -3.5], [-8.0, -5.0], [-8.0, -5.0], [0.02, 0.08], [0.001, 0.003], [0.001, 0.005],
+#              [0.0001,0.001], [0.0001, 0.001], [0.0001, 0.001]]
 
 #GPP
 #new_bounds=[[400, 430], [267, 270], [173,178], [960, 970], [97, 100]]
+new_bounds=[[400, 500], [380, 400], [220, 240], [250, 350], [88, 92]]
 
 
 for i in range(len(driver.params)):
