+++ parameters/cmt_calparbgc.txt
@@ -248,10 +248,10 @@
 //===========================================================
 // CMT15 // Boreal Black Spruce // 11/11/23 boreal black spruce with caribou-poker creek research watershed climate
 //EverTree  DecidShrub   EvrShrub       Moss         Lichen       PFT5      PFT6     PFT7       PFT8     PFT9       // pftnames:
-628.69      267.48       177.65         300.08       98.90        0.0        0.0      0.0       0.0       0.0       // cmax: black spruce 408.69 moss 968.08
+650.0      390.0       230.0         300.08       90.0        0.0        0.0      0.0       0.0       0.0       // cmax: black spruce 408.69 moss 968.08
 16.010259   8.805177    4.294516       5.081049     0.939893       0.0        0.0      0.0       0.0       0.0       // nmax:
-0.0050745    0.010765  0.010611       0.012029     0.012596    0.0        0.0      0.0       0.0       0.0       // cfall(0): bs 0.028745 moss 0.020029
-0.002743    0.008302  0.011543       0.0000       0.0       0.0        0.0      0.0       0.0       0.0       // cfall(1): stem
+0.065745    0.010765  0.010611       0.012029     0.012596    0.0        0.0      0.0       0.0       0.0       // cfall(0): bs 0.028745 moss 0.020029
+0.005    0.008302  0.011543       0.0000       0.0       0.0        0.0      0.0       0.0       0.0       // cfall(1): stem bs 0.002743
 0.004894    0.008712  0.010789       0.0000       0.0       0.0        0.0      0.0       0.0       0.0       // cfall(2): root
 0.000972    0.000110  0.005477       0.00327     0.00778       0.0        0.0      0.0       0.0       0.0       // nfall(0): leaf moss 0.00627
 0.000775    0.000733  0.004037       0.0000       0.0       0.0        0.0      0.0       0.0       0.0       // nfall(1): stem
@@ -262,7 +262,7 @@
 -6.137253   -4.531112  -6.817745       0.00         0.0       0.0        0.0      0.0       0.0       0.0       // krb(2): coeff(1) in maintenance resp.(root)
 0.11        0.11        0.11           0.11         0.11       0.0        0.0      0.0       0.0       0.0       // frg: fraction of available NPP (GPP after rm) for growth resp.
 // soil calibrated parameters
-2.195075       // micbnup: 2.195075 1.73 parameter for soil microbial immobialization of N 
+0.4       // micbnup: 2.195075 1.73 parameter for soil microbial immobialization of N 
 0.804759      // kdcrawc: 0.804759 raw-material (litter) C decompositin rates at reference condition   
 0.710771      // kdcsoma:  0.710771
 0.02847     // kdcsompr: 0.029847     // kdcsompr: 0.06847  
