### 2024-10-23 9d303c2
Outputting heat capacity by layer
