### 2024-10-25 d660295
Adding ecosystem sum function to post processing
