### 2024-10-25 137fc4b
current changes for apparent heat capacity in progress
