+++ src/SeismicLine.cpp
@@ -554,51 +554,51 @@ double WildFire::getBurnOrgSoilthick(const int year) {
   //  4 = high + high surface
 
   double total_organic_thickness =  cd->m_soil.mossthick
-                                    + cd->m_soil.shlwthick
-                                    + cd->m_soil.deepthick ;
+                                    + cd->m_soil.shlwthick;
+                                    // + cd->m_soil.deepthick ;
 
   BOOST_LOG_SEV(glg, debug) << "Total organic thickness: " << total_organic_thickness;
 
   //compute the fraction of OL to burn FOLB
   double folb = 0.0;
 
-  if ( this->fri_derived ) {                    // FRI-derived fire regime
-    if (this->fri_severity >= 0) {              // fire severity is available from the input files - so get folb from the parameter file;
-      folb = firpar.foslburn[this->fri_severity];
-    }else {                                     // fire severity is available from the input files - apply the lookup table from Yi et al. 2010;
-      if( cd->drainage_type == 0 ) {            // 0: well-drained; 1: poorly-drained;
-        if ( this->fri_jday_of_burn <= 212 ) {   // Early fire, before July 31st (from Turetsly et al. 2011);
-          if ( this->fri_area_of_burn < 1.0 ) { // Small fire year (less that 1% of the area burned);
-            folb = 0.54;
-          } else {
-            folb = 0.69;
-          }
-        } else {                                // late fire (after July 31st);
-          folb = 0.80;
-        } 
-      } else {                                  // lowland;
-        folb = 0.48;
-      } 
-    }
-  } else {                                      // Explicit fire regime;
-    if (this->exp_fire_severity[year] >= 0) {    // fire severity is available from the input files - so get folb from the parameter file;
-      folb = firpar.foslburn[this->exp_fire_severity[year]];
-    } else {  
-      if ( cd->cmttype > 3) {                   // tundra ecosystems: Mack et al. 2011;
-        folb = 0.01*((21.5-6.1)/21.5);
-      } else {                                  // boreal forest: Genet et al.2013;
-         if(this->slope<2){
-          // AOB in km-2, see coefficient and paper.
-          folb = 0.1276966713-0.0319397467*this->slope+0.0020914862*this->exp_jday_of_burn[year]+0.0127016155* log(this->exp_area_of_burn[year]);
-        } else {
-          folb = -0.2758306315+0.0117609336*this->slope-0.0744057680*cos(this->asp * 3.14159265 / 180 ) +0.0260221684*edall->m_soid.tshlw+0.0011413114*this->exp_jday_of_burn[year]+0.0336302905*log (this->exp_area_of_burn[year]);
-        }
-      }
-    }
-  }
+  // if ( this->fri_derived ) {                    // FRI-derived fire regime
+  //   if (this->fri_severity >= 0) {              // fire severity is available from the input files - so get folb from the parameter file;
+  //     folb = firpar.foslburn[this->fri_severity];
+  //   }else {                                     // fire severity is available from the input files - apply the lookup table from Yi et al. 2010;
+  //     if( cd->drainage_type == 0 ) {            // 0: well-drained; 1: poorly-drained;
+  //       if ( this->fri_jday_of_burn <= 212 ) {   // Early fire, before July 31st (from Turetsly et al. 2011);
+  //         if ( this->fri_area_of_burn < 1.0 ) { // Small fire year (less that 1% of the area burned);
[…]
