### 2024-10-29 #7136 @tobeycarman
@kthyng I haven't forgotten. Just back from some leave time and will be working on this. Your feedback/this issue prompted some re-organization and, as suggested, documentation writing that we will commit in addition to updates to the paper.
