### 2024-10-30 470274f
Adding GIPL framework and adjusting temperature thresholds
