+++ src/Layer.cpp
@@ -87,9 +87,10 @@ double Layer::getHeatCapacity() { // volumetric heat capacity
     hcap = getFrzVolHeatCapa();
   }
 
-  this->hcapa = hcap + DENLIQ * LHFUS * getDeltaUnfVolLiq();
+  this->hcapa = hcap + (DENLIQ * LHFUS) * fapp_hcap();
 
   return hcap;
+
 };
 
 double Layer::getThermalConductivity() {
@@ -121,81 +122,30 @@ double Layer::getUnfVolLiq(){
     uwc = 0.0;
   }
 
-  uwc = fmin(uwc, poro-getVolLiq());
+  uwc = fmin(uwc, getVolWater());
 
   if (isSnow){
     uwc = 0.0;
   }
 
-  return uwc;
+  return 0.0;
 }
 
-// double Layer::getDeltaUnfVolLiq(){
-
-//   double d_uwc;
-
-//   if (tem < temp_dep){
-//     d_uwc = -b_parameter * tem * poro * pow(abs(temp_dep), b_parameter) * pow(abs(tem), -b_parameter - 2);
-//   } else {
-//     d_uwc = 0.0;
-//   }
-
-//   if (isSnow){
-//     d_uwc = 0.0;
-//   }
-
-//   return d_uwc;
-// }
-
 double Layer::getDeltaUnfVolLiq(){
-  // Following numerical formulation in GIPL2.0
-  // based on Nicolsky et al. 2007
 
-  double T1, T2, W1, W2, H, D = 0.0;
-  double fhcap=0.0;
+  double d_uwc;
 
-  
-
-  if (isSoil){
-    if (!prevl || !prevl->isSoil){
-      T1 = tem;
-      T2 = (nextl->tem + tem) / 2.0;
-      W1 = getUnfVolLiq();
-      W2 = nextl->getUnfVolLiq();
-      D = 1.0;
-    } else if (prevl->isSoil && nextl->isSoil){
-      T1 = (prevl->tem + tem) / 2.0;
-      T2 = tem;
-      W1 = prevl->getUnfVolLiq();
-      W2 = getUnfVolLiq();
-      D = dz / (prevl->dz + dz);
-    } else if (prevl->isSoil && nextl->isRock){
-      T1 = (prevl->tem + tem) / 2.0;
-      T2 = tem;
-      W1 = prevl->getUnfVolLiq();
-      W2 = getUnfVolLiq();
-      D = 1.0;
-    } else{
-      T1 = 1.0;
-      T2 = 1.0;
-      W1 = 0.0;
-      W2 = 0.0;
-      D = 0.0;
-    }
-  }
-
-  if (abs(T1 - T2) < 1e-6){
-      fhcap = 0.5 * (W1 + W2) * D;
+  if (tem < temp_dep){
+    d_uwc = -b_parameter * tem * poro * pow(abs(temp_dep), b_parameter) * pow(abs(tem), -b_parameter - 2);
   } else {
-      H = 1 / (T1 - T2);
-      fhcap = H * (W1 - W2) * D;
+    d_uwc = 0.0;
   }
 
-  if (!isSoil){
-      fhcap = 0.0;
+  if (isSnow){
+    d_uwc = 0.0;
   }
 
-  return fhcap;
+  return 0.0;
 }
 
 double Layer::getVolWater(){
@@ -232,7 +182,7 @@ double Layer::getVolIce() {
       vice = fmax(0.0, vice);
     }
     else{
-      vice = fmin((double)vice, (double)poro);
+      vice = fmin((double)vice, (double)maxice);
     }
 
     return vice;
@@ -254,6 +204,7 @@ double Layer::getVolLiq() {
     }
     else{
[…]
