### 2024-11-01 289742b
Removes the time dimension from the DA statefile
