+++ src/DAController.cpp
@@ -214,32 +214,32 @@ void DAController::run_DA_LAI(timestep_id current_step){
   //Write accessory variables to file
   //VEGC
   std::array<std::array<double, NUM_PFT>, NUM_PFT_PART> vegc = cohort->get_vegc_pftandcomp_monthly();
-  temutil::output_nc_5dim(&this->vegc_outspec, ".nc", &curr_coords, &vegc[0][0], NUM_PFT_PART, NUM_PFT, 0, 1);
+  temutil::output_nc_4dim_notime(&this->vegc_outspec, ".nc", &curr_coords, &vegc[0][0], NUM_PFT_PART, NUM_PFT);
 
   //STRN
   std::array<std::array<double, NUM_PFT>, NUM_PFT_PART> strn = cohort->get_strn_pftandcomp_monthly();
-  temutil::output_nc_5dim(&this->strn_outspec, ".nc", &curr_coords, &strn[0][0], NUM_PFT_PART, NUM_PFT, 0, 1);
+  temutil::output_nc_4dim_notime(&this->strn_outspec, ".nc", &curr_coords, &strn[0][0], NUM_PFT_PART, NUM_PFT);
 
   //LWC
-  temutil::output_nc_4dim(&this->lwc_outspec, ".nc", &curr_coords, &cohort->edall->m_soid.lwc[0], MAX_SOI_LAY, 0, 1);
+  temutil::output_nc_3dim_notime(&this->lwc_outspec, ".nc", &curr_coords, &cohort->edall->m_soid.lwc[0], MAX_SOI_LAY);
 
   //SOMRAWC
-  temutil::output_nc_4dim(&this->rawc_outspec, ".nc", &curr_coords, &cohort->bdall->m_sois.rawc[0], MAX_SOI_LAY, 0, 1);
+  temutil::output_nc_3dim_notime(&this->rawc_outspec, ".nc", &curr_coords, &cohort->bdall->m_sois.rawc[0], MAX_SOI_LAY);
 
   //SOMA
-  temutil::output_nc_4dim(&this->soma_outspec, ".nc", &curr_coords, &cohort->bdall->m_sois.soma[0], MAX_SOI_LAY, 0, 1);
+  temutil::output_nc_3dim_notime(&this->soma_outspec, ".nc", &curr_coords, &cohort->bdall->m_sois.soma[0], MAX_SOI_LAY);
 
   //SOMPR
-  temutil::output_nc_4dim(&this->sompr_outspec, ".nc", &curr_coords, &cohort->bdall->m_sois.sompr[0], MAX_SOI_LAY, 0, 1);
+  temutil::output_nc_3dim_notime(&this->sompr_outspec, ".nc", &curr_coords, &cohort->bdall->m_sois.sompr[0], MAX_SOI_LAY);
 
   //SOMCR
-  temutil::output_nc_4dim(&this->somcr_outspec, ".nc", &curr_coords, &cohort->bdall->m_sois.somcr[0], MAX_SOI_LAY, 0, 1);
+  temutil::output_nc_3dim_notime(&this->somcr_outspec, ".nc", &curr_coords, &cohort->bdall->m_sois.somcr[0], MAX_SOI_LAY);
 
   //LAYERDZ (for reference, not to be updated by DA)
-  temutil::output_nc_4dim(&this->layerdz_outspec, ".nc", &curr_coords, &cohort->cd.m_soil.dz[0], MAX_SOI_LAY, 0, 1);
[…]
