### 2024-11-04 fc84a7d
Removing snow layer and adjusting burn to mimick seismic line
