### 2024-11-04 5e1ad08
Add sphinx extension for handling references/bibtex, move refs to bibtex file.
Used mendeley and the web importer to convert existing refs page to
bibtex. There were two entries that had to be done by hand (Jordan and
Beringer).

Then from Mendeley Reference Manager, export to bibtex format, save in
the bibliography.bib file.

There are problems with non-unique citation keys will address in future
commits.
