### 2024-11-04 6701bd9
fix duplicate citation keys.
