+++ docs_src/sphinx/source/bibliography.bib
@@ -88,7 +88,7 @@
    volume = {108},
    year = {2003},
 }
-@article{Zhuang2003,
+@article{Zhuang2003a,
    abstract = {There is substantial evidence that soil thermal dynamics are changing in terrestrial ecosystems of the Northern Hemisphere and that these dynamics have implications for the exchange of carbon betwe...},
    author = {Q. Zhuang and A. D. McGuire and J. M. Melillo and J. S. Clein and R. J. Dargaville and D. W. Kicklighter and R. B. Myneni and J. Dong and V. E. Romanovsky and J. Harden and J. E. Hobbie},
    doi = {10.3402/TELLUSB.V55I3.16368},
@@ -305,7 +305,7 @@
    title = {The effects of land cover and land use change on the contemporary carbon balance of the arctic and boreal terrestrial ecosystems of Northern Eurasia},
    year = {2011},
 }
-@article{Hayes2011,
+@article{Hayes2011a,
    abstract = {Studies indicate that, historically, terrestrial ecosystems of the northern high-latitude region may have been responsible for up to 60% of the global net land-based sink for atmospheric CO 2. However, these regions have recently experienced remarkable modification of the major driving forces of the carbon cycle, including surface air temperature warming that is significantly greater than the global average and associated increases in the frequency and severity of disturbances. Whether Arctic tundra and boreal forest ecosystems will continue to sequester atmospheric CO 2 in the face of these dramatic changes is unknown. Here we show the results of model simulations that estimate a 41 Tg C yr -1 sink in the boreal land regions from 1997 to 2006, which represents a 73% reduction in the strength of the sink estimated for previous decades in the late 20th century. Our results suggest that CO 2 uptake by the region in previous decades may not be as strong as previously estimated. The recent decline in sink strength is the combined result of (1) weakening sinks due to warming-induced increases in soil organic matter decomposition and (2) strengthening sources from pyrogenic CO 2 emissions as a result of the substantial area of boreal forest burned in wildfires across the region in recent years. Such changes create positive feedbacks to the climate system that accelerate global warming, putting further pressure on emission reductions to achieve atmospheric stabilization targets. Copyright 2011 by the American Geophysical Union.},
    author = {D. J. Hayes and A. D. McGuire and D. W. Kicklighter and K. R. Gurney and T. J. Burnside and J. M. Melillo},
    doi = {10.1029/2010GB003813},
@@ -403,7 +403,7 @@
    volume = {19},
    year = {2009},
 }
-@article{Euskirchen2014,
+@article{Euskirchen2014a,
[…]
