### 2024-11-04 bd77ca0
Add content from JOSS paper draft for soil and vegetation sections.
