### 2024-11-06 726b59c
re-format C and N to use latex math style
