+++ docs_src/sphinx/source/model_overview.rst
@@ -132,11 +132,11 @@ productivity (GPP) for each PFT. GPP is a function of foliage development
 (seasonal and successional patterns), air and soil temperature, water and
 nutrient availability, photosynthetically active radiation, and maximum
 assimilation rate (a calibrated parameter) (:cite:p:`McGuire1992interactions`;
-:cite:p:`Euskirchen2009changes`). Changes in vegetation C stocks are calculated
-using GPP, autotrophic respiration (Ra), and litterfall (transfer from
-vegetation to soil). Vegetation N stocks are calculated using plant N uptake and
-litterfall. Vegetation C and N stocks may also be modified as a result of
-wildfire burn.
+:cite:p:`Euskirchen2009changes`). Changes in vegetation :math:`C` stocks are
+calculated using GPP, autotrophic respiration (Ra), and litterfall (transfer
+from vegetation to soil). Vegetation :math:`N` stocks are calculated using plant
+:math:`N` uptake and litterfall. Vegetation :math:`C` and :math:`N` stocks may
+also be modified as a result of wildfire burn.
 
 .. raw:: html
 
@@ -159,11 +159,12 @@ Richards equation to predict soil moisture dynamics in the unfrozen layers
 :cite:p:`Zhuang2003modeling`). Snow is also represented with a dynamic stack of
 layers. The physical properties of the snowpack (density, thickness, and
 temperature) are calculated from snowfall, sublimation and snowmelt. Snow cover
-influences soil-thermal and hydrological seasonal dynamics. Changes in soil C
-stocks are a result of litterfall from the vegetation and decomposition of soil
-C stocks by microbes (heterotrophic respiration or Rh). Changes in soil organic
-and available N stocks are a result of litterfall, net mineralization of organic
-N, and plant N uptake. Soil organic layers and soil C and N stocks may also be
+influences soil-thermal and hydrological seasonal dynamics. Changes in soil
+:math:`C` stocks are a result of litterfall from the vegetation and
+decomposition of soil :math:`C` stocks by microbes (heterotrophic respiration or
+Rh). Changes in soil organic and available :math:`N` stocks are a result of
+litterfall, net mineralization of organic :math:`N` , and plant :math:`N`
+uptake. Soil organic layers and soil :math:`C` and :math:`N` stocks may also be
 modified due to wildfire.
 
 
