### 2024-11-07 fa6324f
Sort .bib on citation keys...
...and various other cleanup: sort fields, remove empties, etc.

used web app: https://flamingtempura.github.io/bibtex-tidy/index.html
