+++ docs_src/sphinx/source/bibliography.bib
@@ -1,470 +1,470 @@
-@misc{Jordan1991onedimensional,
-   author = {Rachel E. Jordan},
-   journal = {Special Report 91-16},
-   keywords = {Computer models,Mathematical models,SNTHERM.89 (Computer program),Snow,Snow covers,Thermal properties},
-   publisher = {Cold Regions Research and Engineering Laboratory (U.S.)},
-   title = {A One-dimensional temperature model for a snow cover : technical documentation for SNTHERM.89},
-   url = {http://hdl.handle.net/11681/11677},
-   year = {1991},
-}
-@article{Beringer2001representation,
-   author = {Jason Beringer and Amanda H Lynch and F Stuart Chapin and Michelle Mack and Gordon B Bonan},
-   city = {Boston MA, USA},
-   doi = {10.1175/1520-0442(2001)014<3324:TROASI>2.0.CO;2},
-   issue = {15},
-   journal = {Journal of Climate},
-   pages = {3324 - 3335},
-   publisher = {American Meteorological Society},
-   title = {The Representation of Arctic Soils in the Land Surface Model: The Importance of Mosses},
-   volume = {14},
-   url = {https://journals.ametsoc.org/view/journals/clim/14/15/1520-0442_2001_014_3324_troasi_2.0.co_2.xml},
-   year = {2001},
-}
-@article{Yuan2012assessment,
-   abstract = {Carbon (C) dynamics of boreal forest ecosystems have substantial implications for efforts to mitigate the rise of atmospheric CO2 and may be substantially influenced by warming and changing wildfire regimes. In this study we applied a large-scale ecosystem model that included dynamics of organic soil horizons and soil organic matter characteristics of multiple pools to assess forest C stock changes of the Yukon River Basin (YRB) in Alaska, USA, and Canada from 1960 through 2006, a period characterized by substantial climate warming and increases in wildfire. The model was calibrated for major forests with data from long-term research sites and evaluated using a forest inventory database. The regional assessment indicates that forest vegetation C storage increased by 46 Tg C, but that total soil C storage did not change appreciably during this period. However, further analysis suggests that C has been continuously lost from the mineral soil horizon since warming began in the 1970s, but has increased in the amorphous organic soil horizon. Based on a factorial experiment, soil C stocks would have increased by 158 Tg C if the YRB had not undergone warming and changes in fire regime. The analysis also identified that warming and changes in fire regime were approximately equivalent in their effects on soil C storage, and interactions between these two suggests that the loss of organic horizon thickness associated with increases in wildfire made deeper soil C stocks more vulnerable to loss via decomposition. Subbasin analyses indicate that C stock changes were primarily sensitive to the fraction of burned forest area within each subbasin and that boreal forest ecosystems in the YRB are currently transitioning from being sinks to sources at ;0.7% annual area burned. We conclude that it is important for international mitigation efforts focused on controlling atmospheric CO2 to consider how climate warming and changes in fire regime may concurrently affect the CO 2 sink strength of boreal forests. It is also important for large-scale biogeochemical and earth system models to include organic soil dynamics in applications to assess regional C dynamics of boreal forests responding to warming and changes in fire regime. © 2012 by the Ecological Society of America.},
-   author = {F. M. Yuan and S. H. Yi and A. D. Mcguire and K. D. Johnson and J. Liang and J. W. Harden and E. S. Kasischke and W. A. Kurz},
-   doi = {10.1890/11-1957.1},
-   issn = {10510761},
-   issue = {8},
-   journal = {Ecological Applications},
-   keywords = {Assessment,Basin-scale analysis,Boreal forests,C stock dynamics,Climate warming,Dynamic organic soil,Fire regime change,Terrestrial ecological modeling},
-   month = {12},
-   pages = {2091-2109},
-   pmid = {23387112},
[…]
