### 2024-11-07 3e87934
Change cite key format to use lower case first letter.
