+++ docs_src/sphinx/source/bibliography.bib
@@ -1,4 +1,4 @@
-@article{Balshi2007role,
+@article{balshi2007role,
   title        = {The role of historical fire disturbance in the carbon dynamics of the pan-boreal region: A process-based analysis},
   author       = {M. S. Balshi and A. D. McGuire and Q. Zhuang and J. Melillo and D. W. Kicklighter and E. Kasischke and C. Wirth and M. Flannigan and J. Harden and J. S. Clein and T. J. Burnside and J. McAllister and W. A. Kurz and M. Apps and A. Shvidenko},
   year         = 2007,
@@ -11,7 +11,7 @@
   abstract     = {Wildfire is a common occurrence in ecosystems of northern high latitudes, and changes in the fire regime of this region have consequences for carbon feedbacks to the climate system. To improve our understanding of how wildfire influences carbon dynamics of this region, we used the process-based Terrestrial Ecosystem Model to simulate fire emissions and changes in carbon storage north of 45°N from the start of spatially explicit historically recorded fire records in the twentieth century through 2002, and evaluated the role of fire in the carbon dynamics of the region within the context of ecosystem responses to changes in atmospheric CO2 concentration and climate. Our analysis indicates that fire plays an important role in interannual and decadal scale variation of source/sink relationships of northern terrestrial ecosystems and also suggests that atmospheric CO2 may be important to consider in addition to changes in climate and fire disturbance. There are substantial uncertainties in the effects of fire on carbon storage in our simulations. These uncertainties are associated with sparse fire data for northern Eurasia, uncertainty in estimating carbon consumption, and difficulty in verifying assumptions about the representation of fires that occurred prior to the start of the historical fire record. To improve the ability to better predict how fire will influence carbon storage of this region in the future, new analyses of the retrospective role of fire in the carbon dynamics of northern high latitudes should address these uncertainties. Copyright 2007 by the American Geophysical Union.},
   issue        = 2
 }
-@article{Balshi2009assessing,
+@article{balshi2009assessing,
   title        = {Assessing the response of area burned to changing climate in western boreal North America using a Multivariate Adaptive Regression Splines (MARS) approach},
   author       = {Michael S. Balshi and A. David McGuire and Paul Duffy and Mike Flannigan and John Walsh and Jerry Melillo},
   year         = 2009,
@@ -24,7 +24,7 @@
   issue        = 3,
   keywords     = {Boreal forest,Climate change,Fire,Future area burned,Multivariate Adaptive Regression Splines}
 }
-@article{Balshi2009vulnerability,
+@article{balshi2009vulnerability,
   title        = {Vulnerability of carbon storage in North American boreal forests to wildfires during the 21st century},
   author       = {Michael S. Balshi and A. D. Mcguire and P. Duffy and M. Flannigan and D. W. Kicklighter and J. Melillo},
   year         = 2009,
@@ -37,7 +37,7 @@
   issue        = 6,
   keywords     = {Boreal carbon dynamics,Climate change,Fire emissions}
 }
-@article{Beringer2001representation,
+@article{beringer2001representation,
   title        = {The Representation of Arctic Soils in the Land Surface Model: The Importance of Mosses},
   author       = {Jason Beringer and Amanda H Lynch and F Stuart Chapin and Michelle Mack and Gordon B Bonan},
   year         = 2001,
@@ -50,7 +50,7 @@
   city         = {Boston MA, USA},
   issue        = 15
 }
-@article{Decker2009impact,
+@article{decker2009impact,
   title        = {Impact of Modified Richards Equation on Global Soil Moisture Simulation in the Community Land Model (CLM3.5)},
[…]
