+++ mads_calibration/calibration_files/calibrated/CMT55-Daring Lake Fen/cmt_calparbgc.txt
@@ -1,21 +1,0 @@
-//==========================================================
-// CMT55 // Arctic Fen // Daring Creek
-//  Sphagnum  EgreenShrub         PFT2         PFT3         PFT4         PFT5         PFT6         PFT7         PFT8         PFT9
-  276.386600    96.488860     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // cmax: 
-    3.604999     0.942515     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // nmax: 
-    0.077788     0.035897     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // cfall(0): 
-    0.000000     0.025385     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // cfall(1): 
-    0.000000     0.007820     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // cfall(2): 
-    0.004675     0.002299     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // nfall(0): 
-    0.000000     0.001848     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // nfall(1): 
-    0.000000     0.030999     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // nfall(2): 
-   -0.000081    -0.000081     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // kra: 
-   -2.189131    -3.305160     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // krb(0): 
-    0.000000    -7.405390     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // krb(1): 
-    0.000000    -3.801470     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // krb(2): 
-    0.200000     0.200000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000     0.000000 // frg: 
-0.514619     // micbnup: 
-0.195884     // kdcrawc: 
-0.043923     // kdcsoma: 
-0.000197     // kdcsompr: 
-0.000005     // kdcsomcr: 
