+++ mads_calibration/calibration_files/calibrated/CMT55-Daring Lake Fen/cmt_dimvegetation.txt
@@ -1,44 +1,0 @@
-//==========================================================
-// CMT55 // Arctic Fen // Daring Creek
-//Sphagnum  EgreenShrub  PFT2   PFT3      PFT4         PFT5      PFT6        PFT7         PFT8     PFT9    // name: units // description // comment // refs
-  1          1          0         0          0            0         0           0            0        0      // vegfrac:  vegetation coverage on land - 0 indicates that PFT doesn't exist //  //  //
-  0          1          0         0          0            0         0           0            0        0      // ifwoody:  woody plants (1) or not (0) //  //  // 
-  0          0          0         0          0            0         0           0            0        0      // ifdeciwoody:  deciduous woodland(1) or not (0) //  //  // 
-  1          1          0         0          0            0         0           0           0         0      // ifperenial:  perenial (1) or not (0) //  //  // 
-  1          0          0         0          0            0         0           0           0         0      // nonvascular type:  Vascular (0), sphagnum (1), feathermoss (2), lichen (3) //  //  // 
-0.0072       0.0060       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // sla:  specific leaf area (m2/gC) //  //  // 
-0.1133       0.1133       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // klai:  extinction coeff for conversion from lai to fpc //  //  // 
-0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // minleaf:  this and the following 4 parameters are for EET/PET controlled phenology - f(leaf) //  //  // 
-0.7964       0.7964       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // aleaf:  //  //  // 
-0.4663       0.4663       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // bleaf:  //  //  // 
--2.873e-02   -2.873e-02   0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // cleaf:  //  //  // 
-1.6662       1.6662       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // kfoliage:  this and the following 1 parameters are for non-woody plant's f(foliage) //  //  // 
--0.2320      -0.4470      0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // cov:  //  //  // 
-15.2060      15.2060      0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // m1:  this and the following 3 parameters are for woody plant's f(foliage) //  //  // 
--3.197e-01   -3.197e-01   0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // m2:  //  //  // 
-0.0401       0.0401       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // m3:  //  //  // 
-0.0001       0.0001       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // m4:  //  //  // 
-100.00      35.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // frprod[0]:  fine root allocation percentage vertically in soils (every 10 cm) //  //  // 
[…]
