+++ mads_calibration/calibration_files/calibrated/CMT55-Daring Lake Fen/cmt_envcanopy.txt
@@ -1,16 +1,0 @@
-//==========================================================
-// CMT55 // Arctic Fen // Daring Creek
-// Sphagnum  EgreenShrub    PFT2         PFT3         PFT4         PFT5         PFT6         PFT7         PFT8         PFT9 // name: units // description // comment // refs
-0.1000       0.1000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // albvisnir:  canopy albedo //  //  // 
-0.5000       0.5000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // er:  light extinction coefficient //  //  // 
-0.0410       0.0410       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // ircoef:  rainfall interception //  //  // 
-0.0410       0.0410       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // iscoef:  snow interception //  //  // 
-0.0030       0.0030       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // glmax:  max. canopy conductance (m/s) //  //  // 
-0.0800       0.0800       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // gl_bl:  boundary layer conductance //  //  // 
-0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // gl_c:   cuticular conductance //  //  // 
-930.000    930.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // vpd_open:   vpd (pa) for leaf stomata fully openness //  //  // 
-4100.00    4100.000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // vpd_close:  vpd (pa) for leaf stomata fully closure //  //  // 
-75.0000      75.000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // ppfd50:  //  //  // 
-0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // initvegwater:  initial intercepted rain water in canopy //  //  // 
-0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000       0.0000 // initvegsnow:  initial intercepted snow water in canopy //  //  //
-
