+++ mads_calibration/calibration_files/originals/CMT50-Trail Valley Creek Shrub Tundra/cmt_dimground.txt
@@ -1,21 +1,0 @@
-//==========================================================
-// CMT50 // Shrub Tundra TVC // ##Values from Trail Valley###
-//  dvmdostem parameters // v0.7.0-183-gc44a88a1
-//  cmtdescription // ""A long winded description. Spaces? Quotes? Special charachters?"""
-//  calibration site // ""Shrub Tundra TVC"""
-//  references file // refs.bib
-    300      // snwdenmax: kg/m3 // max. snow density // Wang 2022
-    230     // snwdennew: kg/m3 // new (min.) snow density // Marsh and Pomeroy 1996
-      0.0000  // initsnwthick: m // initial snow thickness 
-      0.0000  // initsnwdense: kg/m3 // initial snow density 
-      0.0350  // maxmossthick: m // max. dead moss thickness 
-      0.0050  // initdmossthick: m // initial dead moss thickness(not using anymore) 
-      3       // mosstype:  // moss type (1:sphagnum-2:feathermoss-3:other) 
-      0.1000  // initfibthick: m //Krogh etal 2017
-      0.1650  // inithumthick: m // Voigt etal 2023
-      0.0652  // coefshlwa: 
-      0.7080  // coefshlwb: 
-      0.0545  // coefdeepa: 
-      0.9750  // coefdeepb: 
-      0.0966  // coefminea: 
-      0.7851  // coefmineb:
