+++ mads_calibration/calibration_files/originals/CMT50-Trail Valley Creek Shrub Tundra/cmt_envground.txt
@@ -1,43 +1,0 @@
-//===========================================================
-// CMT50 // Shrub Tundra TVC // ##Values from Trail Valley###
-      0.8000  // snwalbmax:  
-      0.4000  // snwalbmin:  
-  -1.500e+07  // psimax:  
-      0.0800  // evapmin:  
-      0.0400  // drainmax:  
-      0.2000  // rtdp4gdd:  // root depth for estimating growing degree day 
-      0.2500  // tcsolid(m): W/mK // Soil thermal conductivity for moss 
-      0.3300  // tcsolid(f): W/mK // Soil thermal conductivity for fibric // averaged across first 0.20m of organic layer 
-      0.4000  // tcsolid(h): W/mK // Soil thermal conductivity for humic 
-      0.9800  // porosity(m):  // porosity for moss layers 
-      0.9200  // porosity(f):  // porosity for fibric layers // averaged across first 0.20m of organic layer 
-      0.8000  // porosity(h):  // porosity for humic layers 
-   1.400e+05  // bulkden(m): g/m3 // bulk density for moss 
-   2.600e+05  // bulkden(f): g/m3 // bulk density for fibric // averaged across first 0.20m of organic layer 
-   7.700e+05  // bulkden(h): g/m3 // bulk density for humic 
-      0.1500  // hksat(m): mm/s // hydraulic conductivity at saturation for moss 
-      0.0010  // hksat(f): mm/s // hydraulic conductivity at saturation for fibric // averaged across first 0.20m of organic layer 
-      0.0001  // hksat(h): mm/s // hydraulic conductivity at saturation for humic 
-      1.5000  // nfactor(s):  // Summer nfactor 
-      1.0000  // nfactor(w):  // Winter nfactor 
-     -1.0000  // initsnwtem: deg_C // initial snow temperature //  // fromcmt4
-     -1.0000  // initts(0): deg_C // initial soil temperature at 10 cm interval 
-     -1.0000  // initts(1): deg_C // initial soil temperature at 10 cm interval 
-     -1.0000  // initts(2): deg_C // initial soil temperature at 10 cm interval 
-     -1.0000  // initts(3): deg_C // initial soil temperature at 10 cm interval 
-     -1.0000  // initts(4): deg_C // initial soil temperature at 10 cm interval 
-     -1.0000  // initts(5): deg_C // initial soil temperature at 10 cm interval 
-     -1.0000  // initts(6): deg_C // initial soil temperature at 10 cm interval 
-     -1.0000  // initts(7): deg_C // initial soil temperature at 10 cm interval 
-     -1.0000  // initts(8): deg_C // initial soil temperature at 10 cm interval 
-     -1.0000  // initts(9): deg_C // initial soil temperature at 10 cm interval
-      0.1000  // initvwc(0):  // initial soil volume water content at every 10 cm interval
-      0.1000  // initvwc(1):  // initial soil volume water content at every 10 cm interval 
-      0.1000  // initvwc(2):  // initial soil volume water content at every 10 cm interval
-      0.1000  // initvwc(3):  // initial soil volume water content at every 10 cm interval 
-      0.1000  // initvwc(4):  // initial soil volume water content at every 10 cm interval
-      0.1000  // initvwc(5):  // initial soil volume water content at every 10 cm interval 
-      0.1000  // initvwc(6):  // initial soil volume water content at every 10 cm interval
-      0.1000  // initvwc(7):  // initial soil volume water content at every 10 cm interval
-      0.1000  // initvwc(8):  // initial soil volume water content at every 10 cm interval 
-      0.1000  // initvwc(9):  // initial soil volume water content at every 10 cm interval
