+++ mads_calibration/calibration_files/originals/CMT60_Scotty_Creek_Peat_Plateau/cmt_firepar.txt
@@ -1,21 +1,0 @@
-//==========================================================
-// CMT60 // PEAT PLATEAU // 
-//    EverTree    EricShrub       Lichen         Moss         PFT4         PFT5         PFT6         PFT7         PFT8         PFT9 // name: units // description // comment // refs
-0.00         0.00        0.00        0.00         0.00         0.00         0.00         0.00         0.00         0.00      // fvcomb_sev1:  //  // fraction of PFT vegetation combusted for severity 1 // 
-0.15         0.15        0.15        0.15         0.00         0.00         0.00         0.00         0.00         0.00      // fvcomb_sev2:  //  // fraction of PFT vegetation combusted for severity 2 // 
-0.25         0.25        0.25        0.25         0.00         0.00         0.00         0.00         0.00         0.00      // fvcomb_sev3:  //  // fraction of PFT vegetation combusted for severity 3 // 
-0.30         0.30        0.30        0.30         0.00         0.00         0.00         0.00         0.00         0.00      // fvcomb_sev4:  //  // fraction of PFT vegetation combusted for severity 4 // 
-0.35         0.35        0.35        0.35         0.00         0.00         0.00         0.00         0.00         0.00      // fvcomb_sev5:  //  // fraction of PFT vegetation combusted for severity 5 // 
-0.00         0.00        0.00        0.00         0.00         0.00         0.00         0.00         0.00         0.00      // fvdead_sev1:  //  // fraction of PFT vegetation killed for severity 1 // 
-0.50         0.50        0.50        0.50         0.00         0.00         0.00         0.00         0.00         0.00      // fvdead_sev2:  //  // fraction of PFT vegetation killed for severity 2 // 
-0.60         0.60        0.60        0.60         0.00         0.00         0.00         0.00         0.00         0.00      // fvdead_sev3:  //  // fraction of PFT vegetation killed for severity 3 // 
-0.64         0.64        0.64        0.64         0.00         0.00         0.00         0.00         0.00         0.00      // fvdead_sev4:  //  // fraction of PFT vegetation killed for severity 4 // 
-0.64         0.64        0.64        0.64         0.00         0.00         0.00         0.00         0.00         0.00      // fvdead_sev5:  //  // fraction of PFT vegetation killed for severity 5 // 
-0.00   // foslburn_sev1:  //  // fraction of OS layer burned for severity 1 // 
-0.48   // foslburn_sev2:  //  // fraction of OS layer burned for severity 2 // 
-0.54   // foslburn_sev3:  //  // fraction of OS layer burned for severity 3 // 
-0.69   // foslburn_sev4:  //  // fraction of OS layer burned for severity 4 // 
-0.80   // foslburn_sev5:  //  // fraction of OS layer burned for severity 5 // 
-0.50   // vsmburn:  //  // volume soil moisture to define wet soil for prevent organic layer burning // 
-0.10   // r_retain_c:  //  // ratio of burning residue C (retained into soil) // 
-0.10   // r_retain_n:  //  // ratio of burning residue N (retained into soil) // 

[…]
