### 2024-11-12 415cffb
Update temporal section with content from JOSS paper.
Choose to keep some of the old info but put it in collapse boxes.
