### 2024-11-12 9096828
Move content from JOSS paper draft (spatial discussion)
