### 2024-11-12 4c1a14d
Migrate/clone content from JOSS paper to model overview.
change setting to allow numbered figs, update .bib file.
