+++ docs_src/sphinx/source/bibliography.bib
@@ -50,6 +50,20 @@
   city         = {Boston MA, USA},
   issue        = 15
 }
+@article{briones2024exploring,
+  title        = {Exploring the interplay between soil thermal and hydrological changes and their impact on carbon fluxes in permafrost ecosystems},
+  author       = {Valeria Briones and Elchin E Jafarov and Hélène Genet and Brendan M Rogers and Ruth M Rutter and Tobey B Carman and Joy Clein and Eugénie S Euschkirchen and Edward AG Schuur and Jennifer D Watts and Susan M Natali},
+  year         = 2024,
+  month        = 7,
+  journal      = {Environmental Research Letters},
+  volume       = 19,
+  pages        = {074003},
+  doi          = {10.1088/1748-9326/ad50ed},
+  issn         = {1748-9326},
+  abstract     = {<p> Accelerated warming of the Arctic can affect the global climate system by thawing permafrost and exposing organic carbon in soils to decompose and release greenhouse gases into the atmosphere. We used a process-based biosphere model (DVM-DOS-TEM) designed to simulate biophysical and biogeochemical interactions between the soil, vegetation, and atmosphere. We varied soil and environmental parameters to assess the impact on cryohydrological and biogeochemical outputs in the model. We analyzed the responses of ecosystem carbon balances to permafrost thaw by running site-level simulations at two long-term tundra ecological monitoring sites in Alaska: Eight Mile Lake (EML) and Imnavait Creek Watershed (IMN), which are characterized by similar tussock tundra vegetation but differing soil drainage conditions and climate. Model outputs showed agreement with field observations at both sites for soil physical properties and ecosystem CO <sub>2</sub> fluxes. Model simulations of Net Ecosystem Exchange (NEE) showed an overestimation during the frozen season (higher CO <sub>2</sub> emissions) at EML with a mean NEE of 26.98 ± 4.83 gC/m <sup>2</sup> /month compared to observational mean of 22.01 ± 5.67 gC/m <sup>2</sup> /month, and during the fall months at IMN, with a modeled mean of 19.21 ± 7.49 gC/m <sup>2</sup> /month compared to observation mean of 11.9 ± 4.45 gC/m <sup>2</sup> /month. Our results underscore the importance of representing the impact of soil drainage conditions on the thawing of permafrost soils, particularly poorly drained soils, which will drive the magnitude of carbon released at sites across the high-latitude tundra. These findings can help improve predictions of net carbon releases from thawing permafrost, ultimately contributing to a better understanding of the impact of Arctic warming on the global climate system. </p>},
+  issue        = 7
+}
+
 @article{decker2009impact,
   title        = {Impact of Modified Richards Equation on Global Soil Moisture Simulation in the Community Land Model (CLM3.5)},
   author       = {Mark Decker and Xubin Zeng},
@@ -200,6 +214,17 @@
   doi          = {10.1007/978-90-481-9118-5_6},
   isbn         = 9789048191178,
[…]
