### 2024-11-12 f5144ee
remove empty processes WRITE THIS sections...
...seems like things have evolved such that we don't need these empty
sections...
