### 2024-11-12 c0e3a43
Remove sections from JOSS paper that have been migrated to docs:
- snow and soil
 - vegetation
 - run stages
