+++ joss-paper/paper.md
@@ -213,80 +213,6 @@ for light, water, and available N, shown with the purple arrow in the top
 center.\label{fig:modeloverview}](dvmdostem-overview-export_2024-08-19.jpg)
 
 
-## Snow and soil structure and processes
-
-The soil column is structured as a sequence of layers organized by soil horizons
-(i.e. fibric, humic, mineral, and parent material). The number and physical
-properties of layers may change throughout the simulation based on vegetation,
-thermal, hydrologic, and seasonal properties that are calculated at each time
-step [@Zhuang2003; @Euskirchen2014; @Yi2009; @McGuire2018]. The model uses the
-two-directional Stefan algorithm to predict freezing/thawing fronts and the
- Richards equation to predict soil moisture dynamics in the unfrozen layers
-[@Yi2009; @Yi2010; @Zhuang2003]. Snow is also represented with a dynamic stack
-of layers. The physical properties of the snowpack (density, thickness, and
-temperature) are calculated from snowfall, sublimation and snowmelt. Snow
-cover influences soil-thermal and hydrological seasonal dynamics. Changes in
-soil C stocks are a result of litterfall from the vegetation and decomposition
-of soil C stocks by microbes (heterotrophic respiration or Rh). Changes in soil
-organic and available N stocks are a result of litterfall, net mineralization of
-organic N, and plant N uptake. Soil organic layers and soil C and N stocks may
-also be modified due to wildfire.
-
-
-## Vegetation structure and processes
-
-Each vegetation CMT (e.g. “wet-sedge tundra”, “white spruce forest”, etc.), is
-modeled with up to ten PFTs (e.g., “deciduous shrubs”, “sedges”, “mosses”), 
-each of which may have up to three compartments: leaf, stem, and root.
-Vegetation C and N fluxes are calculated at each time step based on
-environmental factors and soil properties. Assimilation of atmospheric $CO_2$ by
-the vegetation is estimated by computing gross primary productivity (GPP) for
-each PFT. GPP is a function of foliage development (seasonal and
-successional patterns), air and soil temperature, water and nutrient
-availability, photosynthetically active radiation, and maximum assimilation rate
-(a calibrated parameter) [@McGuire1992; @Euskirchen2009]. Changes
-in vegetation C stocks are calculated using GPP, autotrophic respiration (Ra),
-and litterfall (transfer from vegetation to soil). Vegetation N stocks are
-calculated using plant N uptake and litterfall. Vegetation C and N stocks may
-also be modified as a result of wildfire burn.
-
-
-## Run stages
-
-To initialize historical or future simulations, `DVMDOSTEM` needs to compute a
-quasi steady-state (QSS) solution. This solution is forced by using averaged
-historical atmospheric and ecosystem properties (e.g. soil texture) to drive the
-model. QSS of physical processes (e.g. soil temperature and water content) are
-usually achieved in less than 100 years, while QSS of biogeochemical processes
-(e.g. soil and vegetation C and N stocks) are achieved in 1,000 to >10,000
-years. However, to decrease overall run-times, `DVMDOSTEM` uses two QSS stages: 
-“Pre-run” and “Equilibrium”. The list of all `DVMDOSTEM` run stages is as follows:
-
-* Pre-run (pr): QSS computation for the physical state variables.
-* Equilibrium (eq): QSS computation for the biogeochemical state variables. 
-* Spinup (sp): introduction of pre-industrial climate variability and fire
-  regime.
-* Transient (tr): historical simulation.
-* Scenario (sc): future simulation.
-
-Model simulation requires advancing the model consecutively through
-all of the run stages as needed (pr-eq-sp-tr).
-
-## Spatial considerations
-
-`DVMDOSTEM` can be applied at the site level or across large regions. Spatially,
[…]
