### 2024-11-12 8946f47
catch bib entry for spatial section. migrate from JOSS paper.
