+++ docs_src/sphinx/source/bibliography.bib
@@ -226,6 +226,17 @@
   abstract     = {Fire is an important factor controlling the composition and thickness of the organic layer in the black spruce forest ecosystems of interior Alaska. Fire that burns the organic layer can trigger dramatic changes in the underlying permafrost, leading to accelerated ground thawing within a relatively short time. In this study, we addressed the following questions. (1) Which factors determine post-fire ground temperature dynamics in lowland and upland black spruce forests? (2) What levels of burn severity will cause irreversible permafrost degradation in these ecosystems? We evaluated these questions in a transient modeling-sensitivity analysis framework to assess the sensitivity of permafrost to climate, burn severity, soil organic layer thickness, and soil moisture content in lowland (with thick organic layers, ∼80 cm) and upland (with thin organic layers, ∼30 cm) black spruce ecosystems. The results indicate that climate warming accompanied by fire disturbance could significantly accelerate permafrost degradation. In upland black spruce forest, permafrost could completely degrade in an 18 m soil column within 120 years of a severe fire in an unchanging climate. In contrast, in a lowland black spruce forest, permafrost is more resilient to disturbance and can persist under a combination of moderate burn severity and climate warming. © 2013 IOP Publishing Ltd.},
   issue        = 3
 }
+@article{johnstone2010changes,
+  title        = {Changes in fire regime break the legacy lock on successional trajectories in Alaskan boreal forest},
+  author       = {Jill F. Johnstone and Teresa N. Hollingsworth and F. Stuart Chapin and Michelle C. Mack},
+  year         = 2010,
+  journal      = {Global Change Biology},
+  volume       = 16,
+  doi          = {10.1111/j.1365-2486.2009.02051.x},
+  issn         = 13541013,
+  abstract     = {Predicting plant community responses to changing environmental conditions is a key element of forecasting and mitigating the effects of global change. Disturbance can play an important role in these dynamics, by initiating cycles of secondary succession and generating opportunities for communities of long-lived organisms to reorganize in alternative configurations. This study used landscape-scale variations in environmental conditions, stand structure, and disturbance from an extreme fire year in Alaska to examine how these factors affected successional trajectories in boreal forests dominated by black spruce. Because fire intervals in interior Alaska are typically too short to allow relay succession, the initial cohorts of seedlings that recruit after fire largely determine future canopy composition. Consequently, in a dynamically stable landscape, postfire tree seedling composition should resemble that of the prefire forest stands, with little net change in tree composition after fire. Seedling recruitment data from 90 burned stands indicated that postfire establishment of black spruce was strongly linked to environmental conditions and was highest at sites that were moist and had high densities of prefire spruce. Although deciduous broadleaf trees were absent from most prefire stands, deciduous trees recruited from seed at many sites and were most abundant at sites where the fires burned severely, consuming much of the surface organic layer. Comparison of pre- and postfire tree composition in the burned stands indicated that the expected trajectory of black spruce self-replacement was typical only at moist sites that burned with low fire severity. At severely burned sites, deciduous trees dominated the postfire tree seedling community, suggesting these sites will follow alternative, deciduous-dominated trajectories of succession. Increases in the severity of boreal fires with climate warming may catalyze shifts to an increasingly deciduous-dominated landscape, substantially altering landscape dynamics and ecosystem services in this part of the boreal forest. © 2009 Blackwell Publishing Ltd.},
+  issue        = 4
+}
 @misc{jordan1991onedimensional,
   title        = {A One-dimensional temperature model for a snow cover : technical documentation for SNTHERM.89},
   author       = {Rachel E. Jordan},
