### 2024-11-12 971f6e7
migrate parameterization section from JOSS paper to docs.
Also move some items on output prep/processing into the running docs
page.
