+++ docs_src/sphinx/source/bibliography.bib
@@ -215,6 +215,12 @@
   isbn         = 9789048191178,
   abstract     = {Recent changes in climate, disturbance regimes and land use and management systems in Northern Eurasia have the potential to disrupt the terrestrial sink of atmospheric CO2 in a way that accelerates global climate change. To determine the recent trends in the carbon balance of the arctic and boreal ecosystems of this region, we performed a retrospective analysis of terrestrial carbon dynamics across northern Eurasia over a recent 10-year period using a terrestrial biogeochemical process model. The results of the simulations suggest a shift in direction of the net flux from the terrestrial sink of earlier decades to a net source on the order of 45 Tg C year-1 between 1997 and 2006. The simulation framework and subsequent analyses presented in this study attribute this shift to a large loss of carbon from boreal forest ecosystems, which experienced a trend of decreasing precipitation and a large area burned during this time period. © 2011 Springer Science+Business Media B.V.}
 }
+@article{jafarovINPREP2024,
+  title        = {Estimation of above- and below-ground ecosystem parameters for the DVM-DOS-TEM model using MADS: a synthetic case study},
+  author       = {Elchin E. Jafarov and Hélène Genet Genet and Velimir V. (Monty) Vesselinov and Valeria Briones and Aiza Kabeer and Andrew L. Mullen and Benjamin Maglio and Tobey Carman and Ruth Rutter and Joy Clein and Chu-Chun Chang and Dogukan Teber and Trevor Smith and Joshua Rady and Christina Schädel and Jennifer D. Watts and Brendan M. Rogers and Susan Natali},
+  journal      = {Manuscript submitted for publication},
+  year         = 2024,
+}
 @article{jafarov2013effects,
   title        = {The effects of fire on the thermal stability of permafrost in lowland and upland black spruce forests of interior Alaska in a changing climate},
   author       = {E. E. Jafarov and V. E. Romanovsky and H. Genet and A. D. McGuire and S. S. Marchenko},
