### 2024-11-12 979aec3
Migrate I/O section from JOSS paper to docs.
