+++ docs_src/sphinx/source/bibliography.bib
@@ -77,6 +77,12 @@
   abstract     = {A fundamental deficiency has been found in the numerical solution of the soil moisture‐based Richards equation using the mass‐conservative scheme in the Community Land Model (CLM) in the first part of our efforts ( Zeng and Decker 2009). This study implements the revised form of the Richards equation from that study (which doesn't change the property of the differential equation but does remove the deficiency of the numerical solution) along with a new bottom boundary condition into the current version of CLM (CLM3.5) for global offline modeling evaluations. CLM3.5 represents a significant improvement over its earlier version (CLM3.0), but it also introduces a new deficiency in the vertical distribution of the soil moisture variability. Mean soil moisture in CLM3.5 is also too wet. It is found that the new treatments (primarily a numerically correct solution of Richards equation with a new bottom boundary condition) with minimal tuning are able to maintain the improvements of the CLM3.5 over CLM3.0 and, at the same time, remove the new deficiencies of CLM3.5 based on in situ and satellite data analysis. Because the deficiency in the numerical solution of the soil moisture‐based Richards equation is also expected in other land models, implementation details are provided to facilitate similar tests using other land models in the future.},
   issue        = 3
 }
+@misc{eaton2011netcdf,
+  title        = {NetCDF Climate and Forecast (CF) Metadata Conventions},
+  author       = {B. Eaton and J. Gregory and B. Drach and K. Taylor and S. Hankin and J. Caron and R. Signell and P. Bentley and G. Rappa and H. Höck and A. Pamment and M. Juckes},
+  year         = 2011,
+  url          = {https://cfconventions.org/Data/cf-conventions/cf-conventions-1.11/cf-conventions.html}
+}
 @article{euskirchen2006importance,
   title        = {Importance of recent shifts in soil thermal dynamics on growing season length, productivity, and carbon sequestration in terrestrial high-latitude ecosystems},
   author       = {E. S. Euskirchen and A. D. McGuire and D. W. Kicklighter and Q. Zhuang and J. S. Clein and R. J. Dargaville and D. G. Dye and J. S. Kimball and K. C. McDonald and J. M. Melillo and V. E. Romanovsky and N. V. Smith},
@@ -376,6 +382,17 @@
   issn         = {00431397},
   abstract     = {Snow surface temperature (Ts) is important to the snowmelt energy balance and landatmosphere interactions, but in situ measurements are rare, thus limiting evaluation of remote sensing data sets and distributed models. Here we test simple Ts approximations with standard height (2-4 m) air temperature (Ta), wet-bulb temperature (Tw), and dew point temperature (Td), which are more readily available than Ts. We used hourly measurements from seven sites to understand which Ts approximation is most robust and how Ts representation varies with climate, time of day, and atmospheric conditions (stability and radiation). Td approximated Ts with the lowest overall bias, ranging from 22.3 to 12.6°C across sites and from 22.8 to 1.5°C across the diurnal cycle. Prior studies have approximated Ts with Ta, which was the least robust predictor of Ts at all sites. Approximation of Ts with Td was most reliable at night, at sites with infrequent clear sky conditions, and at windier sites (i.e., more frequent turbulent instability). We illustrate how mean daily T d can help detect surface energy balance bias in a physically based snowmelt model. The results imply that spatial Td data sets may be useful for evaluating snow models and remote sensing products in data sparse regions, such as alpine, cold prairie, or Arctic regions. To realize this potential, more routine observations of humidity are needed. Improved understanding of Td variations will advance understanding of T s in space and time, providing a simple yet robust measure of snow surface feedback to the atmosphere. © 2013. American Geophysical Union. All Rights Reserved.},
[…]
