+++ include/diagnostics.h
@@ -236,10 +236,13 @@ struct soidiag_bgc {
   double orgnsum;
   double avlnsum;
 
+  //Variables used for or resulting from CH4 flux calculation
   double ch4ebulsum; //Total ebullition
+
   soidiag_bgc(): shlwc(UIN_D), deepc(UIN_D), mineac(UIN_D), minebc(UIN_D),
       minecc(UIN_D), rawcsum(UIN_D), somasum(UIN_D), somprsum(UIN_D),
-      somcrsum(UIN_D), orgnsum(UIN_D), avlnsum(UIN_D) {
+      somcrsum(UIN_D), orgnsum(UIN_D), avlnsum(UIN_D),
+      ch4ebulsum(UIN_D) {
 
     for (int i = 0; i < MAX_SOI_LAY; ++i) {
       knmoist[i] = UIN_D;
