### 2024-11-14 e5c1a83
Properly clears the yearly transport struct entry
It also skips the middle output cases, as they
are not implemented yet.
