+++ src/Runner.cpp
@@ -1600,6 +1600,7 @@ void Runner::output_netCDF(std::map<std::string, OutputSpec> &netcdf_outputs, in
 
     #pragma omp critical(outputCH4EBULLITION)
     {
+      //Ebullition is not offered as a sum
 
       if(curr_spec.layer){
         std::array<double, MAX_SOI_LAY> ch4ebul_arr{};
@@ -1630,19 +1631,16 @@ void Runner::output_netCDF(std::map<std::string, OutputSpec> &netcdf_outputs, in
           }
         }
         //yearly
-
-      }
-      //Sum instead of by layer
-      else{
-
-        //daily
-        if(curr_spec.daily){
-        }
-        else if(curr_spec.monthly){
-          output_nc_3dim(&curr_spec, file_stage_suffix, &cohort.bdall->m_soid.ch4ebulsum, 1, month_timestep, 1);
-        }
         else if(curr_spec.yearly){
-//          output_nc_3dim(&curr_spec, file_stage_suffix, &, 1, year, );
+          for(int il=0; il<MAX_SOI_LAY; il++){
+            ch4ebul_arr[il] = cohort.bdall->y_soi2soi.ch4_ebul[il];
+          }
+          outhold.ch4ebul_layer_for_output.push_back(ch4ebul_arr);
+
+          if(output_this_timestep){
+            output_nc_4dim(&curr_spec, file_stage_suffix, &outhold.ch4ebul_layer_for_output[0], MAX_SOI_LAY, year_start_idx, years_to_output);
+            outhold.ch4ebul_layer_for_output.clear();
+          }
         }
       }
     }//end critical(outputCH4EBULLITION)
