### 2024-11-15 ff8b741
Remove parameterization and software design sections from JOSS paper.
