### 2024-11-16 2717df1
JOSS paper, move Figures for better text flow.
