### 2024-11-16 3f177c6
User Guide, various:
- update bibliography
 - add software design seciton from JOSS paper
 - add figure. This used to be an embedded drawing that was in T.
   Carman's account. Moved the drawing to the shared drive and changed
   to a figure rather than embed, so that figure with caption can be
   used.
