+++ docs_src/sphinx/source/bibliography.bib
@@ -344,6 +344,19 @@
   keywords     = {Alaska,Alaska carbon cycle,Landscape Conservation Cooperative,boreal forest,climate change,fire,inland aquatic ecosystems,maritime conifer forest,permafrost,tundra,uplands,wetlands},
   pmid         = 29923353
 }
+@article{merkel2014docker,
+  title        = {Docker: lightweight Linux containers for consistent development and deployment},
+  author       = {Dirk Merkel},
+  year         = 2014,
+  month        = 3,
+  journal      = {Linux J.},
+  publisher    = {Belltown Media},
+  volume       = 2014,
+  issn         = {1075-3583},
+  abstract     = {Docker promises the ability to package applications and their dependencies into lightweight containers that move easily between different distros, start up quickly and are isolated from each other.},
+  city         = {Houston, TX},
+  issue        = 239
+}
 @article{niu2005simple,
   title        = {A simple TOPMODEL-based runoff parameterization (SIMTOP) for use in global climate models},
   author       = {Guo Yue Niu and Zong Liang Yang and Robert E. Dickinson and Lindsey E. Gulden},
