+++ docs_src/sphinx/source/running_dvmdostem.rst
@@ -412,8 +412,10 @@ Parallel Options
 ==================================
 Processing Outputs
 ==================================
-    
-    ------------------
+
+    WRITE THIS....
+
+------------------
 Output Selection
 ------------------
     WRITE THIS...
