+++ src/Soil_Bgc.cpp
@@ -265,8 +265,8 @@ void Soil_Bgc::CH4Flux(const int mind, const int id) {
           pft_transport[ip] = 0.0;
         }
 
-        // accumulating across all pfts
-        plant += pft_transport[ip];
+        // accumulating across all pfts, assuming 50% oxidation in rhizosphere
+        plant += 0.5 * pft_transport[ip];
 
         //Storing plant transport values for output
         bd->d_soi2a.ch4_transport[il][ip] = pft_transport[ip];
@@ -315,21 +315,23 @@ void Soil_Bgc::CH4Flux(const int mind, const int id) {
       //>>> Why >= -2.0? and why 0.25?
       //>>> BM: https://doi.org/10.1073/pnas.1420797112 ?
       //>>> ^^ I don't think this is the one, but probably some kind of winter Q10
-      if(currl->tem >= -2.0 && currl->tem < 0.000001){
-        TResp_unsat = getQ10(calpar.oxidq10_ch4, -2.1) * 0.25;
-        TResp_sat = getQ10(calpar.prodq10_ch4, -6.0) * 0.25;
-      }
-      else if(currl->tem < -2.0){
+      // if(currl->tem >= -2.0 && currl->tem < 0.000001){
+      //   TResp_unsat = getQ10(calpar.oxidq10_ch4, -2.1) * 0.25;
+      //   TResp_sat = getQ10(calpar.prodq10_ch4, -6.0) * 0.25;
+      // }
+      // else 
+      if(currl->tem < 0.0){
         TResp_unsat = 0.0;
         TResp_sat = 0.0;
       }
       else{
-        TResp_unsat = getQ10(calpar.oxidq10_ch4, currl->tem - 2.1);
-        TResp_sat = getQ10(calpar.prodq10_ch4, currl->tem - 6.0);
+        TResp_unsat = getQ10(calpar.oxidq10_ch4, currl->tem - calpar.oxidTref_ch4);
+        TResp_sat = getQ10(calpar.prodq10_ch4, currl->tem - calpar.prodTref_ch4);
       }
 
-      // Fan 2013 Eq 18, Vmax and km are Michaelis-Menten kinetics parameters
-      oxid = (1 - saturated_fraction) * (calpar.oxidkm_ch4 * currl->ch4 * TResp_unsat / (calpar.oxidVmax_ch4 + currl->ch4));
+      // Fan 2013 Eq 18, Vmax and km are Michaelis-Menten kinetics parameters.
+      // Plant is added here to account for 50% rhizospheric oxidation
+      oxid = (1 - saturated_fraction) * (calpar.oxidkm_ch4 * currl->ch4 * TResp_unsat / (calpar.oxidVmax_ch4 + currl->ch4)) + plant;
 
       if (oxid < 0.000001){
         oxid = 0.0;
@@ -371,9 +373,8 @@ void Soil_Bgc::CH4Flux(const int mind, const int id) {
         prod_somcr_ch4[il] = 0.0;
       }
 
-      // output for production
       prod_gm2hr = (prod_rawc_ch4[il] + prod_soma_ch4[il] + prod_sompr_ch4[il] + prod_somcr_ch4[il]);
-
+          
       // Fan 2013 Eq. 9, converted back to umolL^-1hr^-1 for solver
       prod = (convert_gm2_to_umolL) * prod_gm2hr; 
 
@@ -408,10 +409,10 @@ void Soil_Bgc::CH4Flux(const int mind, const int id) {
 
       // is not at soil surface
       if (wtlayer != ground->fstshlwl){
-        bubble += ebul;
         if (currl == wtlayer){
[…]
