### 2024-11-18 8becb7c
Remove unneeded refs from JOSS paper.
