+++ joss-paper/joss-paper.bib
@@ -11,12 +11,6 @@
    volume = {19},
    year = {2024},
 }
-@misc{Eaton2011,
-   author = {B. Eaton and J. Gregory and B. Drach and K. Taylor and S. Hankin and J. Caron and R. Signell and P. Bentley and G. Rappa and H. Höck and A. Pamment and M. Juckes},
-   title = {NetCDF Climate and Forecast (CF) Metadata Conventions},
-   url = {https://cfconventions.org/Data/cf-conventions/cf-conventions-1.11/cf-conventions.html},
-   year = {2011},
-}
 @article{Euskirchen2022,
    abstract = {<p>As the Arctic region moves into uncharted territory under a warming climate, it is important to refine the terrestrial biosphere models (TBMs) that help us understand and predict change. One fundamental uncertainty in TBMs relates to model parameters, configuration variables internal to the model whose value can be estimated from data. We incorporate a version of the Terrestrial Ecosystem Model (TEM) developed for arctic ecosystems into the Predictive Ecosystem Analyzer (PEcAn) framework. PEcAn treats model parameters as probability distributions, estimates parameters based on a synthesis of available field data, and then quantifies both model sensitivity and uncertainty to a given parameter or suite of parameters. We examined how variation in 21 parameters in the equation for gross primary production influenced model sensitivity and uncertainty in terms of two carbon fluxes (net primary productivity and heterotrophic respiration) and two carbon (C) pools (vegetation C and soil C). We set up different parameterizations of TEM across a range of tundra types (tussock tundra, heath tundra, wet sedge tundra, and shrub tundra) in northern Alaska, along a latitudinal transect extending from the coastal plain near Utqiaġvik to the southern foothills of the Brooks Range, to the Seward Peninsula. TEM was most sensitive to parameters related to the temperature regulation of photosynthesis. Model uncertainty was mostly due to parameters related to leaf area, temperature regulation of photosynthesis, and the stomatal responses to ambient light conditions. Our analysis also showed that sensitivity and uncertainty to a given parameter varied spatially. At some sites, model sensitivity and uncertainty tended to be connected to a wider range of parameters, underlining the importance of assessing tundra community processes across environmental gradients or geographic locations. Generally, across sites, the flux of net primary productivity (NPP) and pool of vegetation C had about equal uncertainty, while heterotrophic respiration had higher uncertainty than the pool of soil C. Our study illustrates the complexity inherent in evaluating parameter uncertainty across highly heterogeneous arctic tundra plant communities. It also provides a framework for iteratively testing how newly collected field data related to key parameters may result in more effective forecasting of Arctic change.</p>},
    author = {Eugénie S. Euskirchen and Shawn P. Serbin and Tobey B. Carman and Jennifer M. Fraterrigo and Hélène Genet and Colleen M. Iversen and Verity Salmon and A. David McGuire},
@@ -41,17 +35,6 @@
    title = {Changes in vegetation in northern Alaska under scenarios of climate change, 2003–2100: implications for climate feedbacks},
    volume = {19},
    year = {2009},
-}
-@article{Euskirchen2014,
[…]
