### 2024-11-21 2c8f448
Port changes to docs from the JOSS paper draft (and branch)
The two big changes here are:
- Adding sphinx tools to manage the references list and allow citations
   from within the documentation.
 - Adding content that was originally written for the JOSS paper draft.

These changes were inspired as part of the JOSS paper review process.
