### 2024-11-21 f49a7ca
Update PYTHONPATH reccomendation for building docs
