+++ include/OutputHolder.h
@@ -71,6 +71,7 @@ public:
 
   std::vector<double> ch4transport_sum_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> ch4transport_layer_for_output;
+  std::vector<std::array<double, NUM_PFT>> ch4transport_pft_for_output;
   std::vector<std::array<std::array<double, NUM_PFT>, MAX_SOI_LAY>> ch4transport_for_output; 
 
   std::vector<std::array<double, MAX_SOI_LAY>> layerdepth_for_output;
