### 2024-12-02 6395e25
Adds basic summed output comparison plotter
For example, this can be used to examine SOCFROZEN
and SOCUNFROZEN vs SOC to sanity check those outputs.
