### 2024-12-03 ab1454d
Adds recalculated sum to the comparison plotter
