### 2024-12-03 968674f
Remove indirect dependencies in mapping support environment due to version conflict.
This is not the most ideal way to hande this issue, but a better
solution is forthcoming in a branch that rearranges the requirements
tracking for the project.
