+++ requirements_mapping.txt
@@ -4,9 +4,9 @@ backcall==0.2.0
 bokeh==3.0.3
 certifi==2024.7.4
 cftime==1.2.1
-click==7.1.2
+#click==7.1.2
 click-plugins==1.1.1
-cligj==0.7.1
+#cligj==0.7.1
 commentjson==0.8.3
 configobj==5.0.6
 cycler==0.10.0
