### 2024-12-03 87891bd
Replaces deprecated methods for soil plotter
