### 2024-12-03 a4c3fe0
Replaces deprecated methods for veg plotter
