### 2024-12-03 d8aa5b7
Adds old outstanding changes to veg plotter
These changes were added long ago, but the plotter
works for a trivial case so they are acceptable.
