### 2024-12-04 #7136 @tobeycarman
@kthyng I think we are ready to move this forward again. I cut a new release of ``dvmdostem`` yesterday (v0.8.1) and published the documentation which has updated information that we ported from the original JOSS draft. And the paper is compiling, see above ☝️. It would be great to get the status "in review" as there are some folks who would like something to cite for some upcoming AGU presentations.
