### 2024-12-09 #7136 @kthyng
@tobeycarman Looks like some great changes! I see a few "write this" sections in your docs that would be great to address before reviewers get to them but I won't hold up review based on them. This will be going to the waitlist until a relevant editor has capacity so you'll probably have some time to work on this. (https://uaf-arctic-eco-modeling.github.io/dvm-dos-tem/SA_UQ.html, https://uaf-arctic-eco-modeling.github.io/dvm-dos-tem/examples_and_tutorials/example_experiment_0.html#plot-soil-temperatures).
