### 2024-12-09 #7136 @tobeycarman
re: capitalization on references, fixed in [dvm-dos-tem PR#761](https://github.com/uaf-arctic-eco-modeling/dvm-dos-tem/pull/761). Note that this will not be reflected in the published documentation until the next version release.
