### 2024-12-11 a36775b
Adds general utility python tools to generated sphinx docs.
