### 2024-12-11 2eff521
Adds doctest file for general utility python method
