### 2024-12-16 a6f90ea
Bump configobj from 5.0.6 to 5.0.9
Bumps [configobj](https://github.com/DiffSK/configobj) from 5.0.6 to 5.0.9.
- [Release notes](https://github.com/DiffSK/configobj/releases)
- [Changelog](https://github.com/DiffSK/configobj/blob/release/CHANGES.rst)
- [Commits](https://github.com/DiffSK/configobj/compare/v5.0.6...v5.0.9)

---
updated-dependencies:
- dependency-name: configobj
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
