+++ src/Soil_Bgc.cpp
@@ -77,39 +77,26 @@ void Soil_Bgc::assignCarbonBd2LayerMonthly() {
   // What about mineral C ??
 }
 
+void Soil_Bgc::TriSolver(int matrix_size, double *A, double *B, double *C, double *D, double *X) {
+    // Forward elimination
+    for (int i = 1; i < matrix_size; i++) { 
+        double xmult = A[i] / B[i - 1];    // A[i] not A[i-1] used as  A[-1] is
+        B[i] -= xmult * C[i - 1];          // unused given open upper boundary 
+        D[i] -= xmult * D[i - 1];
+    }
 
-void Soil_Bgc::TriSolver(int matrix_size, double *A, double *D, double *C, double *B,double *X) {
-
-  // B is known state, X is solution, A, D, C are used to construct tridiagonal matrix
-  // note: commonly a, b, and c are in first matrix, here b and d are reversed
-  // |  d1  c1  0   0  .  0  |   | x1 |   | b1 |
-  // |  a2  d2  c2  0  .  0  |   | x2 |   | b2 |
-  // |  0  a3  d3  c3  0  0  |   | x3 |   | b3 |
-  // |  .  .              .  | x | .. | = | .. |
-  // |  .  .             .   |   | .. |   | .. |
-  // |  .  .             cn  |   | .. |   | .. |
-  // |  0  0  0  0   an  dn  |   | xn |   | bn |
-
-	int i;
-	double xmult;
-
-	for (i = 2; i <= matrix_size; i++) {
-		xmult = A[i - 1] / D[i - 1];
-		D[i] -= xmult * C[i - 1];
-		B[i] -= xmult * B[i - 1];
-	}
-
-	X[matrix_size] = B[matrix_size] / D[matrix_size];
-	for (i = matrix_size - 1; i >= 1; i--)
-		X[i] = (B[i] - C[i] * X[i + 1]) / D[i];
+    // Backward substitution
+    X[matrix_size - 1] = D[matrix_size - 1] / B[matrix_size - 1]; 
+    for (int i = matrix_size - 2; i >= 0; i--) { 
+        X[i] = (D[i] - C[i] * X[i + 1]) / B[i];
+    }
 }
-
 
 void Soil_Bgc::CH4Flux(const int mind, const int id) {
 
   // Initial atmospheric CH4 concentration upper boundary condition
   // Fan et al. 2010 supplement Eq. 13. Units: umol L^-1
-  const double upper_bound = 0.076; // >>> This should be input forcing 
+  const double upper_bound = 0.076; // >>> This should be input forcing 0.076
   // Defining hourly time step
   double dt = 1.0 / HR_IN_DAY;
   // Unit conversion factors
@@ -119,26 +106,27 @@ void Soil_Bgc::CH4Flux(const int mind, const int id) {
   // 1 L^-1   = 1000  m^-3
   // multiplying by the layer thickness dz gives m^-2
   double convert_umolL_to_gm3 = 0.012;
-  // From Fan Eq. 8, components of CH4 mass balance prior to solver 
-  // equal to prod - oxid - ebul - plant, units : umol L^-1 hr^-1 
-  double partial_delta_ch4;
-  // Diffusion-specific efflux from top soil layer to atmosphere 
+  // From Fan Eq. 8, components of CH4 mass balance prior to solver
+  // equal to prod - oxid - ebul - plant, units : umol L^-1 hr^-1
+  // representing the reaction term in the reaction-diffusion equation
+  double reaction_term;
+  // Diffusion-specific efflux from top soil layer to atmosphere
   // in units umolL^-1hr^-1, umolL^-1day^-1, gm^-2day^-1 respectively
[…]
