### 2024-12-17 01e6040
Removing bonanza creek and reinstating toolik demo-data
