+++ demo-data/cru-ts40_ar5_rcp85_ncar-ccsm4_toolik_field_station_10x10/notes.txt
@@ -1,4 +1,4 @@
-2019-12-12T14:21:58.140828
+2019-08-08T14:50:31.116819
 =====================
 Modified the vegetation file to set all CMT 08 pixels to CMT 07.
 Saved a copy of the original, just in case.
