### 2024-12-23 147577d
Need to initialize CH4 as positive number
