### 2024-12-23 07c1f87
Commented code for ebullition efflux via preferential pathways
