+++ requirements_mapping.txt
@@ -18,7 +18,7 @@ geopandas==0.8.1
 ipython==8.10.0
 ipython-genutils==0.2.0
 jedi==0.17.2
-Jinja2==3.1.4
+Jinja2==3.1.5
 kiwisolver==1.2.0
 lark-parser==0.7.8
 MarkupSafe==2.0
