### 2024-12-26 3483553
removing additional CMTs from 61-SCB
