+++ mads_calibration/calibration_files/calibrated/CMT61-Scotty_Creek_Bog/cmt_bgcvegetation.txt
@@ -3,618 +3,6 @@
 // Maximum of 10 PFTs
 // PFT names should be CamelCase and must not have a period in them!
 //
-//===========================================================
-// CMT00 // BARE GROUND OPEN WATER SNOW AND ICE  // ##JUNK VALUES!!!###
-//PFT0       PFT1         PFT2         PFT3         PFT4         PFT5         PFT6         PFT7         PFT8         PFT9         // names: comments                  
-400.0        400.0        400.0        400.0        400.0        400.0        400.0        400.0        400.0        400.0        // kc:
-75           75           75           75           75           75           75           75           75           75           // ki:
--8.0         -8.0        -8.0         -8.0         -8.0         -8.0         -8.0         -8.0         -8.0         -8.0          // tmin:
-14.2         14.2         14.2         14.2         14.2         14.2         14.2         14.2         14.2         14.2         // toptmin:
-22.4         22.4         22.4         22.4         22.4         22.4         22.4         22.4         22.4         22.4         // toptmax:
-29.0         29.0         29.0         29.0         29.0         29.0         29.0         29.0         29.0         29.0         // tmax:
-2.3566500    2.3566500    2.3566500    2.3566500    2.3566500    2.3566500    2.3566500    2.3566500    2.3566500    2.3566500    // raq10a0:
--0.053077   -0.053077    -0.0530770   -0.0530770   -0.0530770   -0.0530770   -0.0530770   -0.0530770   -0.0530770   -0.0530770    // raq10a1:
-0.0023842    0.0023842    0.0023842    0.0023842    0.0023842    0.0023842    0.0023842    0.0023842    0.0023842    0.0023842    // raq10a2:
--0.0000411  -0.0000411   -0.0000411   -0.0000411   -0.0000411   -0.0000411   -0.0000411   -0.0000411   -0.0000411   -0.0000411    // raq10a3:
-0.0042       0.0042       0.0042       0.0042       0.0042       0.0042       0.0042       0.0042       0.0042       0.0042       // knuptake: 
-0.191        0.322        0.644        0.718        0.375        0.375        0.375        1.000        1.000        0.000        // cpartl:
-0.335        0.651        0.277        0.120        0.000        0.000        0.000        0.000        0.000        0.000        // cpartw:
-0.474        0.027        0.079        0.162        0.625        0.625        0.625        0.000        0.000        0.000        // cpartr:
-70.57        24.98        23.53        29.88        21.31        21.04        22.44        35.57        43.38        0.000        // initcnevenl:
-554.76       38.53        35.76        33.35        0.000        0.000        0.000        0.000        0.000        0.000        // initcnevenw:
-70.57        63.49        61.35        64.72        50.00        52.08        54.05        0.000        0.000        0.000        // initcnevenr:
-70.57        24.98        23.53        29.88        21.31        21.04        22.44        35.57        43.38        0.000        // cnevenbl:
-554.76       38.53        35.76        33.35        0.000        0.000        0.000        0.000        0.000        0.000        // cnevenbw:
-70.57        63.49        61.35        64.72        50.00        52.08        54.05        0.000        0.000        0.000        // cnevenbr:
[…]
