### 2024-12-30 932aa38
Additional code clean-up Logging statements have been added, some commented code removed, code written for ebullition efflux via preferential pathways in two forms (50% assumed efflux and exponentially decreasing with depth).
