+++ src/Soil_Bgc.cpp
@@ -96,7 +96,8 @@ void Soil_Bgc::CH4Flux(const int mind, const int id) {
 
   // Initial atmospheric CH4 concentration upper boundary condition
   // Fan et al. 2010 supplement Eq. 13. Units: umol L^-1
-  const double upper_bound = 0.076; // >>> This should be input forcing 0.076
+  const double upper_bound = 0.076;
+  // upper_bound = ed->y_atms.ch4;
   // Defining hourly time step
   double dt = 1.0 / HR_IN_DAY;
   // Unit conversion factors
@@ -116,8 +117,8 @@ void Soil_Bgc::CH4Flux(const int mind, const int id) {
   // Individual layer fluxes production, ebullition, oxidation
   // plant-mediation declared in loop to sum pfts, units : umol L^-1 hr^-1
   double prod, ebul, ebul_efflux, oxid = 0.0;
-  // Individual layer fluxes and efflux units : g m^-2 hr^-1
-  double prod_gm2hr, ebul_gm2hr, oxid_gm2hr, plant_gm2hr, efflux_gm2hr = 0.0; // >> efflux per hour not used?
+  // Individual layer fluxes units : g m^-2 hr^-1
+  double prod_gm2hr, ebul_gm2hr, oxid_gm2hr, plant_gm2hr = 0.0;
   // Flux components cumulated over a day in units of umol L^-1 day^-1
   double plant_daily, ebul_daily = 0.0;
   // Flux components cumulated over a day in units of g m^-2 day^-1
@@ -139,7 +140,7 @@ void Soil_Bgc::CH4Flux(const int mind, const int id) {
   // Defining number of layers for solving (including boundaries)
   int numlay = numsoill + 1;
 
-  // Storing pool + R * dt prior to solver for calculating efflux
+  // Storing pool + reaction term * dt prior to solver for calculating efflux
   double prev_pool_R[MAX_SOI_LAY] = {0};
   
   // For storing methane movement data by layer for output
@@ -154,23 +155,23 @@ void Soil_Bgc::CH4Flux(const int mind, const int id) {
   // Outputting solution to solver
   double solution[numlay] = {0};
 
-  // alpha = Ddt/2dx^2
+  // alpha constant in crank nicolson discretization, Ddt/2dx^2
   double alpha = 0.0;
   // Storing diffusion coefficient
   double diff[numlay] = {0};
 
-    // ch4 production responses: kdc_ch4 * Tresp(q10) * SOM pool, Fan Eq. 9
+  // ch4 production responses: kdc_ch4 * Tresp(q10) * SOM pool, Fan Eq. 9
   double prod_rawc_ch4[MAX_SOI_LAY] = {0};
   double prod_soma_ch4[MAX_SOI_LAY] = {0};
   double prod_sompr_ch4[MAX_SOI_LAY] = {0};
   double prod_somcr_ch4[MAX_SOI_LAY] = {0};
 
-  // >>> q10 used for RH needs to be different for CH4 
   // Temperature response (q10) for unsaturated and saturated layers
   double TResp_unsat, TResp_sat = 0;
 
   // Function of LAI to describe relative plant size for ch4 transport
   double fLAI[NUM_PFT];
+
   //Looping through pfts and assigning fgrow (0-1) from Fan Eq. 20
   for(int ip=0; ip<NUM_PFT; ip++){
     if(cd->m_vegd.ffoliage[ip]<=0){
@@ -181,10 +182,7 @@ void Soil_Bgc::CH4Flux(const int mind, const int id) {
     }
   }
 
-  // >>> not sure if this is needed or whether we should follow this 
-  // >>> syntax convention for all / other variables in this loop?
[…]
