### 2025-01-03 c734c47
Replacing interception equations with CLM5.0 This appears to predict similar values, though ircoef and iscoef are both being used here, and can be tuned to assist in calibration of thermal and hydrological dynamics if required.
