### 2025-01-06 #7136 @mengqi-z
👋 @JoanyMarino - Would you be available to review a submission for JOSS? The review process is straightforward, handled through GitHub issues, and follows a checklist-driven approach: https://joss.readthedocs.io/en/latest/review_criteria.html
