### 2025-01-07 3d15bb2
Re-enabling succession flag
