### 2025-01-09 #7136 @mengqi-z
👋 @r-barnes @ruillercoaster - Would any of you be available to review a submission for JOSS? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
