### 2025-01-09 b4855b8
Partly writing nitrogen redistribution
Need to debug further.
