+++ include/ModelData.h
@@ -91,7 +91,7 @@ public:
   int baseline_start;//Start year for baseline EQ climate
   int baseline_end;//End year for baseline EQ climate
 
-  bool veg_succession;//Allow cmt change following disturbance
+  bool cmt_change;//Allow cmt change following disturbance
 
   // Maps holding data about variables to be output at specific timesteps
   // C++11 would allow the use of unordered_maps, which have a faster
