### 2025-01-10 b7fb105
initial groundwater input test
