+++ include/Richards.h
@@ -28,6 +28,7 @@ public :
 
   //Array to hold lateral drainage values for output to file.
   double layer_drain[MAX_SOI_LAY];
+  double layer_input[MAX_SOI_LAY];
   double percolation[MAX_SOI_LAY];
 
   void update(Layer *fstsoill, Layer* bdrainl, const double & bdraindepth,
