### 2025-01-14 #7667 @editorialbot
[REVIEW]: DVMDOSTEM: a terrestrial ecosystem model designed to represent arctic, boreal and permafrost ecosystem dynamics
**Submitting author:** @tobeycarman (<a href="http://orcid.org/0000-0003-4617-4674">Tobey Carman</a>)
**Repository:** https://github.com/uaf-arctic-eco-modeling/dvm-dos-tem
**Branch with paper.md** (empty if default branch): master
**Version:** v0.8.3
