### 2025-01-14 #7667 @mengqi-z
👋🏼 @kdorheim @arbennett - this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```
