### 2025-01-21 781eeee
Fix indexing but when sticthcing stages
