+++ scripts/util/output.py
@@ -396,7 +396,7 @@ def stitch_stages(var, timestep, stages, fileprefix='', with_dti=False):
 
   if with_dti:
     ds_begin = nc.Dataset(expected_file_names[0])
-    ds_end = nc.Dataset(expected_file_names[1])
+    ds_end = nc.Dataset(expected_file_names[-1])
     dti = build_full_datetimeindex(ds_begin, ds_end, timeres=timestep)
     return (full_ds, units_str, dimensions, dti)
   else:
