### 2025-01-21 3ec7c4c
Major re-organization to support pyddt as a standalone package.
- changed some filenames to use lowercase to avoid name conflict with
class names.
  - adjusted imports to relative for sibiling packages

Have not run any of the tests, those will need to be adjusted. But if
you set you PYTHONPATH to include the pyddt directory, then imports in
a script or interpreter should work like this:

   import pyddt.util.general as pug
   pug.breakdown_output_filename()

or:

    In [1]: import pyddt.drivers.sensitivity as psd
    In [2]: psd.Sensitivity()

The other major thing to change is moving the cl
[…]
