+++ src/Richards.cpp
@@ -36,6 +36,7 @@ void Richards::update(Layer *fstsoill, Layer* bdrainl,
   qinfil = infil;
   qevap  = evap;
   qdrain = 0.0;
+  qgroundwater = 0.0;
 
   if (bdraindepth <= fstsoill->z) {// no percolation
     if(qinfil > 0.0){//add infil back to ponding/qover
@@ -224,6 +225,13 @@ void Richards::update(Layer *fstsoill, Layer* bdrainl,
   double eq7103_den = 0.0;
   bool sat_soil = false;//If there is at least one saturated layer
 
+  // Using an inverted for of the above equations to calculate an
+  // approximate subsurface input in unsaturated layers. This is
+  // to be used for estimating groundwater inputs in lowlands.
+  double eq7103_unsat_num = 0.0;
+  double eq7103_unsat_den = 0.0;
+  bool unsat_soil = false; // If there is at least one saturated layer
+
   for(int ind=topind; ind<=drainind; ind++){
     //For any saturated layer
     if(vol_liq[ind] / liq_poro[ind] >= 0.9){
@@ -233,6 +241,10 @@ void Richards::update(Layer *fstsoill, Layer* bdrainl,
       //but impedance was removed to achieve sufficient drainage
       eq7103_num += ksat[ind] * dzmm[ind]/1.e3;
       eq7103_den += dzmm[ind]/1.e3;
+    } else {
+      unsat_soil = true;
+      eq7103_unsat_num += imped[ind] * ksat[ind] * dzmm[ind] / 1.e3;
+      eq7103_unsat_den += dzmm[ind] / 1.e3;
     }
   }
 
@@ -277,61 +289,48 @@ void Richards::update(Layer *fstsoill, Layer* bdrainl,
         currl = currl->nextl;
       }
     }
-  } else { // for unsaturated layers
-    double fgroundwater = 1.0;
+  } 
+  else { // for unsaturated layers
+
+    double fgroundwater = 1000.0;
      //CLM5 Equations 7.103 and 7.102
     double slope_rads = cell_slope * PI / 180;//Converting to radians
     double kgroundwater_perch = 10.e-5* sin(slope_rads) //original factor 10e-5; to be adjusted per S.Swenson
-                        * (eq7103_num / eq7103_den);
+                        * (eq7103_unsat_num / eq7103_unsat_den);
     double qgroundwater_perch = kgroundwater_perch * (bdraindepth - watertab)
                         * fgroundwater; // adding a ground water input factor to unsaturated layers
     
-    //Calculate input to unsaturated layers
+    // Calculate liquid space available in each layer
     currl = fstsoill;
     while(currl != NULL && currl->indl<=drainind && currl->isSoil){
       int ind = currl->indl;
       if(vol_liq[ind] / liq_poro[ind] <= 0.9){
-        double layer_max_input = currl->poro - effminliq[ind];
+        double layer_max_input = (currl->poro * DENLIQ * currl->dz) - currl->liq - effminliq[ind];
         double layer_calc_input = qgroundwater_perch * SEC_IN_DAY
-                                * ((dzmm[ind]/1.e3) / eq7103_den);
+                                * ((dzmm[ind]/1.e3) / eq7103_unsat_den);
         layer_input[ind] = fmin(layer_max_input, layer_calc_input);
-        qdrain -= layer_input[ind]; //mm/day
+        qgroundwater += layer_input[ind]; //mm/day
       }
       currl = currl->nextl;
     }
-    //Drainage has been calculated from saturated layers (usually bottom layers);
-    //however, to avoid creating isolated unsaturated deep layers,
-    //remove the calculated drainage from the top down.
-    if(qdrain > 0){
-      double to_drain = qdrain;
[…]
