### 2025-01-24 01c06e6
Fixing NaNs in nitrogen reassignment
Starting to add a section within to the documentation.
