+++ docs_src/sphinx/source/model_overview.rst
@@ -243,6 +243,18 @@ modified due to wildfire.
 
 
 *********************
+Processes
+*********************
+
+Adding process descriptions here...
+
+
+========
+CMT Change
+========
+
+
+*********************
 Inputs/Outputs (IO)
 *********************
 
