### 2025-01-27 201889a
updated values for CMT66
