### 2025-01-28 13d318c
loading only new vegetation parameters
adding a function to load vegetation parameters
from calparbgc without changing soil parameters.
