### 2025-01-28 80b02e0
Added assignCalparBgcSoil function
