+++ include/CohortLookup.h
@@ -27,6 +27,7 @@ public:
   void init();
   void assignBgcCalpar(string & dirname);
   void assignBgcCalparVeg(string & dirname);
+  void assignBgcCalparSoil(std::string & dirname);
   // For loading only vegetation related parameters
   // following a cmt change
   void loadVegetationParams();
