### 2025-01-29 281e425
moved files to match reccomended directory layout, tweaked .toml file, notes.
puts all source files in a subdirectory named src

This allowed us to do

    # (from root of dvm-dos-tem repo)
    pip install -e pyddt

And then in python to do

    import pyddt.util.output as puo
    puo.get_last_n_eq(...)

 etc.

Have not addressed dependencies, etc yet...
