+++ pyddt/TEMPORARY_notes.yaml
@@ -12,10 +12,11 @@ pyDDT
 
 
 
-pyddt
+pyddt/
+
   README
   pyproject.toml
-  
+
   docs/
 
   tests/
@@ -23,27 +24,29 @@ pyddt
     doctests/
       doctests_outspec_utils.md
 
-  util
-    general.py
-    metrics.py
-    param.py
-    runmask.py
-    outspec.py
-    
-  drivers
-    BaseDriver.py
-    MadsTEMDriver.py
-    Sensitivity.py
-  
-  calibration
-    ca-config-demo.yaml
-    AC-MADS-TEM.jl
-    AC-SCIPY-TEM.py
+  src/
+    pyddt/
+      util
+        general.py
+        metrics.py
+        param.py
+        runmask.py
+        outspec.py
+        
+      drivers
+        BaseDriver.py
+        MadsTEMDriver.py
+        Sensitivity.py
+      
+      calibration
+        ca-config-demo.yaml
+        AC-MADS-TEM.jl
+        AC-SCIPY-TEM.py
 
-  sensitivity
-    SA_setup_and_run.py
-    SA_post_hoc_analysis.py
-    sa-demo-config.yaml
+      sensitivity
+        SA_setup_and_run.py
+        SA_post_hoc_analysis.py
+        sa-demo-config.yaml
   
   
 
