### 2025-01-30 db3e82c
Implements loading atmospheric ch4 from input files
