+++ src/EnvData.cpp
@@ -891,6 +891,8 @@ void EnvData::update_from_climate(const Climate& clm, const int midx, const int 
   this->d_atmd.svp = clm.svp_d[doy_idx];
   this->d_atmd.vpd = clm.vpd_d[doy_idx];
 
+  this->atmospheric_ch4 = clm.ch4_d;
+
   // Adjust days-since-rain
   if(this->d_a2l.prec > 0.0) {
     this->d_atms.dsr = 0;
