+++ src/ModelData.cpp
@@ -97,6 +97,8 @@ ModelData::ModelData(Json::Value controldata):force_cmt(-1) {
   soil_texture_file = controldata["IO"]["soil_texture_file"].asString();
   co2_file          = controldata["IO"]["co2_file"].asString();
   proj_co2_file     = controldata["IO"]["proj_co2_file"].asString();
+  ch4_file          = controldata["IO"]["ch4_file"].asString();
+  proj_ch4_file     = controldata["IO"]["proj_ch4_file"].asString();
   runmask_file      = controldata["IO"]["runmask_file"].asString();
   output_dir        = controldata["IO"]["output_dir"].asString();
   output_spec_file  = controldata["IO"]["output_spec_file"].asString();
