+++ include/EnvData.h
@@ -88,6 +88,9 @@ public:
   snw2atm_env y_snw2a;
   snw2soi_env y_snw2soi;
 
+  //The atmospheric ch4 for the year
+  double atmospheric_ch4;
+
   //Arrays to hold a month's worth of variables for daily
   // netCDF output.
   double daily_eet[31];
