+++ include/ModelData.h
@@ -77,6 +77,8 @@ public:
   string drainage_file;
   string co2_file;
   string proj_co2_file;
+  string ch4_file;
+  string proj_ch4_file;
   string runmask_file;
   string output_dir;
   string output_spec_file;
