### 2025-02-04 9bdba25
Testing groundwater input formulation (in progress)
