### 2025-02-06 09e7a5a
Update runmask.py and associated doctest to use new pacakge layout.
