### 2025-02-06 9adf438
Fix setup_working_directory.py and associated tests to new package layout.
