+++ pyddt/src/pyddt/drivers/madstem.py
@@ -6,16 +6,13 @@ import subprocess
 import shutil
 import pandas as pd
 
-#import pyddt.drivers.BaseDriver
 from . import basedriver
 
-# import util.param as pu
-# import util.setup_working_directory
-# import util.runmask
-# import util.outspec
-# import util.output
-
-# from drivers.BaseDriver import BaseDriver
+from ..util import param
+from ..util import setup_working_directory
+from ..util import runmask
+from ..util import outspec
+from ..util import output
 
 class MadsTEMDriver(basedriver.BaseDriver):
   '''
@@ -92,12 +89,12 @@ class MadsTEMDriver(basedriver.BaseDriver):
     assert len(self.paramnames) == len(self.pftnums), "params list and pftnums list must be same length!"
 
     self.params = []
-    plu = pu.build_param_lookup(self._seedpath)
+    plu = param.build_param_lookup(self._seedpath)
 
     for pname, pftnum in zip(self.paramnames, self.pftnums):
-      original_pdata_file = pu.which_file(self._seedpath, pname, lookup_struct=plu)
-      p_db = pu.get_CMT_datablock(original_pdata_file, self.cmtnum)
-      p_dd = pu.cmtdatablock2dict(p_db)
+      original_pdata_file = param.which_file(self._seedpath, pname, lookup_struct=plu)
+      p_db = param.get_CMT_datablock(original_pdata_file, self.cmtnum)
+      p_dd = param.cmtdatablock2dict(p_db)
       if pname in p_dd.keys():
         p_val = p_dd[pname]
       else:
@@ -120,14 +117,14 @@ class MadsTEMDriver(basedriver.BaseDriver):
     run_dir = os.path.join(self.work_dir, 'run')
 
     # Make the working directory
-    util.setup_working_directory.cmdline_entry([
+    setup_working_directory.cmdline_entry([
       '--input-data-path', self.site, 
       '--seed-parameters', self._seedpath,
       run_dir
     ])
 
     # Adjust run mask for appropriate pixel
-    util.runmask.cmdline_entry([
+    runmask.cmdline_entry([
       '--reset',
       '--yx',self.PXy, self.PXx,
       '{}/run-mask.nc'.format(run_dir)
@@ -137,14 +134,14 @@ class MadsTEMDriver(basedriver.BaseDriver):
     # copy of the template output spec file.
     # maybe a flag should be added to setup_working_directory.py that can
     # enforce starting with an empty outspec file...
-    util.outspec.cmdline_entry([
+    outspec.cmdline_entry([
       '{}/config/output_spec.csv'.format(run_dir),
       '--empty',
     ])
 
     # Next run thru the requested outputs toggling them on in the file.
     for output_spec in self.outputs:
-      util.outspec.cmdline_entry([
+      outspec.cmdline_entry([
         '{}/config/output_spec.csv'.format(run_dir),
         '--on', output_spec['ncname'], 
         output_spec['timeres'],
@@ -155,7 +152,7 @@ class MadsTEMDriver(basedriver.BaseDriver):
 
     # Setup all the auxillary outputs...
     for output_spec in self.aux_outputs:
-      util.outspec.cmdline_entry([
+      outspec.cmdline_entry([
         '{}/config/output_spec.csv'.format(run_dir),
         '--on', output_spec['ncname'], 
         output_spec['timeres'],
@@ -166,13 +163,13 @@ class MadsTEMDriver(basedriver.BaseDriver):
 
 
     # Make sure CMTNUM output is on
-    util.outspec.cmdline_entry([
+    outspec.cmdline_entry([
       '{}/config/output_spec.csv'.format(run_dir),
       '--on','CMTNUM','y'
     ])
 
[…]
