### 2025-02-06 029129a
Adding ch4 files to setup_working_directory.py
