### 2025-02-06 3161861
Adding unit conversion for atmospheric ch4 ppb->umol/L
Also testing some ebullition mechanisms
