### 2025-02-07 9f54305
Reviewing initial declaration of variables
