### 2025-02-13 acfb57c
adding doxygen strings and some clean up
Also prepped ebul_efflux variable to be output
need to work on structs to output correctly
