### 2025-02-14 uaf-arctic-eco-modeling/dvm-dos-tem#767 @kdorheim
Installation instructions review for JOSS
Hello, this is related to [openjournals/joss-reviews#7667](https://github.com/openjournals/joss-reviews/issues/7667). I am following along with the installation instructions and have run into a few problems/questions, one of which  (question 3) may hold up my review.

1. Could you please specify which version of docker desktop I should be using? It seems like the first set of issues I ran into were resolved when I reinstalled it but I would like to confirm ```Docker version 27.1.1, build 63125853e3``` that this is the correct version to use. If there is a docker version requirement, it could be helpful to include it in the installation instructions.

2. When I build the docker image I get a message related to 6 warnings, are these warnings okay to ignore?

```
 6 warnings found (use docker --debug to expand):
 - FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 65)
 - FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 161)
 - FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 193)
 - FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 240)
 - FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 262)
 - FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 49)
```

3. When I run `V_TAG=$(git describe) docker compose up -d dvmdostem-dev` I get the following error

```
...
=> ERROR [dvmdostem-dev internal] load metadata for docker.io/library/cpp-dev:latest                                                                                        1.1s
 => CANCELED [dvmdostem-dev internal] load metadata for docker.io/library/dvmdostem-dev:latest                                                                               1.1s
 => [dvmdostem-dev auth] library/cpp-dev:pull token for registry-1.docker.io                                                                                                0.0s
 => [dvmdostem-dev auth] library/develop:pull token for registry-1.docker.io                                                                                                 0.0s
 => [dvmdostem-dev auth] library/dvmdostem-dev:pull token for registry-1.docker.io                                                                                  0.0s
------
 > [dvmdostem-dev internal] load metadata for docker.io/library/cpp-dev:latest:
------
failed to solve: cpp-dev:latest: failed to resolve source metadata for docker.io/library/cpp-dev:latest: pull access denied, repository does not exist or may require authorization: server message: insufficient_scope: authorization failed
```
Any suggestions for a work around?
