### 2025-02-14 uaf-arctic-eco-modeling/dvm-dos-tem#767 @tobeycarman
@kdorheim shoot. Lets see:

re: (1), docker version I am not aware of a specific requirement, I am using the following with success:
```
$ docker --version
Docker version 24.0.2, build cb74dfc
```

re: (2), I think the warnings are safe to ignore, but looks like we should standardize the `FROM/AS` case in the Dockerfile.

re: (3), hmm, I think this comes up when your images did not build correctly and or you have a mis-match between your checked out version of the code and the `V_TAG` environment variable. The `V_TAG` env variable can be set and passed several ways; in the example you showed above, it is set only for that command (not exported). Alternatively you could export the environment variable (making it available to subsequent commands in the same shell) or use the `.env` file. I have more info on this error in issue #636.
