### 2025-02-14 uaf-arctic-eco-modeling/dvm-dos-tem#767 @tobeycarman
@kdorheim note that in the QuickStart directions, there are a bunch of steps that are labeled "(optional)" - most of this comes into play when you are doing "real" work where you start needing to run command more than one time (i.e export the `V_TAG` or use the `.env` file rather than prefixing every command with `V_TAG=$(git describe)`  and keeping your inputs and outputs in locations separate from your source code folder.
