### 2025-02-17 uaf-arctic-eco-modeling/dvm-dos-tem#767 @kdorheim
Thanks @tobeycarman for getting back to me so quickly. I am trying to follow along with the sample run included in the getting started section but am running into some issues. After uninstalling and reinstalling docker I am still running into some docker issues

running `./docker-build-wrapper.sh cpp-dev && ./docker-build-wrapper.sh dev`

```
(base) dorh012@NNNN dvm-dos-tem % docker images
REPOSITORY   TAG                   IMAGE ID       CREATED          SIZE
cpp-dev      v0.8.1-10-g1984a49f   fc7417110cb7   16 minutes ago   940MB
ubuntu       focal                 8e5c4f0285ec   4 months ago     101MB
```

trying to follow along the instruction and running `V_TAG=$(git describe) docker compose up -d dvmdostem-dev` returns

```
WARN[0000] The "DDT_INPUT_CATALOG" variable is not set. Defaulting to a blank string.
WARN[0000] The "DDT_WORKFLOWS" variable is not set. Defaulting to a blank string.
WARN[0000] /Users/dorh012/Documents/2025/dvm-dos-tem/docker-compose.yml: the attribute `version` is obsolete, it will be ignored, please remove it to avoid potential confusion
[+] Running 1/1
 ! dvmdostem-dev Warning pull access denied for dvmdostem-dev, repository does not exist or may require 'docker login'                                                                                                                     2.0s
[+] Building 0.3s (7/7) FINISHED                                                                                                                                                                                                 docker:default
 => [dvmdostem-dev internal] load build definition from Dockerfile                                                                                                                                                                         0.0s
 => => transferring dockerfile: 16.89kB                                                                                                                                                                                                    0.0s
 => WARN: FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 49)                                                                                                                                                            0.0s
 => WARN: FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 65)                                                                                                                                                            0.0s
 => WARN: FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 161)                                                                                                                                                           0.0s
 => WARN: FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 193)                                                                                                                                                           0.0s
 => WARN: FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 240)                                                                                                                                                           0.0s
 => WARN: FromAsCasing: 'as' and 'FROM' keywords' casing do not match (line 262)                                                                                                                                                           0.0s
 => ERROR [dvmdostem-dev internal] load metadata for docker.io/library/develop:latest                                                                                                                                                      0.2s
 => ERROR [dvmdostem-dev internal] load metadata for docker.io/library/cpp-dev:latest                                                                                                                                                      0.2s
 => CANCELED [dvmdostem-dev internal] load metadata for docker.io/library/dvmdostem-dev:latest                                                                                                                                             0.2s
 => [dvmdostem-dev auth] library/dvmdostem-dev:pull token for registry-1.docker.io                                                                                                                                                         0.0s
 => [dvmdostem-dev auth] library/cpp-dev:pull token for registry-1.docker.io                                                                                                                                                               0.0s
 => [dvmdostem-dev auth] library/develop:pull token for registry-1.docker.io                                                                                                                                                               0.0s
------
 > [dvmdostem-dev internal] load metadata for docker.io/library/develop:latest:
------
------
 > [dvmdostem-dev internal] load metadata for docker.io/library/cpp-dev:latest:
------
failed to solve: develop: failed to resolve source metadata for docker.io/library/develop:latest: pull access denied, repository does not exist or may require authorization: server message: insufficient_scope: authorization failed

```

The errors related to `dvmdostem-dev` are where I think I am getting stuck. Is the `dvmdostem-dev` image publicly available on the docker hub? I am somewhat familiar with docker but am by no means an expert, so I may be missing an important docker step. I am curious if @arbennett has gotten to work or if they have run into similar problems.
