### 2025-02-17 #7667 @kdorheim
> I can also have my review done by Feb 25! Apologies for the delays.

@arbennett I am having issues getting dvmdostem installed and running, see [this issue](https://github.com/uaf-arctic-eco-modeling/dvm-dos-tem/issues/767) and was curious if you have had any success.
