### 2025-02-18 uaf-arctic-eco-modeling/dvm-dos-tem#767 @tobeycarman
@kdorheim ah ha! OK, I think that something happened when building the images - we use a set of layered images so that you don't have to constantly be rebuilding the lower layers. This is what the `cpp-dev` image is. But you need the `dvmdostem-dev` image, which is built on top of the `cpp-dev` image, hence the two subsequent invocations of `docker-build-wrapper.sh`. But what I notice is that when you run `docker images`, I am not seeing the `dvmdostem-dev` image listed, so I am thinking that something went wrong building it. I would try again to run `./docker-build-wrapper.sh dev`; it should pick up where it left off by using the already-built `cpp-dev` image. It might take 10 minutes or so to build the `dvmdostem-dev` image (depending on your network speed, etc). Once that succeeds you should see it in the output of `docker images`. Then the subsequent commands should start working.

Re: a publicly available image, we have not managed to push images to a container registry yet, although I think that would help a lot in this case. Still working on the process for automatically building and publishing images when we release, this issue is making me think it is higher priority...

Thanks so much for having the patience to try this!
