### 2025-02-18 uaf-arctic-eco-modeling/dvm-dos-tem#768 @kdorheim
Automated Testing for KD JOSS Review
Hello! I am working on my checklist for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7667#issuecomment-2652398626). This specific github issue concerns the **Automated Testing** item on my checklist.

It sounds like JOSS is fine with the absence of automated tests and I would be able to check off this item on their list but the tests directory makes me think that there is some sort of testing that is going on. So I wanted to clarify, are the shell scripts in the tests directory being run on GitHub actions, Jenkins, or are they now depreciate?
