### 2025-02-18 uaf-arctic-eco-modeling/dvm-dos-tem#769 @kdorheim
Documentation for KD JOSS Review
Hello! I am working on my checklist for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7667#issuecomment-2652398626), this github issue is related to **Documentation** element of my checklist.
****
Should I be reviewing the [user guide](https://uaf-arctic-eco-modeling.github.io/dvm-dos-tem/examples_and_tutorials/basic_model_run.html) or the [wiki](https://github.com/uaf-arctic-eco-modeling/dvm-dos-tem/wiki)?
