### 2025-02-18 uaf-arctic-eco-modeling/dvm-dos-tem#768 @tobeycarman
@kdorheim everything in `tests/` is deprecated, although I don't think this is explicitly noted anywhere (until now 😄).

In the last several years we have been using the files in `scripts/tests`. There are not automated yet. The current usage is discussed in the [User Guide > Dev Info > Testing and Deployment](https://uaf-arctic-eco-modeling.github.io/dvm-dos-tem/software_development_info.html#testing-and-deployment) section.
