### 2025-02-18 uaf-arctic-eco-modeling/dvm-dos-tem#767 @kdorheim
@tobeycarman thank you, yes after running that the `dvmdostem-dev` properly builds!  `docker images` now returns

```
REPOSITORY      TAG                   IMAGE ID       CREATED          SIZE
dvmdostem-dev   v0.8.1-10-g1984a49f   8858134836fe   42 seconds ago   5.39GB
cpp-dev         v0.8.1-10-g1984a49f   688342a68732   6 minutes ago    940MB
```
