### 2025-02-18 uaf-arctic-eco-modeling/dvm-dos-tem#767 @kdorheim
@tobeycarman alrighty I am now running into new issues (so progress!) with `V_TAG=$(git describe) docker compose up -d dvmdostem-dev`

it returns the following

```
WARN[0000] The "DDT_INPUT_CATALOG" variable is not set. Defaulting to a blank string.
WARN[0000] The "DDT_WORKFLOWS" variable is not set. Defaulting to a blank string.
WARN[0000] /Users/dorh012/Documents/2025/dvm-dos-tem/docker-compose.yml: the attribute `version` is obsolete, it will be ignored, please remove it to avoid potential confusion
[+] Running 0/1
 ⠹ Container dvm-dos-tem-dvmdostem-dev-1  Starting                                                                                                                                           0.1s
Error response from daemon: error while mounting volume '/var/lib/docker/volumes/dvm-dos-tem_inputcatalog/_data': missing device in volume options
```
