### 2025-02-18 uaf-arctic-eco-modeling/dvm-dos-tem#768 @kdorheim
Ah! okay here are some follow-up suggestions and questions

(1) Remove the tests directory since the tests are depreciated. If it would be useful for those tests to stick around for a bit while you are implementing new testing suites, stick them in a legacy code directory or change the name for users that are trying to understand the model by diving into the code 🤷‍♂ - just a suggestion
(2) Looking at the `.github/workflows/main_sample.yml`, the [actions v2](https://github.com/actions/cache) is being depreciated. consider upgrading now (adding it to my own todo for my repos) so you don't have to do it latter
(3) Also related to `.github/workflows/main_sample.yml` please correct me if I am wrong but it looks like it is only testing if the image of one of the dependencies builds? Is that correct?
