### 2025-02-18 uaf-arctic-eco-modeling/dvm-dos-tem#768 @tobeycarman
Suggestions welcome! It is great to get the perspective from someone unfamiliar with the project and all the little oddities that we have internalized and forgotten about.

re: (3), yes the GitHub actions stuff is just a stub - so doesn't do anything too meaningful - yet. Our build times and test run times are large enough that we quickly bump out of the free usage tiers, which makes it more complicated to add more tests (either more engineering time to figure out how to make the builds/tests more efficient or tracking down funding source for the paid tiers). So for now it's a stub - all the testing is happening by individual developers building and running the docker images and the tests in `scripts/tests`.

Somewhat related, we have a [branch in progress](https://github.com/uaf-arctic-eco-modeling/dvm-dos-tem/tree/reorg2), (no PR yet) that will significantly re-factor the layout of the `scripts/` directory and make available a "`pyddt`" Python package. This will be a step toward getting the project organized in a more standard way that will make integration with container registries and CI/CD pipelines easier.
