### 2025-02-18 uaf-arctic-eco-modeling/dvm-dos-tem#770 @kdorheim
Software paper for KD JOSS Review
Hello! I am working on my checklist for the [JOSS review]
(https://github.com/openjournals/joss-reviews/issues/7667#issuecomment-2652398626), this github issue addresses some of the elements of the **Software paper** portion of my checklist.
****

**Statement of Need**
In lines 43–46, the authors mention that the model has evolved over time and presumably cite previous literature that has used it. The Acknowledgements section also suggests that the current version builds on prior work. However, the statement of need in the JOSS manuscript and the main READ me of the repository is currently unclear.

Is this manuscript intended to document a specific version of the model (v0.8.1)? If so, it would be helpful to explicitly describe the differences between this release and the previous major version. Alternatively, if this is the first formal documentation of DVM-DOS-STEM, that should be clearly stated. Either scenario provides a compelling justification for publication, but a more explicit statement of need is currently missing.

Additionally, I encourage the authors to include the model version number in the title, as this would help differentiate the publication/release from future versions of DVM-DOS-STEM that are likely in development.

**References**
There are some “?” in citations throughout the paper (see the footnote on page one and citations in like 46) are these notes to include additional citations or an issue related to compiling the paper?
From my POV I think the authors cite good references but additional might be needed if they choose to address my comments r.w.t the State of the field

**State of the field**
Currently, this section is missing, and I'm not sure if it is crucial to include. However, referencing other ecosystem models, such as FATES or ED, could be helpful to readers/users. Would DVM-DOS-EM fall into the same modeling category as these other two? Providing readers with context—whether through similarities or differences to other models—could enhance their understanding of potential applications of DVM-DOS-EM.
