### 2025-02-18 uaf-arctic-eco-modeling/dvm-dos-tem#767 @tobeycarman
@kdorheim Ah ha, ok, so sometimes there are issues with Docker volumes that linger from earlier failed attempts. I would start by running `docker volume ls`. Then I'd delete any dvmdostem volumes you have: `docker volume rm <VOLUME_NAME>`.

If that doesn't work, then maybe it is a problem with the empty `DDT_INPUT_CATALOG` and `DDT_WORKFLOWS` variables.
You shouldn't need these - they are for "real work" where you want to mount your (larger) storage folders to the docker container.

In the simple test/demo case, you should not need them - the source code folder is mounted inside the container and for the simple demo, this should be sufficient as it will make the `demo-data/` directory available; the default run should work with this data and put the output in `demo-data/output`

If after deleting the lingering volumes, you still have problems, I guess you'd need to set those two offending environment variables.
