### 2025-02-18 uaf-arctic-eco-modeling/dvm-dos-tem#768 @kdorheim
Re - `pyddt` that is awesome!

Okay I would be satisfied with checking off the automated testing on the JOSS list if you were to include a statement on the main repository READ ME that no CI is presently in use but that it is an issue to address in the future.
