### 2025-02-18 uaf-arctic-eco-modeling/dvm-dos-tem#767 @kdorheim
@tobeycarman to set the `DDT_INPUT_CATALOG` and `DDT_WORKFLOWS` manually what do you mean by that? Could you include example for what you would do on your machine?
