### 2025-02-19 uaf-arctic-eco-modeling/dvm-dos-tem#767 @kdorheim
Hmmm, @tobeycarman, thanks for troubleshooting with me. I kept running into issues with the quick starter instructions. Here is the documentation of the steps I had to take to make some progress. I guess my question for you is, does this look like the expected process/terminal output a user should expect when walking through the quick start?

****
Install git and docker (for me it was for mac)
```
(base) user@machine~ % git version
git version 2.39.3 (Apple Git-146)
(base) user@machine~ % docker version
Client: Docker Engine - Community
 Version:           27.1.1
 API version:       1.46
 Go version:        go1.22.5
 Git commit:        63125853e3
 Built:             Fri Jul 19 17:35:01 2024
 OS/Arch:           darwin/arm64
 Context:           desktop-linux

Server: Docker Desktop 4.38.0 (181591)
 Engine:
  Version:          27.5.1
  API version:      1.47 (minimum version 1.24)
  Go version:       go1.22.11
  Git commit:       4c9b3b0
  Built:            Wed Jan 22 13:41:25 2025
  OS/Arch:          linux/arm64
  Experimental:     false
 containerd:
  Version:          1.7.25
  GitCommit:        bcc810d6b9066471b0b6fa75f557a15a1cbf31bb
 runc:
  Version:          1.1.12
  GitCommit:        v1.1.12-0-g51d5e946
 docker-init:
  Version:          0.19.0
  GitCommit:        de40ad0
```

Clone the dvm-dos-tes repository and switch to the joss review branch

```
git clone https://github.com/uaf-arctic-eco-modeling/dvm-dos-tem.git
git switch joss-paper-2
```

I defined `DDT_INPUT_CATALOG`, `DDT_WORKFLOWS`, `V_TAG` in my .bash_profile (my version of .env) as the following and created the my-personal-runs directory.
```
export DDT_INPUT_CATALOG="/Users/name/Documents/2025/dvm-dos-tem/demo-data"
export DDT_WORKFLOWS="/Users/name/Documents/2025/dvm-dos-tem/my-personal-runs"
export V_TAG="v0.8.1"
```
cd to the root `dvm-dos-tem` directory.

then ran the following, the of the command line included in the quick starter instructions did not work for me and the user MUST be logged into docker account otherwise there are permission errors!
```
./docker-build-wrapper.sh cpp-dev
./docker-build-wrapper.sh dev
```

the following docker images were created
```
(base) user@machinedvm-dos-tem % docker images
REPOSITORY      TAG                   IMAGE ID       CREATED        SIZE
dvmdostem-dev   v0.8.0-87-g8becb7c5   4b3be9f6f63c   21 hours ago   5.39GB
cpp-dev         v0.8.0-87-g8becb7c5   9e140745bec6   21 hours ago   940MB
```

login to docker again, otherwise will fail

```
(base) user@machinedvm-dos-tem % V_TAG=$(git describe) docker compose up -d dvmdostem-dev
WARN[0000] /Users/name/Documents/2025/dvm-dos-tem/docker-compose.yml: the attribute `version` is obsolete, it will be ignored, please remove it to avoid potential confusion
[+] Running 5/5
 ✔ Network dvm-dos-tem_default            Created                                                                                                                                                                                    0.1s
 ✔ Volume "dvm-dos-tem_inputcatalog"      Created                                                                                                                                                                                    0.0s
 ✔ Volume "dvm-dos-tem_workflows"         Created                                                                                                                                                                                    0.0s
 ✔ Volume "dvm-dos-tem_sourcecode"        Created                                                                                                                                                                                    0.0s
 ✔ Container dvm-dos-tem-dvmdostem-dev-1  Started
```

Now checking the docker setup we have...
```
(base) user@machinedvm-dos-tem % docker container ls
CONTAINER ID   IMAGE                               COMMAND       CREATED              STATUS              PORTS                    NAMES
6dd952411b65   dvmdostem-dev:v0.8.0-87-g8becb7c5   "/bin/bash"   About a minute ago   Up About a minute   0.0.0.0:7001->7001/tcp   dvm-dos-tem-dvmdostem-dev-1
(base) user@machinedvm-dos-tem % docker volume ls
DRIVER    VOLUME NAME
local     dvm-dos-tem_inputcatalog
local     dvm-dos-tem_sourcecode
local     dvm-dos-tem_workflows
```

Then run `docker compose exec dvmdostem-dev bash`
```
(base) user@machinedvm-dos-tem % docker compose exec dvmdostem-dev bash
WARN[0000] /Users/name/Documents/2025/dvm-dos-tem/docker-compose.yml: the attribute `version` is obsolete, it will be ignored, please remove it to avoid potential confusion
```

now can run `make`

```
develop@6dd952411b65:/work$ make
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/TEM.cpp -o obj/TEM.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/TEMLogger.cpp -o obj/TEMLogger.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/CalController.cpp -o obj/CalController.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/ArgHandler.cpp -o obj/ArgHandler.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/TEMUtilityFunctions.cpp -o obj/TEMUtilityFunctions.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Climate.cpp -o obj/Climate.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/OutputEstimate.cpp -o obj/OutputEstimate.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Runner.cpp -o obj/Runner.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/BgcData.cpp -o obj/BgcData.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/CohortData.cpp -o obj/CohortData.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/EnvData.cpp -o obj/EnvData.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/EnvDataDly.cpp -o obj/EnvDataDly.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/FireData.cpp -o obj/FireData.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/RestartData.cpp -o obj/RestartData.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/WildFire.cpp -o obj/WildFire.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/DoubleLinkedList.cpp -o obj/DoubleLinkedList.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Ground.cpp -o obj/Ground.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/MineralInfo.cpp -o obj/MineralInfo.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Moss.cpp -o obj/Moss.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Organic.cpp -o obj/Organic.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Snow.cpp -o obj/Snow.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/SoilParent.cpp -o obj/SoilParent.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Vegetation.cpp -o obj/Vegetation.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/CohortLookup.cpp -o obj/CohortLookup.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Cohort.cpp -o obj/Cohort.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Integrator.cpp -o obj/Integrator.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/ModelData.cpp -o obj/ModelData.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Richards.cpp -o obj/Richards.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Snow_Env.cpp -o obj/Snow_Env.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Soil_Bgc.cpp -o obj/Soil_Bgc.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Soil_Env.cpp -o obj/Soil_Env.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/SoilParent_Env.cpp -o obj/SoilParent_Env.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Stefan.cpp -o obj/Stefan.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/TemperatureUpdator.cpp -o obj/TemperatureUpdator.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/CrankNicholson.cpp -o obj/CrankNicholson.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/tbc-debug-util.cpp -o obj/tbc-debug-util.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Vegetation_Bgc.cpp -o obj/Vegetation_Bgc.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Vegetation_Env.cpp -o obj/Vegetation_Env.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/Layer.cpp -o obj/Layer.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/MineralLayer.cpp -o obj/MineralLayer.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/MossLayer.cpp -o obj/MossLayer.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/OrganicLayer.cpp -o obj/OrganicLayer.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/ParentLayer.cpp -o obj/ParentLayer.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/SnowLayer.cpp -o obj/SnowLayer.o
g++ -c -ansi -g -gdwarf-2 -std=c++11 -fPIC -DBOOST_ALL_DYN_LINK -Werror  -DGIT_SHA=\"v0.8.0-87-g8becb7\"   -I/usr/include/jsoncpp  src/SoilLayer.cpp -o obj/SoilLayer.o
g++  -o dvmdostem -I/usr/include/jsoncpp obj/ArgHandler.o obj/TEMLogger.o obj/CalController.o obj/TEMUtilityFunctions.o obj/Climate.o obj/OutputEstimate.o obj/Runner.o obj/BgcData.o obj/CohortData.o obj/EnvData.o obj/EnvDataDly.o obj/FireData.o obj/RestartData.o obj/WildFire.o obj/DoubleLinkedList.o obj/Ground.o obj/MineralInfo.o obj/Moss.o obj/Organic.o obj/Snow.o obj/SoilParent.o obj/Vegetation.o obj/CohortLookup.o obj/Cohort.o obj/Integrator.o obj/ModelData.o obj/Richards.o obj/Snow_Env.o obj/Soil_Bgc.o obj/Soil_Env.o obj/SoilParent_Env.o obj/Stefan.o obj/CrankNicholson.o obj/tbc-debug-util.o obj/Vegetation_Bgc.o obj/Vegetation_Env.o obj/Layer.o obj/MineralLayer.o obj/MossLayer.o obj/OrganicLayer.o obj/ParentLayer.o obj/SnowLayer.o obj/SoilLayer.o obj/TemperatureUpdator.o obj/TEM.o -I/usr/lib -lnetcdf -lboost_system -lboost_filesystem -lboost_program_options -lboost_thread -lboost_log -ljsoncpp -lpthread -lreadline -llapacke
```

Now launch model run
```
develop@6dd952411b65:/work$ ./dvmdostem --log-level monitor -p 100 -e 1000 -s 250 -t 115 -n 85
Setting up logging...
[monitor] [EQ] Equilibrium Initial Year Count: 1000
[monitor] [EQ] Running Equilibrium, 1000 years.
[monitor] [SP] Running Spinup, 250 years.
[monitor] [TR] Running Transient, 115 years
[monitor] [SC] Running Scenario, 85 years.
cell 0, 0 complete.
[monitor] [] Skipping cell (0, 1)
[monitor] [] Skipping cell (0, 2)
[monitor] [] Skipping cell (0, 3)
[monitor] [] Skipping cell (0, 4)
[monitor] [] Skipping cell (0, 5)
[monitor] [] Skipping cell (0, 6)
[monitor] [] Skipping cell (0, 7)
[monitor] [] Skipping cell (0, 8)
[monitor] [] Skipping cell (0, 9)
[monitor] [] Skipping cell (1, 0)
[monitor] [EQ] Equilibrium Initial Year Count: 1000
[monitor] [EQ] Running Equilibrium, 1000 years.
[monitor] [SP] Running Spinup, 250 years.
[monitor] [TR] Running Transient, 115 years
[monitor] [SC] Running Scenario, 85 years.
cell 1, 1 complete.
[monitor] [] Skipping cell (1, 2)
[monitor] [] Skipping cell (1, 3)
[monitor] [] Skipping cell (1, 4)
[monitor] [] Skipping cell (1, 5)
[monitor] [] Skipping cell (1, 6)
[monitor] [] Skipping cell (1, 7)
[monitor] [] Skipping cell (1, 8)
[monitor] [] Skipping cell (1, 9)
[monitor] [] Skipping cell (2, 0)
[monitor] [] Skipping cell (2, 1)
[monitor] [] Skipping cell (2, 2)
[monitor] [] Skipping cell (2, 3)
[monitor] [] Skipping cell (2, 4)
[monitor] [] Skipping cell (2, 5)
[monitor] [] Skipping cell (2, 6)
[monitor] [] Skipping cell (2, 7)
[monitor] [] Skipping cell (2, 8)
[monitor] [] Skipping cell (2, 9)
[monitor] [] Skipping cell (3, 0)
[monitor] [] Skipping cell (3, 1)
[monitor] [] Skipping cell (3, 2)
[monitor] [] Skipping cell (3, 3)
[monitor] [] Skipping cell (3, 4)
[monitor] [] Skipping cell (3, 5)
[monitor] [] Skipping cell (3, 6)
[monitor] [] Skipping cell (3, 7)
[monitor] [] Skipping cell (3, 8)
[monitor] [] Skipping cell (3, 9)
[monitor] [] Skipping cell (4, 0)
[monitor] [] Skipping cell (4, 1)
[monitor] [] Skipping cell (4, 2)
[monitor] [] Skipping cell (4, 3)
[monitor] [] Skipping cell (4, 4)
[monitor] [] Skipping cell (4, 5)
[monitor] [] Skipping cell (4, 6)
[monitor] [] Skipping cell (4, 7)
[monitor] [] Skipping cell (4, 8)
[monitor] [] Skipping cell (4, 9)
[monitor] [] Skipping cell (5, 0)
[monitor] [] Skipping cell (5, 1)
[monitor] [] Skipping cell (5, 2)
[monitor] [] Skipping cell (5, 3)
[monitor] [] Skipping cell (5, 4)
[monitor] [] Skipping cell (5, 5)
[monitor] [] Skipping cell (5, 6)
[monitor] [] Skipping cell (5, 7)
[monitor] [] Skipping cell (5, 8)
[monitor] [] Skipping cell (5, 9)
[monitor] [] Skipping cell (6, 0)
[monitor] [] Skipping cell (6, 1)
[monitor] [] Skipping cell (6, 2)
[monitor] [] Skipping cell (6, 3)
[monitor] [] Skipping cell (6, 4)
[monitor] [] Skipping cell (6, 5)
[monitor] [] Skipping cell (6, 6)
[monitor] [] Skipping cell (6, 7)
[monitor] [] Skipping cell (6, 8)
[monitor] [] Skipping cell (6, 9)
[monitor] [] Skipping cell (7, 0)
[monitor] [] Skipping cell (7, 1)
[monitor] [] Skipping cell (7, 2)
[monitor] [] Skipping cell (7, 3)
[monitor] [] Skipping cell (7, 4)
[monitor] [] Skipping cell (7, 5)
[monitor] [] Skipping cell (7, 6)
[monitor] [] Skipping cell (7, 7)
[monitor] [] Skipping cell (7, 8)
[monitor] [] Skipping cell (7, 9)
[monitor] [] Skipping cell (8, 0)
[monitor] [] Skipping cell (8, 1)
[monitor] [] Skipping cell (8, 2)
[monitor] [] Skipping cell (8, 3)
[monitor] [] Skipping cell (8, 4)
[monitor] [] Skipping cell (8, 5)
[monitor] [] Skipping cell (8, 6)
[monitor] [] Skipping cell (8, 7)
[monitor] [] Skipping cell (8, 8)
[monitor] [] Skipping cell (8, 9)
[monitor] [] Skipping cell (9, 0)
[monitor] [] Skipping cell (9, 1)
[monitor] [] Skipping cell (9, 2)
[monitor] [] Skipping cell (9, 3)
[monitor] [] Skipping cell (9, 4)
[monitor] [] Skipping cell (9, 5)
[monitor] [] Skipping cell (9, 6)
[monitor] [] Skipping cell (9, 7)
[monitor] [] Skipping cell (9, 8)
[monitor] [] Skipping cell (9, 9)
Total Seconds: 182
```
Then I get the following error when trying to run `scripts/plot_output_var.py --yx 0 0 --file output/GPP_yearly_tr.nc`
```
develop@6dd952411b65:/work$ ./scripts/plot_output_var.py --yx 0 0 --file output/GPP_yearly_tr.nc
bash: ./scripts/plot_output_var.py: No such file or directory
```
