### 2025-02-19 uaf-arctic-eco-modeling/dvm-dos-tem#767 @tobeycarman
@kdorheim awesome! 🥳  Looks like everything ran (phew). I think that we moved the `plot_output_var.py` script into the `scripts/viewers` directory and have not updated the instructions, so I think the following should work:

```
./scripts/viewers/plot_output_var.py --yx 0 0 --file output/GPP_yearly_tr.nc
```

Thanks for your patience in working through all this. We will review all your other feedback at our weekly meeting today and try to address everything ASAP.
