### 2025-02-20 06b986b
Adding ebul efflux output
Also in the process of fixing co2 and ch4 averaging.
There is currently a bug in the avg_over_yearly function
