+++ src/Runner.cpp
@@ -1734,11 +1734,8 @@ void Runner::output_netCDF(std::map<std::string, OutputSpec> &netcdf_outputs, in
       if(curr_spec.daily){
         for(int id=0; id<DINM[month]; id++){
 
-          double effluxebul = cohort.bdall->daily_ch4_efflux[id]
-                            - cohort.bdall->daily_ch4_effdiff[id]
-                            - cohort.bdall->daily_total_plant_ch4[id];
+          outhold.ch4effebul_for_output.push_back(cohort.bdall->daily_ch4_efflux_ebul[id]);
 
-          outhold.ch4effebul_for_output.push_back(effluxebul);
         }
 
         if(end_of_year){
@@ -1749,20 +1746,7 @@ void Runner::output_netCDF(std::map<std::string, OutputSpec> &netcdf_outputs, in
       //Monthly
       else if(curr_spec.monthly){
 
-        double ch4trans_tot = 0;
-        for(int il=0; il<MAX_SOI_LAY; il++){
-          for(int ip=0; ip<NUM_PFT; ip++){
-            if(cohort.cd.m_veg.vegcov[ip] > 0.0){
-              ch4trans_tot += cohort.bdall->m_soi2a.ch4_transport[il][ip];
-            }
-          }
-        }
-
-        double effluxebul = cohort.bdall->m_soi2a.ch4efflux
-                          - cohort.bdall->m_soi2a.ch4effdiff
-                          - ch4trans_tot;
-
-        outhold.ch4effebul_for_output.push_back(effluxebul);
+        outhold.ch4effebul_for_output.push_back(cohort.bdall->m_soi2a.ch4efflux_ebul);
 
         if(output_this_timestep){
           output_nc_3dim(&curr_spec, file_stage_suffix, &outhold.ch4effebul_for_output[0], 1, month_start_idx, months_to_output);
@@ -1772,18 +1756,8 @@ void Runner::output_netCDF(std::map<std::string, OutputSpec> &netcdf_outputs, in
       //Yearly
       else if(curr_spec.yearly){
 
-        double ch4trans_tot = 0;
-        for(int il=0; il<MAX_SOI_LAY; il++){
-          for(int ip=0; ip<NUM_PFT; ip++){
-            ch4trans_tot += cohort.bdall->y_soi2a.ch4_transport[il][ip];
-          }
-        }
-
-        double effluxebul = cohort.bdall->y_soi2a.ch4efflux
-                          - cohort.bdall->y_soi2a.ch4effdiff
-                          - ch4trans_tot;
-
-        outhold.ch4effebul_for_output.push_back(effluxebul);
+        outhold.ch4effebul_for_output.push_back(cohort.bdall->y_soi2a.ch4efflux_ebul);
+        
         if(output_this_timestep){
           output_nc_3dim(&curr_spec, file_stage_suffix, &outhold.ch4effebul_for_output[0], 1, year_start_idx, years_to_output);
           outhold.ch4effebul_for_output.clear();
