+++ src/BgcData.cpp
@@ -252,6 +252,9 @@ void BgcData::soil_endOfDay(const int& dinm, const int& doy){
   m_soi2a.ch4efflux += d_soi2a.ch4efflux;
   d_soi2a.ch4efflux = 0.;
 
+  m_soi2a.ch4efflux_ebul += d_soi2a.ch4efflux_ebul;
+  d_soi2a.ch4efflux_ebul = 0.;
+
   daily_ch4_effdiff[dom] = d_soi2a.ch4effdiff;
   m_soi2a.ch4effdiff += d_soi2a.ch4effdiff;
   d_soi2a.ch4effdiff = 0.;
@@ -291,6 +294,7 @@ void BgcData::soil_beginOfMonth() {
 
   m_soid.ch4ebulsum = 0.;
   m_soi2a.ch4efflux = 0.;
+  m_soi2a.ch4efflux_ebul = 0.;
   m_soi2a.ch4effdiff = 0.;
 };
 
@@ -317,6 +321,7 @@ void BgcData::soil_beginOfYear() {
   y_soi2a.rhsomcrsum=0.;
   y_soi2a.ch4_oxid_sum = 0.;
   y_soi2a.ch4efflux = 0.;
+  y_soi2a.ch4efflux_ebul = 0.;
   y_soi2a.ch4effdiff = 0.;
   y_soi2soi.netnminsum = 0.;
   y_soi2soi.nimmobsum  = 0.;
@@ -496,6 +501,7 @@ void BgcData::soil_endOfMonth(const int currmind) {
 
   y_soi2a.ch4_oxid_sum += m_soi2a.ch4_oxid_sum;
   y_soi2a.ch4efflux += m_soi2a.ch4efflux;
+  y_soi2a.ch4efflux_ebul += m_soi2a.ch4efflux_ebul;
   y_soi2a.ch4effdiff += m_soi2a.ch4effdiff;
 
   for (int il =0; il<MAX_SOI_LAY; il++) {
