+++ src/Soil_Bgc.cpp
@@ -645,16 +645,17 @@ void Soil_Bgc::CH4Flux(const int mind, const int id) {
   bd->daily_total_plant_ch4[id] = plant_gm2day;
 
   diff_efflux_daily = 0.0; 
-  plant_gm2day =0.0; 
+  plant_gm2day = 0.0; 
   ebul_gm2day = 0.0;
   ebul_efflux_gm2day = 0.0; 
 
   bd->d_soi2a.ch4efflux = efflux_gm2day;
+  bd->d_soi2a.ch4efflux_ebul = ebul_efflux_gm2day;
+  bd->d_soi2a.ch4effdiff = diff_efflux_gm2day;
 
   //Store daily values for output
   bd->daily_ch4_efflux[id] = bd->d_soi2a.ch4efflux;
-  bd->d_soi2a.ch4effdiff = diff_efflux_gm2day;
-
+  bd->daily_ch4_efflux_ebul[id] = bd->d_soi2a.ch4efflux_ebul;
 }
 
 /** Writes Carbon values from each of the Ground object's Layers into the bd 
