### 2025-02-21 8f0ee4b
Fixing bug in co2 and ch4 averaging functioning
