### 2025-02-26 56c6bde
conitnued debugging, ebullition method, and ebul output
