+++ pyddt/src/pyddt/drivers/sensitivity.py
@@ -19,7 +19,11 @@ from contextlib import contextmanager
 
 
 from . import basedriver
-from .. import util
+from ..util import param as pu
+from ..util import runmask as ru
+from ..util import output as ou
+from ..util import outspec osu
+
 
 @contextmanager
 def log_wrapper(message):
@@ -219,13 +223,13 @@ def params_from_seed(seedpath, params, pftnums, percent_diffs, cmtnum):
   assert len(params) == len(percent_diffs), "params list and percent_diffs list must be same length"
 
   final = []
-  plu = util.param.build_param_lookup(seedpath)
+  plu = pu.build_param_lookup(seedpath)
 
   for pname, pftnum, perturbation in zip(params, pftnums, percent_diffs):
-    original_pdata_file = util.param.which_file(seedpath, pname, lookup_struct=plu)
+    original_pdata_file = pu.which_file(seedpath, pname, lookup_struct=plu)
 
-    p_db = util.param.get_CMT_datablock(original_pdata_file, cmtnum)
-    p_dd = util.param.cmtdatablock2dict(p_db)
+    p_db = pu.get_CMT_datablock(original_pdata_file, cmtnum)
+    p_dd = pu.cmtdatablock2dict(p_db)
 
     if pname in p_dd.keys():
       p_initial = p_dd[pname]
@@ -421,7 +425,7 @@ class Sensitivity(basedriver.BaseDriver):
         return pft_verbose_name
 
       if row.pftnum >= 0 and row.pftnum < 10:
-        pft_verbose_name = util.param.get_pft_verbose_name(
+        pft_verbose_name = pu.get_pft_verbose_name(
           cmtnum=self.cmtnum, pftnum=row.pftnum, 
           lookup_path=lookup_path,
         )
@@ -765,14 +769,14 @@ class Sensitivity(basedriver.BaseDriver):
       sample_specific_folder = self._ssrf_name(idx)
 
     # Make the working directory
-    util.setup_working_directory.cmdline_entry([
+    setup_working_directory.cmdline_entry([
       '--input-data-path', self.site, 
       '--seed-parameters', self._seedpath,
       sample_specific_folder
     ])
 
     # Strip all extra CMTs from parameter files
-    util.param.cmdline_entry([
+    pu.cmdline_entry([
       '--extract-cmt',
       os.path.join(sample_specific_folder, 'parameters'),
       f'CMT{self.cmtnum:02d}',
@@ -786,7 +790,7 @@ class Sensitivity(basedriver.BaseDriver):
               os.path.join(sample_specific_folder, 'parameters'))
 
     # Adjust run mask for appropriate pixel
-    util.runmask.cmdline_entry([
+    runmask.cmdline_entry([
       '--reset',
       '--yx',self.PXy, self.PXx,
       '{}/run-mask.nc'.format(sample_specific_folder)
@@ -797,7 +801,7 @@ class Sensitivity(basedriver.BaseDriver):
     # configurable in the .yaml file...which worked, but was inefficient when
     # the user is only looking at one or two of the output variables.
     #
-    #util.outspec.cmdline_entry([
+    #outspec.cmdline_entry([
     #  '--enable-cal-vars',
     #  '{}/config/output_spec.csv'.format(sample_specific_folder),
     #])
@@ -806,14 +810,14 @@ class Sensitivity(basedriver.BaseDriver):
     # copy of the template output spec file.
     # maybe a flag should be added to setup_working_directory.py that can
     # enforce starting with an empty outspec file...
-    util.outspec.cmdline_entry([
+    outspec.cmdline_entry([
       '{}/config/output_spec.csv'.format(sample_specific_folder),
       '--empty',
     ])
 
[…]
