### 2025-02-27 fecc5bd
update version for the "checkout" action, experiment with adding more steps
