+++ .github/workflows/main_sample.yml
@@ -18,20 +18,23 @@ jobs:
   # This workflow contains a single job called "build"
   build:
     # The type of runner that the job will run on
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-focal
 
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
       # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
 
       # Runs a single command using the runners shell
       - name: Run a one-line script
         run: echo "Hello, world!"
       - name: Run thing
         run: echo $(pwd)
-      - name: test docker build
-        run: docker build --target cpp-dev --tag cpp-dev:0.0.1 . 
+      - name: build first image layer
+        run: docker-build-wrapper.sh --cpp-dev     #run: docker build --target cpp-dev --tag cpp-dev:0.0.1 . 
+      - name: build second image layer
+        run : docker-build-wrapper.sh --dev
+
 
       # Runs a set of commands using the runners shell
       - name: Run a multi-line script
