### 2025-02-27 86510c3
change ubuntu tag in GitHub action yaml file.
had the wrong descriptor for the ubuntu release.
