### 2025-02-27 05dbaf6
Adding labile and nall into loop
Need to re-initialize cohortdata as this is preventing
more PFTs being added and certain compartments growing
