+++ .github/workflows/main_sample.yml
@@ -32,6 +32,10 @@ jobs:
         run: echo $(pwd)
       - name: where am i
         run: echo "$(ls)"
+      - name: get git sha 
+        run: export GIT_VERSION=$(git rev-parse --short "$GITHUB_SHA")
+      - name: print something
+        run: echo $GIT_VERSION
       - name: build first image layer
         run: bash docker-build-wrapper.sh --cpp-dev
       - name: build second image layer
