+++ .github/workflows/main_sample.yml
@@ -37,7 +37,7 @@ jobs:
       - name: git sha
         run: echo "$(echo $GITHUB_SHA | cut -c1-7)"
       - name: get git sha 
-        run: GIT_VERSION=$(git rev-parse --short "$GITHUB_SHA")
+        run: GIT_VERSION="$(echo $GITHUB_SHA | cut -c1-7)"
       - name: print something
         run: echo $GIT_VERSION
       - name: build first image layer
