### 2025-02-27 3f7ed15
Creating copies of Wildfire called Thermokarst
This is a first pass to begin developing thermokarst modeling
capabilities.
