+++ src/Thermokarst.cpp
@@ -1,0 +1,686 @@
+/*
+ * Thermokarst.cpp
+ *
+ * NOTE: this is the first iteration for testing thermokarst in DVM-DOS-TEM (subject to many changes)
+ * also, this was based on Thermokarst.cpp so there may be some / lots of duplicate code      
+ *
+ * This is for thermokarst (or abrupt thaw) occurrence and C/N pool dynamics due to thermokarst
+ *   (1) Thermokarst occurrence must be given by input initially for testing 
+ *   (2) Eventually we hope to force using an external or integrated state and transition model
+ *
+ *   (3) C/N pools updated here through operation on 'bd' -
+ *         NOTE this 'bd' is for all PFTs and soil.
+ *         this is to say, individual PFT's C/N pools updating need to
+ *         do elsewhere
+ *
+ *   (4) ROOTFRAC is updated here through operation on 'ed' - also for ALL PFTS.
+ *
+ *   (5) soil layer structure is NOT changed here, but will re-do in
+ *         Ground::adjustSoilAfterThermokarst()
+ *
+ *   (6) NOTE: assuming ONE root pool, and its index is 'NUM_PFT_PART'
+ *         (i.e. the last).
+ *
+ */
+
+#include <string>
+#include <sstream>
+
+#include "../include/Thermokarst.h"
+
+#include "../include/TEMUtilityFunctions.h"
+#include "../include/TEMLogger.h"
+
+extern src::severity_logger< severity_level > glg;
+
+Thermokarst::Thermokarst() {}
+
+Thermokarst::~Thermokarst() {}
+
+Thermokarst::Thermokarst(const std::string& fri_fname,
+                   const std::string& exp_fname, const double cell_slope,
+                   const double cell_aspect, const double cell_elevation,
+                   const int y, const int x) {
+
+  #pragma omp critical(load_input)
+  {
+    BOOST_LOG_SEV(glg, info) << "Setting up FRI fire data...";
+    this->fri = temutil::get_scalar<int>(fri_fname, "fri", y, x);
+    this->fri_severity = temutil::get_scalar<int>(fri_fname, "fri_severity", y, x);
+    this->fri_jday_of_burn = temutil::get_scalar<int>(fri_fname, "fri_jday_of_burn", y, x);
+    this->fri_area_of_burn = temutil::get_scalar<int>(fri_fname, "fri_area_of_burn", y, x);
+  }//End critical(fri)
+
+  #pragma omp critical(load_input)
+  {
+    BOOST_LOG_SEV(glg, info) << "Setting up explicit fire data...";
+    this->exp_burn_mask = temutil::get_timeseries<int>(exp_fname, "exp_burn_mask", y, x);
+    this->exp_fire_severity = temutil::get_timeseries<int>(exp_fname, "exp_fire_severity", y, x);
+    this->exp_jday_of_burn = temutil::get_timeseries<int>(exp_fname, "exp_jday_of_burn", y, x);
+    this->exp_area_of_burn = temutil::get_timeseries<int64_t>(exp_fname, "exp_area_of_burn", y, x);
+  }//End critical(exp_fir) 
+
+  this->slope = cell_slope;
+  this->asp = cell_aspect;
+  this->elev = cell_elevation;
+
+  BOOST_LOG_SEV(glg, debug) << "Done making Thermokarst object.";
+  BOOST_LOG_SEV(glg, debug) << this->report_fire_inputs();
+
+}
+
[…]
