+++ src/Cohort.cpp
@@ -84,6 +84,10 @@ Cohort::Cohort(int y, int x, ModelData* modeldatapointer):
   this->fire = WildFire(modeldatapointer->fri_fire_file, 
     modeldatapointer->hist_exp_fire_file, this->cd.cell_slope, this->cd.cell_aspect, this->cd.cell_elevation, y, x);
 
+  // BOOST_LOG_SEV(glg, debug) << "Setup the thermokarst information, defaulting to HISTORIC explicit data.";
+  // this->fire = Thermokarst(modeldatapointer->fri_fire_file, 
+  //   modeldatapointer->hist_exp_fire_file, this->cd.cell_slope, this->cd.cell_aspect, this->cd.cell_elevation, y, x);
+
 
   this->soilenv = Soil_Env();
   
