+++ .github/workflows/main_sample.yml
@@ -24,7 +24,8 @@ jobs:
     steps:
       # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
       - uses: actions/checkout@v4
-
+      - run: git fetch --prune --unshallow
+      
       # Runs a single command using the runners shell
       - name: Run a one-line script
         run: echo "Hello, world!"
