+++ .github/workflows/main_sample.yml
@@ -29,6 +29,14 @@ jobs:
       # Runs a single command using the runners shell
       - name: Run a one-line script
         run: echo "Hello, world!"
+      - name: investigate
+        run: |
+          echo "$(logname)"
+          echo $(logname)
+          echo "$?"
+          echo "$HOSTUID"
+          echo "$HOSTGID"
+          echo "here...
       - name: Run thing
         run: echo $(pwd)
       - name: where am i
