### 2025-02-28 a7690bf
Creating a parameter for groundwater scaler
