+++ src/CohortLookup.cpp
@@ -416,7 +416,7 @@ void CohortLookup::assignEnv4Ground(string &dircmt) {
 
   // get a list of data for the cmt number
   std::list<std::string> datalist = temutil::parse_parameter_file(
-      dircmt + "cmt_envground.txt", temutil::cmtcode2num(this->cmtcode), 41
+      dircmt + "cmt_envground.txt", temutil::cmtcode2num(this->cmtcode), 42
   );
 
   // pop each line off the front of the list
@@ -442,6 +442,7 @@ void CohortLookup::assignEnv4Ground(string &dircmt) {
   temutil::pfll2data(datalist, nfactor_s);
   temutil::pfll2data(datalist, nfactor_w);
   temutil::pfll2data(datalist, initsnwtem);
+  temutil::pfll2data(datalist, inflow_factor);
 
   for (int i = 0; i < 10; i++) {
     temutil::pfll2data(datalist, initts[i]);
