+++ include/CohortLookup.h
@@ -152,6 +152,7 @@ public:
   double initsnwtem; //1 snow input: initial snow temperature
                      //  (note: initial water of snow can be from
                      //   initial thickness and density)
+  double inflow_factor; // groundwater scaler
   double initvwc[10]; //10 soil initial input for each 10 cm thickness of layers
   double initts[10];
 
