### 2025-02-28 f4ace2f
Reworking fire scripts for thermokarst
