+++ src/Thermokarst.cpp
@@ -2,7 +2,10 @@
  * Thermokarst.cpp
  *
  * NOTE: this is the first iteration for testing thermokarst in DVM-DOS-TEM (subject to many changes)
- * also, this was based on Thermokarst.cpp so there may be some / lots of duplicate code      
+ * also, this was based on Thermokarst.cpp so there may be some / lots of duplicate code.      
+ *
+ * First iteration will be focused on representing a retrogressive thaw slump, however thermokarst
+ * will be used generally throughout code for the time being.
  *
  * This is for thermokarst (or abrupt thaw) occurrence and C/N pool dynamics due to thermokarst
  *   (1) Thermokarst occurrence must be given by input initially for testing 
@@ -37,135 +40,112 @@ Thermokarst::Thermokarst() {}
 
 Thermokarst::~Thermokarst() {}
 
-Thermokarst::Thermokarst(const std::string& fri_fname,
-                   const std::string& exp_fname, const double cell_slope,
+Thermokarst::Thermokarst(const std::string& exp_fname, const double cell_slope,
                    const double cell_aspect, const double cell_elevation,
                    const int y, const int x) {
 
-  #pragma omp critical(load_input)
-  {
-    BOOST_LOG_SEV(glg, info) << "Setting up FRI fire data...";
-    this->fri = temutil::get_scalar<int>(fri_fname, "fri", y, x);
-    this->fri_severity = temutil::get_scalar<int>(fri_fname, "fri_severity", y, x);
-    this->fri_jday_of_burn = temutil::get_scalar<int>(fri_fname, "fri_jday_of_burn", y, x);
-    this->fri_area_of_burn = temutil::get_scalar<int>(fri_fname, "fri_area_of_burn", y, x);
-  }//End critical(fri)
-
-  #pragma omp critical(load_input)
-  {
-    BOOST_LOG_SEV(glg, info) << "Setting up explicit fire data...";
-    this->exp_burn_mask = temutil::get_timeseries<int>(exp_fname, "exp_burn_mask", y, x);
-    this->exp_fire_severity = temutil::get_timeseries<int>(exp_fname, "exp_fire_severity", y, x);
-    this->exp_jday_of_burn = temutil::get_timeseries<int>(exp_fname, "exp_jday_of_burn", y, x);
-    this->exp_area_of_burn = temutil::get_timeseries<int64_t>(exp_fname, "exp_area_of_burn", y, x);
-  }//End critical(exp_fir) 
+  #pragma omp critical(load_input){
+    BOOST_LOG_SEV(glg, info) << "Setting up explicit thermokarst data...";
+    this->exp_thermokarst_mask = temutil::get_timeseries<int>(exp_fname, "exp_thermokarst_mask", y, x);
+    this->exp_thermokarst_severity = temutil::get_timeseries<int>(exp_fname, "exp_thermokarst_severity", y, x);
+    this->exp_jday_of_thermokarst = temutil::get_timeseries<int>(exp_fname, "exp_jday_of_thermokarst", y, x);
+    this->exp_area_of_thermokarst = temutil::get_timeseries<int64_t>(exp_fname, "exp_area_of_thermokarst", y, x);
+  }//End critical(exp_thermokarst) 
 
   this->slope = cell_slope;
   this->asp = cell_aspect;
   this->elev = cell_elevation;
 
   BOOST_LOG_SEV(glg, debug) << "Done making Thermokarst object.";
-  BOOST_LOG_SEV(glg, debug) << this->report_fire_inputs();
+  BOOST_LOG_SEV(glg, debug) << this->report_thermokarst_inputs();
 
 }
 
[…]
