+++ src/ThermokarstData.cpp
@@ -1,0 +1,76 @@
+#include <string>
+
+#include "../include/ThermokarstData.h"
+
+ThermokarstData::ThermokarstData()
+{
+    fire_a2soi.orgn = 0.0;
+    useseverity = false;
+};
+
+ThermokarstData::~ThermokarstData() {
+};
+
+/** Returns a multi-line string showing the state of the FirData object. */
+std::string ThermokarstData::report_to_string(const std::string& msg) {
+  std::stringstream report;
+  report << "  " << msg << "  ==== FirData REPORT ====\n";
+  report << "   useseverity: " << useseverity << "\n";
+  report << "   fire_soid.burnthick: " << fire_soid.burnthick << "\n";
+  report << "\n";
+  report << "   fire_v2a.orc: " << fire_v2a.orgc << "\n";
+  report << "   fire_v2a.orn: " << fire_v2a.orgn << "\n";
+  report << "\n";
+  report << "   fire_v2soi.abvc: " << fire_v2soi.abvc << "\n";
+  report << "   fire_v2soi.abvn: " << fire_v2soi.abvn << "\n";
+  report << "   fire_v2soi.blwc: " << fire_v2soi.blwc << "\n";
+  report << "   fire_v2soi.blwn: " << fire_v2soi.blwn << "\n";
+  report << "\n";
+  report << "   fire_soi2a.orgc: " << fire_soi2a.orgc << "\n";
+  report << "   fire_soi2a.orgn: " << fire_soi2a.orgn << "\n";
+  report << "\n";
+  report << "   fire_a2soi.orgn: " << fire_a2soi.orgn << "\n";
+
+  return report.str();
+
+}
+
+void ThermokarstData::clear() {
+  useseverity = false;
+  fire_soid = soidiag_fir();
+  fire_v2a  = veg2atm_fir();
+  fire_v2soi= veg2soi_fir();
+  fire_soi2a= soi2atm_fir();
+  fire_v2dead = veg2dead_fir();
+  //fire_a2soi is not cleared until FRI*12 months post fire, as it
+  //holds the flux value for transferring N back to the soil.
+  //fire_a2soi= atm2soi_fir();
+};
+
+void ThermokarstData::init() {
+  //
+};
+
+void ThermokarstData::beginOfMonth(){
+  fire_v2dead = veg2dead_fir();
+}
+
+void ThermokarstData::beginOfYear() {
+  fire_soid.burnthick =0.;
+  fire_v2a.orgc =0.;
+  fire_v2a.orgn =0.;
+  fire_v2soi.abvc =0.;
+  fire_v2soi.blwc =0.;
+  fire_v2soi.abvn =0.;
+  fire_v2soi.blwn =0.;
+  fire_soi2a.orgc =0.;
+  fire_soi2a.orgn =0.;
+};
+
+void ThermokarstData::endOfYear() {
+//
+};
+
+void ThermokarstData::thaw() {
+  clear();
+}
