+++ .github/workflows/main_sample.yml
@@ -29,11 +29,11 @@ jobs:
 
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
-     - name: Make input catalog and workflows dir
-       run: |
-         mkdir -p source
-         mkdir -p input-catalog
-         mkdir -p workflows
+      - name: Make input catalog and workflows dir
+        run: |
+          mkdir -p source
+          mkdir -p input-catalog
+          mkdir -p workflows
 
       - name: Checkout
         uses: actions/checkout@v4
@@ -57,10 +57,10 @@ jobs:
       - name: build second image layer
         run: bash docker-build-wrapper.sh --dev
         continue-on-error: true
-        working-direcory: source
+        working-directory: source
 
       - name: Set v tag
-        run | 
+        run: | 
           GIT_VERSION="$(git describe --tags)"
           echo "V_TAG=${GIT_VERSION}" >> $GITHUB_ENV
         working-directory: source
