+++ .github/workflows/main_sample.yml
@@ -69,7 +69,7 @@ jobs:
       - name: Run docker compose
         uses: hoverkraft-tech/compose-action@v2.0.1
         with:
-          compose-file: "./source/docker-compose.yml"
+          compose-file: "docker-compose.yml"
           services: dvmdostem-dev
           cwd: source
         env:
