+++ .github/workflows/main_sample.yml
@@ -46,47 +46,71 @@ jobs:
       - run: git fetch --prune --unshallow
         working-directory: source
 
-      # Not sure why but running the docekr build wrapper in this context
-      # returns 1 which is considered an error and the subsequent steps
-      # don't run, even though by all indications from the information
-      # # printed to the console, the image build just fine.
-      - name: build first image layer
-        run: bash docker-build-wrapper.sh --cpp-dev
-        continue-on-error: true
-        working-directory: source
-      
-      - name: build second image layer
-        run: bash docker-build-wrapper.sh --dev
-        continue-on-error: true
-        working-directory: source
-
       - name: Set v tag
         run: | 
           GIT_VERSION="$(git describe --tags)"
+          DDT_INPUT_CATALOG="$GITHUB_WORKSPACE"/input-catalog
+          DDT_WORKFLOWS="$GITHUB_WORKSPACE"/workflows
+          SOURCE_DIR="$GITHUB_WORKSPACE"/source
+          
           echo "V_TAG=${GIT_VERSION}" >> $GITHUB_ENV
+          echo "DDT_INPUT_CATALOG=${DDT_INPUT_CATALOG}" >> $GITHUB_ENV
+          echo "DDT_WORKFLOWS=${DDT_WORKFLOWS}" >> $GITHUB_ENV
+          echo "PWD=${SOURCE_DIR}" >> $GITHUB_ENV
+
         working-directory: source
 
-      - name: Run docker compose
-        uses: hoverkraft-tech/compose-action@v2.0.1
-        with:
-          compose-file: "docker-compose.yml"
-          services: dvmdostem-dev
-          cwd: source
-        env:
-          DDT_WORKFLOWS: /home/runner/work/dvm-dos-tem/dvm-dos-tem/workflows
-          DDT_INPUT_CATALOG: /home/runner/work/dvm-dos-tem/dvm-dos-tem/input-catalog
-          PWD: /home/runner/work/dvm-dos-tem/dvm-dos-tem/source
+      - name: look at env
+        run: |
+          echo $GITHUB_ENV
+
+      # # Not sure why but running the docekr build wrapper in this context
+      # # returns 1 which is considered an error and the subsequent steps
+      # # don't run, even though by all indications from the information
+      # # # printed to the console, the image build just fine.
+      # - name: build first image layer
+      #   run: bash docker-build-wrapper.sh --cpp-dev
+      #   continue-on-error: true
+      #   working-directory: source
+      
+      # - name: build second image layer
+      #   run: bash docker-build-wrapper.sh --dev
+      #   continue-on-error: true
+      #   working-directory: source
+
+      # - name: Set v tag
+      #   run: | 
+      #     GIT_VERSION="$(git describe --tags)"
+      #     echo "V_TAG=${GIT_VERSION}" >> $GITHUB_ENV
+      #   working-directory: source
+
+      # - name: Run docker compose
+      #   uses: hoverkraft-tech/compose-action@v2.0.1
+      #   with:
+      #     compose-file: "docker-compose.yml"
+      #     services: dvmdostem-dev
+      #     cwd: source
+      #   env:
+      #     DDT_WORKFLOWS: /home/runner/work/dvm-dos-tem/dvm-dos-tem/workflows
+      #     DDT_INPUT_CATALOG: /home/runner/work/dvm-dos-tem/dvm-dos-tem/input-catalog
+      #     PWD: /home/runner/work/dvm-dos-tem/dvm-dos-tem/source
       
 
-      - name: compile in the running service
-        run: |
-          docker compose exec dvmdostem-dev make
-        working-directory: source
+      # - name: compile in the running service
+      #   run: |
+      #     docker compose exec dvmdostem-dev make
[…]
