### 2025-03-02 0892ec0
Work on github action.
Now this seems to work! Finally a useful CI workflow/test!
