### 2025-03-04 280b3bc
Corrects daily driving variable outputs
These outputs were never updated when we switched
from outputting every in-simulation month to
outputting at the end of the in-simulation year.
