### 2025-03-04 4a97cc6
Fix path to plotting script in quick install/demo instructions.
