### 2025-03-04 e4d0a3e
Remove version attribute from docker compose file
No longer needed with recent versions of docker compose and will result
in the following warning:

  > the attribute `version` is obsolete, it will be ignored, please
  > remove it to avoid potential confusion
