### 2025-03-05 79f53ba
Re-format readme. Include commands in full codeblock form, fix wording in a few spots.
