### 2025-03-05 2690b74
Change quickstart in the UserGuide to match readme.
