### 2025-03-05 29ef822
Development on groundwater logic
Adding parameter and making sure water is being correctly
accumulated. Some more debugging may be needed.
