### 2025-03-05 c0e017a
Add version to title, add clarifying sentence to statement of need.
We haven't released v0.8.2 yet but will ASAP - JOSS paper should be for
the latest release, which will be v0.8.2 shortly. Will merge in
some last remaning changes from JOSS review process among other things.
