### 2025-03-06 68da301
A lot of debugging related changes
