+++ mads_calibration/sa-demo-config.yaml
@@ -8,10 +8,10 @@ seed_path: /work/parameters
 observations: /work/calibration
 
 # The number of samples to draw over the parameter space
-N_samples: 5
+N_samples: 100
 
 # The location where all the individual sample folders will be setup.
-work_dir: /data/workflows/calibration/CMT06-IMNAVIAT/output/cmax_gppallignoringnitrogen/sa-N5
+work_dir: /data/workflows/groundwater_test_SA
 
 # The following keys (params, pftnums, and percent_diffs) should be lists of
 # identical length. I.e. for each parameter we need to specify the PFT and a
@@ -20,22 +20,22 @@ work_dir: /data/workflows/calibration/CMT06-IMNAVIAT/output/cmax_gppallignoringn
 # the parameter.
 
 # Which parameters are being tested/analyzed
-params: [cmax, cmax, cmax, cmax, cmax, cmax, cmax]
+params: [inflowfactor]
 
 # Which PFTs are being tested/analyzed
-pftnums: [0, 1, 2, 3, 4, 5, 6]
+pftnums: [None]
 
 # The "perturbations" around the initial parameter values
-percent_diffs: [.1, .1, .1, .1, .1, .1, .1]
+# percent_diffs: [.1, .1, .1, .1, .1, .1, .1]
 
 # Manually set parameter bounds. This option should be mutually exclusive 
 # with percent_diffs. In other words only specify one of percent_diffs or 
 # p_bounds. For each parameter specify the min and max. Must use square 
 # brackets for yaml to parse it correctly.
-#p_bounds: [[0,1],[0,1],[0,1],[0,1],[0,1],[0,1],[0,1]]
+p_bounds: [[0.0, 30.0]]
 
 # The path to the driving input data that will be used.
-site: /data/input-catalog/cru-ts40_ar5_rcp85_ncar-ccsm4_IMNAVIAT_CREEK_10x10
+site: /data/input-catalog/bnz_lter_vpd
 
 # The pixel from the driving inputs that should be used.
 PXx: 0
@@ -46,7 +46,7 @@ PXy: 0
 #  - False/None/0: dvmdostem default which is dsl and nfeed on for eq, sp, tr, sc stages
 #  - GPPALLIgnoringNitrogen: turns dsl and nfeed OFF for eq stage
 #  - VEGC or NPPAll: turns dsl and nfeed ON for eq (and subsequent) stages
-calib_mode: GPPAllIgnoringNitrogen
+calib_mode: False
 
 # What mathematical scheme to use to choose samples based on bounds, 
 # and perturbations...
@@ -57,21 +57,30 @@ sampling_method: lhc
 # calibration_targets.py file. Internally the program looks up the correct 
 # NetCDF output to enable.
 target_names: 
-- GPPAllIgnoringNitrogen
+- CarbonShallow
+- CarbonDeep
+- CarbonMineralSum
 
 # If you want additional outputs beyond the ones needed for the target
 # comparison then you can specify them here. In this case use the NetCDF names.
 # WARNING you should make sure that you don't explicity change the resolution
 # for a variable that is one of your target variables. This will make the
 # plotting in SA_post_hoc_analysis not work.
-# aux_outputs:
-# - VEGC p m
-# - AVLN y
-# - GPP y p c 
-# - TLAYER m l
+aux_outputs:
+- TLAYER m
+- LWCLAYER m
+- VWCLAYER m
+- WATERTAB m
+- ALD y
+- LAYERDEPTH m
+- QRUNOFF m
+- QINFILTRATION m
+- EET m
+- TRANSPIRATION m
+- QDRAINAGE m
 
[…]
