### 2025-03-06 70b4706
update image in paper
