### 2025-03-06 3bad56b
remove extraneous html markup from references.
