+++ joss-paper/joss-paper.bib
@@ -1,5 +1,5 @@
 @article{Briones2024,
-   abstract = {<p> Accelerated warming of the Arctic can affect the global climate system by thawing permafrost and exposing organic carbon in soils to decompose and release greenhouse gases into the atmosphere. We used a process-based biosphere model (DVM-DOS-TEM) designed to simulate biophysical and biogeochemical interactions between the soil, vegetation, and atmosphere. We varied soil and environmental parameters to assess the impact on cryohydrological and biogeochemical outputs in the model. We analyzed the responses of ecosystem carbon balances to permafrost thaw by running site-level simulations at two long-term tundra ecological monitoring sites in Alaska: Eight Mile Lake (EML) and Imnavait Creek Watershed (IMN), which are characterized by similar tussock tundra vegetation but differing soil drainage conditions and climate. Model outputs showed agreement with field observations at both sites for soil physical properties and ecosystem CO <sub>2</sub> fluxes. Model simulations of Net Ecosystem Exchange (NEE) showed an overestimation during the frozen season (higher CO <sub>2</sub> emissions) at EML with a mean NEE of 26.98 ± 4.83 gC/m <sup>2</sup> /month compared to observational mean of 22.01 ± 5.67 gC/m <sup>2</sup> /month, and during the fall months at IMN, with a modeled mean of 19.21 ± 7.49 gC/m <sup>2</sup> /month compared to observation mean of 11.9 ± 4.45 gC/m <sup>2</sup> /month. Our results underscore the importance of representing the impact of soil drainage conditions on the thawing of permafrost soils, particularly poorly drained soils, which will drive the magnitude of carbon released at sites across the high-latitude tundra. These findings can help improve predictions of net carbon releases from thawing permafrost, ultimately contributing to a better understanding of the impact of Arctic warming on the global climate system. </p>},
+   abstract = { Accelerated warming of the Arctic can affect the global climate system by thawing permafrost and exposing organic carbon in soils to decompose and release greenhouse gases into the atmosphere. We used a process-based biosphere model (DVM-DOS-TEM) designed to simulate biophysical and biogeochemical interactions between the soil, vegetation, and atmosphere. We varied soil and environmental parameters to assess the impact on cryohydrological and biogeochemical outputs in the model. We analyzed the responses of ecosystem carbon balances to permafrost thaw by running site-level simulations at two long-term tundra ecological monitoring sites in Alaska: Eight Mile Lake (EML) and Imnavait Creek Watershed (IMN), which are characterized by similar tussock tundra vegetation but differing soil drainage conditions and climate. Model outputs showed agreement with field observations at both sites for soil physical properties and ecosystem CO <sub>2</sub> fluxes. Model simulations of Net Ecosystem Exchange (NEE) showed an overestimation during the frozen season (higher CO <sub>2</sub> emissions) at EML with a mean NEE of 26.98 ± 4.83 gC/m <sup>2</sup> /month compared to observational mean of 22.01 ± 5.67 gC/m <sup>2</sup> /month, and during the fall months at IMN, with a modeled mean of 19.21 ± 7.49 gC/m <sup>2</sup> /month compared to observation mean of 11.9 ± 4.45 gC/m <sup>2</sup> /month. Our results underscore the importance of representing the impact of soil drainage conditions on the thawing of permafrost soils, particularly poorly drained soils, which will drive the magnitude of carbon released at sites across the high-latitude tundra. These findings can help improve predictions of net carbon releases from thawing permafrost, ultimately contributing to a better understanding of the impact of Arctic warming on the global climate system. },
[…]
