### 2025-03-06 676c3bd
add 2 missing references
