+++ joss-paper/joss-paper.bib
@@ -74,6 +74,17 @@
    volume = {11},
    year = {2014},
 }
+@article{Jafarov2013,
+   abstract = {Fire is an important factor controlling the composition and thickness of the organic layer in the black spruce forest ecosystems of interior Alaska. Fire that burns the organic layer can trigger dramatic changes in the underlying permafrost, leading to accelerated ground thawing within a relatively short time. In this study, we addressed the following questions. (1) Which factors determine post-fire ground temperature dynamics in lowland and upland black spruce forests? (2) What levels of burn severity will cause irreversible permafrost degradation in these ecosystems? We evaluated these questions in a transient modeling-sensitivity analysis framework to assess the sensitivity of permafrost to climate, burn severity, soil organic layer thickness, and soil moisture content in lowland (with thick organic layers, ∼80 cm) and upland (with thin organic layers, ∼30 cm) black spruce ecosystems. The results indicate that climate warming accompanied by fire disturbance could significantly accelerate permafrost degradation. In upland black spruce forest, permafrost could completely degrade in an 18 m soil column within 120 years of a severe fire in an unchanging climate. In contrast, in a lowland black spruce forest, permafrost is more resilient to disturbance and can persist under a combination of moderate burn severity and climate warming. © 2013 IOP Publishing Ltd.},
+   author = {E. E. Jafarov and V. E. Romanovsky and H. Genet and A. D. McGuire and S. S. Marchenko},
+   doi = {10.1088/1748-9326/8/3/035030},
+   issn = {17489326},
+   issue = {3},
+   journal = {Environmental Research Letters},
+   title = {The effects of fire on the thermal stability of permafrost in lowland and upland black spruce forests of interior Alaska in a changing climate},
+   volume = {8},
+   year = {2013},
+}
 @article{Koven2011,
    abstract = {Permafrost soils contain enormous amounts of organic carbon, which could act as a positive feedback to global climate change due to enhanced respiration rates with warming. We have used a terrestrial ecosystem model that includes permafrost carbon dynamics, inhibition of respiration in frozen soil layers, vertical mixing of soil carbon from surface to permafrost layers, and CH 4 emissions from flooded areas, and which better matches new circumpolar inventories of soil carbon stocks, to explore the potential for carbon-climate feedbacks at high latitudes. Contrary to model results for the Intergovernmental Panel on Climate Change Fourth Assessment Report (IPCC AR4), when permafrost processes are included, terrestrial ecosystems north of 60° N could shift from being a sink to a source of CO 2 by the end of the 21st century when forced by a Special Report on Emissions Scenarios (SRES) A2 climate change scenario. Between 1860 and 2100, the model response to combined CO 2 fertilization and climate change changes from a sink of 68 Pg to a 27 + -7 Pg sink to 4 + -18 Pg source, depending on the processes and parameter values used. The integrated change in carbon due to climate change shifts from near zero, which is within the range of previous model estimates, to a climate-induced loss of carbon by ecosystems in the range of 25 + -3 to 85+ -16 Pg C, depending on processes included in the model, with a best estimate of a 62 + -7 Pg C loss. Methane emissions from high-latitude regions are calculated to increase from 34 Tg CH 4/y to 41-70 TgCH 4/y, with increases due to CO 2 fertilization, permafrost thaw, and warming-induced increased CH 4 flux densities partially offset by a reduction in wetland extent.},
    author = {Charles D. Koven and Bruno Ringeval and Pierre Friedlingstein and Philippe Ciais and Patricia Cadule and Dmitry Khvorostyanov and Gerhard Krinner and Charles Tarnocai},
@@ -95,6 +106,17 @@
    title = {Interactions between carbon and nitrogen dynamics in estimating net primary productivity for potential vegetation in North America},
    volume = {6},
    year = {1992},
+}
[…]
