### 2025-03-06 uaf-arctic-eco-modeling/dvm-dos-tem#770 @tobeycarman
@hgenet I think I have addressed all the feedback here except for the "state of the field".  I could use a sentence or two from you describing this. I think we could fit this in either at the end of the "Model Design" section, or if you come up with more than a sentence or two, we could add a full section "State of the field".

You can edit the `paper.md` directly on the `joss-paper-2` branch and push your changes or send edits to me and I will make the commits and push them.

Also note that I added the version to the title, but I used v0.8.2 as I would like to make a release right before this is published.
