### 2025-03-06 #7667 @tobeycarman
@mengqi-z @kdorheim @arbennett,  I think I have fixed everything mentioned by @kdorheim except adding a mention of other similar models. I think this is a good idea and I have asked @hgenet for help with this. I pushed changes to our `joss-paper-2` branch, and also made several updates to our `master` branch because some of the feedback involved changing our README and documentation. The most recent draft was compiled by the GitHub action in our repo and is available here:
https://github.com/uaf-arctic-eco-modeling/dvm-dos-tem/actions/runs/13707476466/artifacts/2706509209
