### 2025-03-06 d5bd855
updating lai and veg coverage, more checks needed
We noticed that even when we update carbon pools
processes are delineated using variables in different
structures which do not get updated, so additional
checks and tests are needed. This may fix part of
the issue, but we will need to debug and probably
add further checks.
