+++ src/Cohort.cpp
@@ -763,7 +763,7 @@ void Cohort::updateMonthly_Dsb(const int & yrind, const int & currmind, std::str
   if(md->cmt_change
      && fire.getSeverity(yrind) >= 4 // we may this to be based on remaining OLT
      && cd.mthsdist == 0){
-    veg.cmtChange();
+    veg.cmtChange(currmind);
   }
 }
 
