+++ include/Vegetation.h
@@ -45,7 +45,7 @@ public:
   void setEnvData(const int &ip, EnvData * edp);
   void setBgcData(const int &ip, BgcData * bdp);
 
-  void cmtChange();
+  void cmtChange(const int & currmind);
 
 private:
   CohortLookup * chtlu;
