### 2025-03-07 64f840b
In development, debugging tests active
Logic seems good, exploring feedbacks between
groundwater, soil and vegetation
