### 2025-03-10 uaf-arctic-eco-modeling/dvm-dos-tem#770 @hgenet
My apologies for the slow response. We should indeed harmonize the qualification of the type of model DVM-DOS-TEM is, in the paper, the readme file and the documentation. Furthermore, I agree that it should be made clear that DVM-DOS-TEM is not a new model, but rather carries the long legacy of previous versions of the terrestrial ecosystem model (TEM).
