### 2025-03-10 uaf-arctic-eco-modeling/dvm-dos-tem#770 @hgenet
Tobey, please review the following edits - if you agree with them, I'll let you integrate them - THANK YOU!
1- In the manuscript:
1a- 25/26: The model couples two previous versions of the Terrestrial Ecosystem Model (TEM, McGuire et al. 1991): DVMTEM that includes a dynamic vegetation module (DVM, Euskirchen et al. 2009), and DOSTEM that includes a dynamic organic soil module (DOS, Yi et al. 2010, Genet et al. 2013). DVMDOSTEM simulates processes at a yearly and monthly scales, ....
1b- 43/46: The DVMDOSTEM ecosystem model is the latest version of TEM that incorporates critical vegetation and soil structural aspects that improve the representation of two critical drivers of change on high latitude ecosystems: climate change and natural disturbances such as wildfire. This model has been developed with particular attention to physical and biological processes driving permafrost and carbon cycling in high latitude ecosystems. DVMDOSTEM is therefore well suited to assessing and informing our understanding of the PCCF.
2- The readme file - first sentence:
The dvm-dos-tem (DVMDOSTEM) model is a process based terrestrial ecosystem model that focuses on representing ecosystem carbon and nitrogen cycles as well as physical processes (e.g. soil temperature and moisture) and vegetation dynamics in high latitude ecosystems.
3- Documentation - Welcome section: DVMDOSTEM is a process-based terrestrial ecosystem model, designed to simulate biophysical and biogeochemical processes between the soil, the vegetation and the atmosphere.

Also, for some reason, the references in the manuscript are sometimes including author's first name. This would need to be fixed. References should likely only refer to author's last names.
