### 2025-03-10 uaf-arctic-eco-modeling/dvm-dos-tem#770 @hgenet
And Tobey - yes on the new release - good idea.
