+++ src/Thermokarst.cpp
@@ -1,0 +1,747 @@
+/*
+ * Thermokarst.cpp
+ *
+ * 3/12/2025
+ *
+ * Starting with a copy of Wildfire.cpp and adapting
+ * script for thermokarst disturbance. Initially, 
+ * this will focus on retrogressive thaw slump and 
+ * possibly active layer detachment as a precursor.
+ *
+ * Thermokarst disturbance initiation will first be
+ * prescribed using a similar method as the explicit
+ * historic fire input forcing. Later we hope to use
+ * either and external or integrated predictive model
+ * to determine thermokarst type, occurrence, and 
+ * severity (or extent).
+ *
+ * C/N pools can be updated here using 'bd' but may 
+ * require additional modification in other places.
+ * Changes to C/N soil and vegetation pools will vary
+ * depending on thermokarst type.
+ *
+ * Soil layer structure needs to be updated in ground
+ * though the function for this can be defined here.
+ *
+ * Roots will need to be considered, but not entirely 
+ * sure of the best method at this point.
+ */
+
+#include <string>
+#include <sstream>
+
+#include "../include/WildFire.h" // >>> this will need to be Thermokarst.h
+
+#include "../include/TEMUtilityFunctions.h"
+#include "../include/TEMLogger.h"
+
+extern src::severity_logger< severity_level > glg;
+
+Thermokarst::Thermokarst() {}
+
+Thermokarst::~Thermokarst() {}
+
+Thermokarst::Thermokarst(const std::string &fri_fname, // >>> fri not required, exp will be a thermokarst specific input which needs updating
+                         const std::string &exp_fname, const double cell_slope,
+                         const double cell_aspect, const double cell_elevation,
+                         const int y, const int x){
+
+  // >>> We will not require a FRI input
+
+// #pragma omp critical(load_input)
+//   {
+//     BOOST_LOG_SEV(glg, info) << "Setting up FRI fire data...";
+//     this->fri = temutil::get_scalar<int>(fri_fname, "fri", y, x);
+//     this->fri_severity = temutil::get_scalar<int>(fri_fname, "fri_severity", y, x);
+//     this->fri_jday_of_burn = temutil::get_scalar<int>(fri_fname, "fri_jday_of_burn", y, x);
+//     this->fri_area_of_burn = temutil::get_scalar<int>(fri_fname, "fri_area_of_burn", y, x);
+//   }//End critical(fri)
+
+  // >>> We can keep the below input loading but will need to change the name once
+  // an input has been established.
+
+  #pragma omp critical(load_input)
+  {
+    BOOST_LOG_SEV(glg, info) << "Setting up explicit fire data...";
+    this->exp_burn_mask = temutil::get_timeseries<int>(exp_fname, "exp_burn_mask", y, x);
+    this->exp_fire_severity = temutil::get_timeseries<int>(exp_fname, "exp_fire_severity", y, x);
+    this->exp_jday_of_burn = temutil::get_timeseries<int>(exp_fname, "exp_jday_of_burn", y, x);
+    this->exp_area_of_burn = temutil::get_timeseries<int64_t>(exp_fname, "exp_area_of_burn", y, x);
+  }//End critical(exp_fir) 
+
+  this->slope = cell_slope;
+  this->asp = cell_aspect;
+  this->elev = cell_elevation;
+
[…]
