### 2025-03-13 9a3cbaf
Added some more variable updating after cmt change
Noticed some other needs for updating in environment
data - specifically, variables affecting EET and PET
calculations resulting in NaNs. Will update next.
