+++ src/Cohort.cpp
@@ -761,7 +761,7 @@ void Cohort::updateMonthly_Dsb(const int & yrind, const int & currmind, std::str
   //updateMonthly_Flood(...)
 
   if(md->cmt_change
-     && fire.getSeverity(yrind) >= 4 // we may this to be based on remaining OLT
+     && fire.getSeverity(yrind) >= 1 // we may this to be based on remaining OLT
      && cd.mthsdist == 0){
     veg.cmtChange(currmind);
   }
