### 2025-03-18 8ac9233
add state of field section comparing to other models.
