+++ joss-paper/paper.md
@@ -222,7 +222,28 @@ effects of wildfire in order to evaluate the role of climate-driven fire
 intensification on ecosystem structure and function [@Yi2010; @Genet2013]. The
 structure of `DVMDOSTEM` is represented visually in \autoref{fig:modeloverview}.
 
+# State of the Field
 
+In the field of ecosystem models, several prominent models such as CLM5
+[@Lawrence2019], ELM-FATES [@Fisher2015], LANDIS [@Scheller2007], and iLand
+[@Seidl2012] have been developed to simulate ecological processes at various
+scales, resolutions and ecotypes. Like `DVMDOSTEM`, ELM-FATES, and early
+versions of CLM, are "offline" land models that do not include feedback with
+atmospheric or oceanic models, focusing instead on land surface processes.
+LANDIS and iLand emphasize forest dynamics with some interaction between grid
+cells, while `DVMDOSTEM` does not model interaction between grid cells. While
+`DVMDOSTEM` contains detailed representations of vegetation - multiple Plant
+Functional Types and individual compartments within Plant Functional Types
+(PFTs) - it also has the concept of community types (collections of PFTs and
+soil properties) and is designed to run at the landscape scale by representing
+more than a single stand of trees or a single forest type. `DVMDOSTEM` is unique
+in its detailed representation of high latitude processes, particularly the
+dynamic organic soils in regions with frozen ground coupled with dynamic
+vegetation and high latitude specific parameterizations. This focus allows
+`DVMDOSTEM` to simulate the complex interactions between soil, vegetation, and
+climate in permafrost ecosystems, providing valuable insights into the PCCF. For
+a more detailed assessment and comparison of high latitude vegetation modeling,
+see [@Heffernan2024].
 
 # Acknowledgements
 
