+++ joss-paper/joss-paper.bib
@@ -36,6 +36,15 @@
    volume = {19},
    year = {2009},
 }
+@article{Fisher2015,
+  title={Taking off the training wheels: the properties of a dynamic vegetation model without climate envelopes},
+  author={Fisher, RA and Muszala, S and Verteinstein, M and Lawrence, P and Xu, C and McDowell, NG and Knox, RG and Koven, C and Holm, J and Rogers, BM and others},
+  journal={Geoscientific Model Development Discussions (Online)},
+  volume={8},
+  number={4},
+  year={2015},
+  publisher={Los Alamos National Laboratory (LANL), Los Alamos, NM (United States)}
+}
 @article{Genet2018,
    abstract = { It is important to understand how upland ecosystems of Alaska, which are estimated to occupy 84% of the state (i.e., 1,237,774 km <sup>2</sup> ), are influencing and will influence state‐wide carbon (C) dynamics in the face of ongoing climate change. We coupled fire disturbance and biogeochemical models to assess the relative effects of changing atmospheric carbon dioxide (CO <sub>2</sub> ), climate, logging and fire regimes on the historical and future C balance of upland ecosystems for the four main Landscape Conservation Cooperatives (LCCs) of Alaska. At the end of the historical period (1950–2009) of our analysis, we estimate that upland ecosystems of Alaska store ~50 Pg C (with ~90% of the C in soils), and gained 3.26 Tg C/yr. Three of the LCCs had gains in total ecosystem C storage, while the Northwest Boreal LCC lost C (−6.01 Tg C/yr) because of increases in fire activity. Carbon exports from logging affected only the North Pacific LCC and represented less than 1% of the state's net primary production (NPP). The analysis for the future time period (2010–2099) consisted of six simulations driven by climate outputs from two climate models for three emission scenarios. Across the climate scenarios, total ecosystem C storage increased between 19.5 and 66.3 Tg C/yr, which represents 3.4% to 11.7% increase in Alaska upland's storage. We conducted additional simulations to attribute these responses to environmental changes. This analysis showed that atmospheric CO <sub>2</sub> fertilization was the main driver of ecosystem C balance. By comparing future simulations with constant and with increasing atmospheric CO <sub>2</sub> , we estimated that the sensitivity of NPP was 4.8% per 100 ppmv, but NPP becomes less sensitive to CO <sub>2</sub> increase throughout the 21st century. Overall, our analyses suggest that the decreasing CO <sub>2</sub> sensitivity of NPP and the increasing sensitivity of heterotrophic respiration to air temperature, in addition to the increase in C loss from wildfires weakens the C sink from upland ecosystems of Alaska and will ultimately lead to a source of CO <sub>2</sub> to the atmosphere beyond 2100. Therefore, we conclude that the increasing regional C sink we estimate for the 21st century will most likely be transitional. },
    author = {Hélène Genet and Yujie He and Zhou Lyu and A. David McGuire and Qianlai Zhuang and Joy Clein and David D'Amore and Alec Bennett and Amy Breen and Frances Biles and Eugénie S. Euskirchen and Kristofer Johnson and Tom Kurkowski and Svetlana (Kushch) Schroder and Neal Pastick and T. Scott Rupp and Bruce Wylie and Yujin Zhang and Xiaoping Zhou and Zhiliang Zhu},
@@ -60,6 +69,19 @@
    title = {Modeling the effects of fire severity and climate warming on active layer thickness and soil carbon storage of black spruce forests across the landscape in interior Alaska},
    volume = {8},
    year = {2013},
+}
+@article{Heffernan2024,
[…]
