### 2025-03-18 546c360
adjust readme and documentation intro text as per comment in iss #770
