### 2025-03-18 uaf-arctic-eco-modeling/dvm-dos-tem#770 @tobeycarman
Addressed with recent commits to `joss-paper-2` branch.
