### 2025-03-18 #7667 @tobeycarman
@kdorheim, @mengqi-z, I have addressed all the feedback thus far. I closed the related issues in our repo. Note that there were several small changes to the README and the documentation that exist only on the `joss-paper-2` branch, and won't be reflected in the published documentation or the landing page of the repository until we merge the `joss-paper-2` branch.

Thanks for your time working on this!
