### 2025-03-18 uaf-arctic-eco-modeling/dvm-dos-tem#776 @arbennett
JOSS Review: Instructions for podman
I had a bit of trouble getting dvm-dos-tem running on my linux machine, which turns out was due to some differences in setting up podman (docker equivalent) to run with non-root access. Eventually I was able to figure this out, and because it requires different steps than docker it might be good to add a link to the docs on this.

I found this post to solve my issue: https://unix.stackexchange.com/questions/731645/podman-w-docker-compose-run-as-user

I am not sure if there are other ways to accomplish this, as I'm not super familiar with podman, but these are the commands that I was able to run to allow me to use `docker compose` as non-root via podman:

```
sudo systemctl disable --now podman podman.socket
systemctl enable --now --user podman podman.socket

# Take output of this command into next command
podman info --format '{{.Host.RemoteSocket.Path}}'

# Example: export DOCKER_HOST=unix:///run/user/1001/podman/podman.sock
export DOCKER_HOST=unix:/{OUTPUT FROM PODMAN INFO ABOVE}

docker compose up -d dvmdostem-dev
```
