### 2025-03-19 #7667 @arbennett
@mengqi-z @tobeycarman @hgenet sorry for falling off the map on this!

I was able to complete my review, and overall I think this is looking good with the changes on the `joss-paper-2` branch. I was able to download and run the model, as well as play around with the configuration thanks to the extensive documentation and detailed instructions. I was able to run on both my Macbook as well as a Linux server. I had a bit of trouble getting things running on the Linux server because it uses podman to emulate docker, so needed to find a workaround for non-root access. I left an issue [here](https://github.com/uaf-arctic-eco-modeling/dvm-dos-tem/issues/776), that I recommend be added to the quick start information. I don't necessarily expect that the solution I found is the best way, but I think it would be good to especially highlight the potential difference between podman and docker so that users at least have a jumping off point to troubleshoot.
