### 2025-03-24 6bfc58c
fixing bug if ch4 file is not read in climate averaging
