### 2025-03-25 d014420
Splits veg N outputs to STR, LAB, TOT
The existing VEGN output was an erroneous mix
of total and structural nitrogen. This commit
splits the outputs to be specifically total,
structural, or labile. Total N is not available
by pft or compartment due to the limitations of
labile N.

Any calibration workflow will need to be modified
to use VEGNSTR instead of VEGN.
