### 2025-03-26 58f7779
Using calibrated ebul param for efflux calculation
