### 2025-03-27 2debc2e
Altering thermokarst / wildfire naming
