### 2025-03-27 dc3c694
Altering thermokarst / wildfire naming
