### 2025-03-27 5a520de
Moving cmtchange to cohort
Moved cmtchange function from vegetation to cohort
as this makes access to data structures simpler.
