+++ src/Vegetation.cpp
@@ -543,167 +543,6 @@ void Vegetation::updateFrootfrac() {
   }
 };
 
-/**
- * Changing CMT vegetation following disturbance, reloading 
- * vegetation parameters, and redistributing carbon and 
- * nitrogen pools based on new init and C:N values
- */
-void Vegetation::cmtChange(const int & currmind){
-  // Determine cmt to succeed to (for testing we are
-  // using cmt1 -> cmt3)
-  std::string new_cmt = "CMT3";
-
-  // Load relevant parameters from new CMT
-  chtlu->cmtcode = new_cmt;
-
-  // reinitializing vegetation parameters
-  chtlu->loadVegetationParams();
-  initializeParameter();
-
-  // // pft needs to be initialized individually for 'envmodule' and 'bgcmodule'
-  // for (int ip = 0; ip < NUM_PFT; ip++)
-  // {
-  //   vegenv[ip].ipft = ip;
-  //   vegbgc[ip].ipft = ip;
-  //   // set-up paramters for vegetation processes
-
-  //   vegbgc[ip].initializeParameter();
-
-  // }
-
-  // reassign pools to new PFTs
-  double cpool = 0.0;
-  double npool = 0.0;
-  double labnpool = 0.0;
-
-  // Looping over previous set of PFTs to accumulate remaining
-  // carbon and nitrogen pools
-  for(int ip=0; ip<NUM_PFT; ip++){
-
-    if(cd->m_veg.vegcov[ip]>0.){//only check PFTs that exist
-
-      labnpool += bd[ip]->m_vegs.labn;
-
-      for(int ipp=0; ipp<NUM_PFT_PART; ipp++){
-
-        cpool += bd[ip]->m_vegs.c[ipp];
-        npool += bd[ip]->m_vegs.strn[ipp];
-
-      }
-    }
-  }
-
-  double initvegc_tot = 0.0;
-  
-  // total expected nitrogen for new cmt based on
-  // cn ratio parameters
-  double initvegn_tot = 0.0;
-
-  double initvegc_pft[NUM_PFT] = {0.0};
-  double initvegn_part_pft[NUM_PFT_PART][NUM_PFT] = {0.0};
-
-  for(int ip=0; ip<NUM_PFT; ip++){
-
-    for(int ipp=0; ipp<NUM_PFT_PART; ipp++){
-
-      // accumulating pft specific carbon values
-      initvegc_pft[ip] += chtlu->initvegc[ipp][ip];
-
-      // only calculate nitrogen if c:n ratio is non-zero
-      // otherwise NaNs created from divide by zero due to
-      // PFTs without all compartments
-      if (chtlu->initc2neven[ipp][ip]>0){
-        initvegn_part_pft[ipp][ip] = chtlu->initvegc[ipp][ip] / chtlu->initc2neven[ipp][ip];
-        initvegn_tot += initvegn_part_pft[ipp][ip];
-      }
-    }
-
-    // break out of loop once all new pfts have been summed
-    if (initvegc_pft[ip] == 0.0){
-      break;
-    }
-
-    initvegc_tot += initvegc_pft[ip];
-  }
-
-  for(int ip=0; ip<NUM_PFT; ip++){
-    // break out of loop once all new pfts have been
-    // accounted for
-    if (initvegc_pft[ip] == 0.0){
-      // bd[ip]->m_vegs.c[ipp] = 0.0;
-      // and other places where needed to be updated
-      // continue
-      break;
-    }
-
-    double new_pft_pool = (initvegc_pft[ip] / initvegc_tot) * cpool;
-    bd[ip]->m_vegs.call = new_pft_pool;
[…]
