+++ src/CohortLookup.cpp
@@ -149,8 +149,11 @@ std::string CohortLookup::calparbgc2str() {
   s << this->kdcsoma << " // kdcsoma:\n";
   s << this->kdcsompr << " // kdcsompr:\n";
   s << this->kdcsomcr << "// kdcsomcr:\n";
-  s << this->s2dfraction << "// s2dfraction \n:";
-  s << this->d2mfraction << "// d2mfraction \n:";
+  s << this->s2dfraction << "// s2dfraction: \n:";
+  s << this->d2mfraction << "// d2mfraction: \n:";
+  s << this->max_ponding_s << "// summermaxponding: \n:";
+  s << this->max_ponding_w << "// wintermaxponding: \n:";
+  s << this->max_ponding_w << "// inflowfactor: \n:";
   return s.str();
 }
 
@@ -160,7 +163,7 @@ void CohortLookup::assignBgcCalpar(std::string & dircmt) {
 
   // get a list of data for the cmt number
   std::list<std::string> l = temutil::parse_parameter_file(
-      dircmt + "cmt_calparbgc.txt", temutil::cmtcode2num(this->cmtcode), 20
+      dircmt + "cmt_calparbgc.txt", temutil::cmtcode2num(this->cmtcode), 23
   );
 
   // pop each line off the front of the list
@@ -186,6 +189,9 @@ void CohortLookup::assignBgcCalpar(std::string & dircmt) {
   temutil::pfll2data(l, kdcsomcr);
   temutil::pfll2data(l, s2dfraction);
   temutil::pfll2data(l, d2mfraction);
+  temutil::pfll2data(l, max_ponding_s);
+  temutil::pfll2data(l, max_ponding_w);
+  temutil::pfll2data(l, inflow_factor);
 }
 
 /** Assign "veg dimension?" from parameter file. */
