### 2025-03-28 uaf-arctic-eco-modeling/dvm-dos-tem#776 @tobeycarman
Closed by #777
