### 2025-03-28 e8f1cf5
Switches daily eet/pet to use dayidx instead of dinm
