### 2025-03-28 342244b
Initial output granularity cross-check script from Hélène
