+++ scripts/tests/m2y_analysis_hg.py
@@ -1,0 +1,170 @@
+#!/usr/bin/env python
+
+#### Run command
+### 1. Prepare the output_spec.csv files for four scenarios (yearly/ecosystem, monthly/ecosystem, monthly/pft/layer, monthly/pft/part)
+### 2. cd /data/workflows/TILE_testing/
+### 3. cp ./run-mask.nc ./run-mask_original.nc
+### 4. ncap2 -O -h -s'run(:,:) = 0' ./run-mask.nc ./run-mask.nc
+### 5. ncap2 -O -h -s'run(0,0) = 1' ./run-mask.nc ./run-mask.nc
+### 6. cp output_spec_scenario.csv output_spec.csv
+### 7. /work/dvmdostem -l err -f ./config/config.js -p 100 -e 200 -s 10 -t 3
+### 8. mv -r ./output ./output_scenario
+### 9. mkdir ./output
+
+#### Analysis of the results
+
+import os
+import xarray as xr
+import pandas as pd
+
+tolerance = 0.000001
+outdir = 'temp_outputs/reco_output_debugging/3_outputs_for_analysis/'
+sc_list = ('y', 'm', 'm_pft_layer', 'm_pft_part')
+
+
+
+
+
+
+### Month to year test
+
+
+varlist = ('AVLN','DEEPC','EET','GPP','HKDEEP','LAI','LTRFALC','LTRFALN','MINEC','NETNMIN','NIMMOB','NPP','NRESORB','NUPTAKELAB','NUPTAKEST','ORGN','PET','QDRAINAGE','QINFILTRATION','QRUNOFF','RAINFALL','RG','RHSOM','RM','SHLWC','SNOWFALL','SNOWTHICK','SOMA','SOMCR','SOMPR','SOMRAWC','SWE','TCDEEP','TDEEP','TRANSPIRATION','TSHLW','VEGC','VEGN','VWCDEEP','WATERTAB')
+statlist = ('stock','stock','flux','flux','mean','mean','flux','flux','stock','flux','flux','flux','flux','flux','flux','stock','flux','flux','flux','flux','flux','flux','flux','flux','stock','flux','stock','stock','stock','stock','stock','mean','mean','mean','flux','mean','stock','stock','mean','mean')
+
+sc = 'y'
+yearly = pd.DataFrame(columns = ['year'])
+for i, var in enumerate(varlist):
+	print(var)
+	dd = xr.open_dataset(os.path.join(outdir,'output_'+str(sc), var + '_yearly_tr.nc')).isel(x=0, y=0)
+	dd['time'] = dd.indexes['time'].to_datetimeindex()
+	dd = dd.to_dataframe()
+	dd.reset_index(inplace=True)
+	dd['year'] = dd['time'].dt.year
+	dd = dd.drop(columns=['time'])
+	yearly = pd.merge(yearly, dd, on=['year'], how='outer')
+
+
+sc = 'm'
+monthly = pd.DataFrame(columns = ['year','month'])
+m2y = pd.DataFrame(columns = ['year'])
+for i, var in enumerate(varlist):
+	dd = xr.open_dataset(os.path.join(outdir,'output_'+str(sc), var + '_monthly_tr.nc')).isel(x=0, y=0)
+	dd['time'] = dd.indexes['time'].to_datetimeindex()
+	dd = dd.to_dataframe()
+	dd.reset_index(inplace=True)
+	dd['year'] = dd['time'].dt.year
+	dd['month'] = dd['time'].dt.month
+	dd = dd.drop(columns=['time'])
+	monthly = pd.merge(monthly, dd, on=['year','month'], how='outer')
+	if statlist[i] == 'flux':
+		dy = pd.DataFrame(dd.groupby('year')[var].sum())
[…]
