+++ scripts/tests/m2y_analysis_hg.py
@@ -35,40 +35,40 @@ statlist = ('stock','stock','flux','flux','mean','mean','flux','flux','stock','f
 sc = 'y'
 yearly = pd.DataFrame(columns = ['year'])
 for i, var in enumerate(varlist):
-	print(var)
-	dd = xr.open_dataset(os.path.join(outdir,'output_'+str(sc), var + '_yearly_tr.nc')).isel(x=0, y=0)
-	dd['time'] = dd.indexes['time'].to_datetimeindex()
-	dd = dd.to_dataframe()
-	dd.reset_index(inplace=True)
-	dd['year'] = dd['time'].dt.year
-	dd = dd.drop(columns=['time'])
-	yearly = pd.merge(yearly, dd, on=['year'], how='outer')
+  print(var)
+  dd = xr.open_dataset(os.path.join(outdir,'output_'+str(sc), var + '_yearly_tr.nc')).isel(x=0, y=0)
+  dd['time'] = dd.indexes['time'].to_datetimeindex()
+  dd = dd.to_dataframe()
+  dd.reset_index(inplace=True)
+  dd['year'] = dd['time'].dt.year
+  dd = dd.drop(columns=['time'])
+  yearly = pd.merge(yearly, dd, on=['year'], how='outer')
 
 
 sc = 'm'
 monthly = pd.DataFrame(columns = ['year','month'])
 m2y = pd.DataFrame(columns = ['year'])
 for i, var in enumerate(varlist):
-	dd = xr.open_dataset(os.path.join(outdir,'output_'+str(sc), var + '_monthly_tr.nc')).isel(x=0, y=0)
-	dd['time'] = dd.indexes['time'].to_datetimeindex()
-	dd = dd.to_dataframe()
-	dd.reset_index(inplace=True)
-	dd['year'] = dd['time'].dt.year
-	dd['month'] = dd['time'].dt.month
-	dd = dd.drop(columns=['time'])
-	monthly = pd.merge(monthly, dd, on=['year','month'], how='outer')
-	if statlist[i] == 'flux':
-		dy = pd.DataFrame(dd.groupby('year')[var].sum())
-	if statlist[i] == 'mean':
-		dy = pd.DataFrame(dd.groupby('year')[var].mean())
-	if statlist[i] == 'stock':
-		dy = dd[dd['month'] == 12]
-		dy = dy.drop(columns=['month'])
-	dy.reset_index(inplace=True)
-	if 'index' in dy.columns:
-		dy = dy.drop(columns=['index'])
-#	dy = dy.rename(columns={var: var + '_m2y'}) 
-	m2y = pd.merge(m2y, dy, on=['year'], how='outer')
+  dd = xr.open_dataset(os.path.join(outdir,'output_'+str(sc), var + '_monthly_tr.nc')).isel(x=0, y=0)
+  dd['time'] = dd.indexes['time'].to_datetimeindex()
+  dd = dd.to_dataframe()
+  dd.reset_index(inplace=True)
+  dd['year'] = dd['time'].dt.year
+  dd['month'] = dd['time'].dt.month
+  dd = dd.drop(columns=['time'])
+  monthly = pd.merge(monthly, dd, on=['year','month'], how='outer')
+  if statlist[i] == 'flux':
+    dy = pd.DataFrame(dd.groupby('year')[var].sum())
+  if statlist[i] == 'mean':
+    dy = pd.DataFrame(dd.groupby('year')[var].mean())
+  if statlist[i] == 'stock':
+    dy = dd[dd['month'] == 12]
+    dy = dy.drop(columns=['month'])
+  dy.reset_index(inplace=True)
+  if 'index' in dy.columns:
[…]
