### 2025-03-31 e5cbd03
Updating output names, adding prints
