### 2025-04-01 11397c4
Updating Thermokarst.cpp to should_ignite/initiate
ModelData.cpp changed due to rebase on to master.
