+++ src/Cohort.cpp
@@ -84,6 +84,8 @@ Cohort::Cohort(int y, int x, ModelData* modeldatapointer):
   this->fire = WildFire(modeldatapointer->fri_fire_file, 
     modeldatapointer->hist_exp_fire_file, this->cd.cell_slope, this->cd.cell_aspect, this->cd.cell_elevation, y, x);
 
+  BOOST_LOG_SEV(glg, debug) << "Setup the thermokarst information.";
+  this->fire = Thermokarst(modeldatapointer->hist_exp_fire_file, this->cd.cell_slope, this->cd.cell_aspect, this->cd.cell_elevation, y, x);
 
   this->soilenv = Soil_Env();
   
@@ -132,6 +134,10 @@ Cohort::~Cohort() {
 /** Loads data from the projected explicit fire input file into internal datastructures. */
 void Cohort::load_proj_explicit_fire(const std::string& proj_exp_fire_file){
   fire.load_projected_explicit_data(proj_exp_fire_file, y, x);
+}
+
+void Cohort::load_proj_explicit_thermokarst(const std::string& proj_exp_thermokarst_file){
+  thermokarst.load_projected_explicit_data(proj_exp_thermokarst_file, y, x);
 }
 
 /** Provides necessary data to Climate for loading projected climate data*/
@@ -757,6 +763,7 @@ void Cohort::updateMonthly_Dsb(const int & yrind, const int & currmind, std::str
   BOOST_LOG_NAMED_SCOPE("dsb");
 
   updateMonthly_Fir(yrind, currmind, stage);
+  updateMonthly_Thermokarst(yrind, currmind, stage);
 
   //updateMonthly_Flood(...)
 }
@@ -830,6 +837,77 @@ void Cohort::updateMonthly_Fir(const int & year, const int & midx, std::string s
   //Transfer monthly fire data, regardless of burn
   year_fd[midx] = *fd;
   fd->clear();
+}
+
+/** Thermokarst Disturbance module. */
+void Cohort::updateMonthly_Thermokarst(const int & year, const int & midx, std::string stage) {
+  // BOOST_LOG_NAMED_SCOPE("fire")
+
+  // // FIX ?? not sure this may no longer be necessary??
+  // // FIX? should this get moved into the "if fire" block?, or do we always zero out the FirData values?
+  // if(cd.mthsdist >= fire.getFRI()*12){
+  //   fd->fire_a2soi.orgn = 0.0;
+  // }
+  // if (midx == 0) {
+  //   fd->beginOfYear();
+  // }
+
+  // // see if it is an appropriate time to burn
+  // if ( fire.should_initiate(year, midx, stage) ) {
+
+  //   BOOST_LOG_SEV(glg, debug) << "Right before fire.burn(..)  " << ground.layer_report_string();
+
+  //   // Fire!
+  //   //  - Update C/N pools for each pft through 'bd', but not soil structure.
+  //   //  - Soil root fraction also updated through 'cd'.
+  //   fire.burn(year);
+    
+  //   BOOST_LOG_SEV(glg, debug) << "Right after fire.burn(..)  " << ground.layer_report_string();
+
+  //   BOOST_LOG_SEV(glg, debug) << "Collect burned veg C/N from individual pfts into bdall...";
+  //   for (int ip=0; ip<NUM_PFT; ip++) {
+  //     if (cd.m_veg.vegcov[ip]>0.) {
+  //       for (int i=0; i<NUM_PFT_PART; i++) {
[…]
