### 2025-04-01 1e4e274
Updating Thermokarst.cpp to should_ignite/initiate
