+++ src/ThermokarstData.cpp
@@ -1,11 +1,10 @@
 #include <string>
 
-// #include "../include/FireData.h"
 #include "../include/ThermokarstData.h"
 
 ThermokarstData::ThermokarstData() {
   // fire_a2soi.orgn = 0.0;
-  // useseverity = false;
+  useThermokarstSeverity = false;
 };
 
 ThermokarstData::~ThermokarstData() {
@@ -15,31 +14,29 @@ ThermokarstData::~ThermokarstData() {
 std::string ThermokarstData::report_to_string(const std::string& msg) {
   std::stringstream report;
 
-  // >>> review these AFTER structs shown in next function
-
-  // report << "  " << msg << "  ==== FirData REPORT ====\n";
-  // report << "   useseverity: " << useseverity << "\n";
-  // report << "   fire_soid.burnthick: " << fire_soid.burnthick << "\n";
-  // report << "\n";
-  // report << "   fire_v2a.orc: " << fire_v2a.orgc << "\n";
-  // report << "   fire_v2a.orn: " << fire_v2a.orgn << "\n";
-  // report << "\n";
-  // report << "   fire_v2soi.abvc: " << fire_v2soi.abvc << "\n";
-  // report << "   fire_v2soi.abvn: " << fire_v2soi.abvn << "\n";
-  // report << "   fire_v2soi.blwc: " << fire_v2soi.blwc << "\n";
-  // report << "   fire_v2soi.blwn: " << fire_v2soi.blwn << "\n";
-  // report << "\n";
-  // report << "   fire_soi2a.orgc: " << fire_soi2a.orgc << "\n";
-  // report << "   fire_soi2a.orgn: " << fire_soi2a.orgn << "\n";
-  // report << "\n";
-  // report << "   fire_a2soi.orgn: " << fire_a2soi.orgn << "\n";
+  report << "  " << msg << "  ==== ThermokarstData REPORT ====\n";
+  report << "   useThermokarstSeverity: " << useThermokarstSeverity << "\n";
+  report << "   fire_soid.burnthick: " << fire_soid.burnthick << "\n";
+  report << "\n";
+  report << "   fire_v2a.orc: " << fire_v2a.orgc << "\n";
+  report << "   fire_v2a.orn: " << fire_v2a.orgn << "\n";
+  report << "\n";
+  report << "   fire_v2soi.abvc: " << fire_v2soi.abvc << "\n";
+  report << "   fire_v2soi.abvn: " << fire_v2soi.abvn << "\n";
+  report << "   fire_v2soi.blwc: " << fire_v2soi.blwc << "\n";
+  report << "   fire_v2soi.blwn: " << fire_v2soi.blwn << "\n";
+  report << "\n";
+  report << "   fire_soi2a.orgc: " << fire_soi2a.orgc << "\n";
+  report << "   fire_soi2a.orgn: " << fire_soi2a.orgn << "\n";
+  report << "\n";
+  report << "   fire_a2soi.orgn: " << fire_a2soi.orgn << "\n";
 
   return report.str();
 
 }
 
 void ThermokarstData::clear() {
-  // useseverity = false;
+  useThermokarstSeverity = false;
   thermokarst_soid = soidiag_thermokarst();
   thermokarst_v2a = veg2atm_thermokarst();
   thermokarst_v2soi = veg2soi_thermokarst();
@@ -78,6 +75,6 @@ void ThermokarstData::endOfYear() {
   //
 };
 
-void ThermokarstData::burn() {
-  // clear();
+void ThermokarstData::clearing() {
+  clear();
 }
