### 2025-04-01 f649e72
Isolates per-month soil pool updates to prevent flux double accounting
