### 2025-04-01 9a88099
Renames VEGN to VEGNTOT, modifies some scripts using VEGN
