### 2025-04-03 11f212d
Adding thermokarst files to config file
