### 2025-04-03 ed6ab61
Adding thermokarst files to config file
