### 2025-04-03 2e8f659
Reassessing ebullition efflux function
