### 2025-04-11 c4ef4ba
add feature to the runmask utility for masking out CMTs that we...
...don't have parameters for...i.e. CMTs that don't exist in the
parameter files.

Only very lightly tested...
