+++ scripts/util/runmask.py
@@ -11,6 +11,7 @@ import os
 import numpy as np
 
 import util.input
+import util.param
 
 '''
 This script is for creating, modifying, and viewing run masks for dvmdostem. 
@@ -88,6 +89,67 @@ def conform_mask_timeseries(rm_m, data):
   #   print c
 
   return rm_m
+
+def conform_mask_to_available_params(param_folder, veg_file, mask_file, verbose=False):
+  """
+  Adjusts the run mask based on available CMTs in the parameter folder.
+
+  This function identifies all available Community Model Types (CMTs) in the 
+  specified parameter folder. It then modifies the run mask in the provided 
+  mask file by turning off (masking) every pixel in the vegetation file 
+  (`veg_file`) whose value is not in the list of available CMTs.
+
+  Parameters
+  ----------
+  param_folder : str
+      Path to the folder containing parameter files with available CMTs.
+  veg_file : str
+      Path to the vegetation NetCDF file containing the `veg_class` variable.
+  mask_file : str
+      Path to the run mask NetCDF file containing the `run` variable.
+  verbose : bool, optional
+      If True, prints additional information during processing. Default is False.
+
+  Returns
+  -------
+  new_mask : numpy.ndarray
+      A 2D numpy array representing the updated run mask, where pixels not 
+      matching available CMTs are turned off.
+  """
+
+  veg = nc.Dataset(os.path.join(veg_file), 'r')
+  if 'veg_class' not in veg.variables.keys():
+    raise RuntimeError(f"Invalid vegegation.nc file...missing veg_class variable {veg_file}")
+
+  mask = nc.Dataset(os.path.join(mask_file), 'r')
+  if 'run' not in mask.variables.keys():
+    raise RuntimeError(f"Invalid mask.nc file...missing run variable {mask_file}")
+
+  available_CMTs = util.param.get_available_CMTs(param_folder)
+
+  if verbose:
+    print("Available CMTs:", available_CMTs)
+    print("Initial vegetation mask:")
+    print(veg['veg_class'][:])
+
+  valid_veg = np.isin(veg['veg_class'][:], available_CMTs)
+  
+  if verbose:
+    print("Valid vegetation mask after applying available CMTs:")
+    print(valid_veg)
+
+  new_mask = np.where(valid_veg, mask['run'][:], valid_veg)
+
+  if verbose:
+    print("New run mask after applying vegetation mask:")
+    print(new_mask)
+
+  valid_veg = np.isin(veg['veg_class'][:], available_CMTs)
+  
+  new_mask = np.where(valid_veg, mask['run'][:], valid_veg)
+
+  return new_mask
+
 
 
 def conform_mask_to_inputs(in_folder, verbose=False):
@@ -271,6 +333,14 @@ def cmdline_define():
           missing data. NOTE!!: modifies run-mask.nc found in %(metavar)s, and
           ignores the files specified 'FILE' argument!'''))
 
+  parser.add_argument("--conform-mask-to-available-params", 
+      metavar=('PARAM_FOLDER', 'VEG_FILE'), nargs=2,
+      help=textwrap.dedent('''Adjust the run mask based on available CMTs in
+                           the parameter folder. This will mask out any pixels 
+                           whose CMT classification in the  VEG_FILE is not in 
+                           the list of available CMTs from the PARAM_FOLDER.
+                           Modifies the run-mask file.'''))
+
   parser.add_argument("--select-only-cmt", metavar=('FOLDER','CMT'), nargs=2,
     help=textwrap.dedent('''Select only pixels with a certain CMT number. This
           will honor any pixels that are alreay masked in the run-mask. As such, 
[…]
