### 2025-04-15 cfd2e7c
Modifies input checking method, makes available on CL
