### 2025-04-15 4947838
Adds input method to return CMT values and counts
