### 2025-04-15 4f33a14
Adds run-mask method to return active, inactive, and invalid counts
