### 2025-04-16 #7667 @tobeycarman
@mengqi-z, I'm sorry but I am a little confused, what is the next step to wrap this up? Do I need to make the tagged release (v0.8.2) and then the review is finalized or does the review (this thread) need to be finalized before I make the tagged release? Also, just noticing that this review issue thread was started referencing v0.8.0, but the paper will actually be for v0.8.2. Will this create any problems?
