### 2025-04-17 bf80cf4
Thermokarst related variables naming changes
