### 2025-04-17 6580222
Thermokarst related variables naming changes
