### 2025-04-17 eab0f03
Fixing reinitialization bug for more pft->less pft
Added code to account for some variables in vegetation
not getting cleared after cmt change. This required
some duplicated function calls and some manual editing
in the cmt_change function. Could be cleaned up but
working for now.
