### 2025-04-21 5b5a12b
Tweaks new run-mask name, swaps np.int for int
