### 2025-04-21 a58a4af
Copy original mask instead of creating a new file
This retains all other variables and global
attributes.
