### 2025-04-22 bebce64
Updates CL help message
