### 2025-04-25 #7667 @mengqi-z
Hi @tobeycarman - Sorry for the delayed reply as I was out of the office for few weeks.

The tagged release will only be required once all reviews are completed and the reviewers have recommended the paper for acceptance. It's fine that the review issue references v0.8.0, even if you've updated to newer versions since then.
