### 2025-05-01 bc5e10d
In progress soil parameter interpolation
