+++ src/Cohort.cpp
@@ -369,7 +369,14 @@ void Cohort::updateMonthly(const int & yrcnt, const int & currmind,
 
   //
   if(currmind==0) {
+
     cd.beginOfYear();
+
+    if(>x_years_from_disturbance){ // come up with a way of working out if dsb has occurred... 
+      interpolateSoilParameters(); // variable DISTURBANCE_HAS_HAPPENED?
+                                   // LAST_DISTURBANCE_MNTH == -9999
+                                   // LAST_DISTURBANCE_YEAR == -9999
+    }
   }
 
   BOOST_LOG_SEV(glg, debug) << "Clean up before a month starts.";
@@ -1498,8 +1505,15 @@ void Cohort::cmtChange(const int & currmind){
   // veg.updateFrootfrac();
 
   // SOILS
+
+  // store / archive old soil params
+
   // load soil parameters
   // chtlu.loadSoilParams();
+
+
+
+
 
   /*
 
@@ -1524,185 +1538,216 @@ void Cohort::cmtChange(const int & currmind){
   */
 };
 
-/** Synchronizes Cohort and CohortData's internal fields from the
- * RestartData object.
-*/
-void Cohort::set_state_from_restartdata() {
-  BOOST_LOG_SEV(glg, info) << "Updating this Cohort and CohortData object with "
-                           << "values from the RestartData object...";
+/*
+ * Interpolating soil parameters following cmt change
+ */
+void Cohort::interpolateSoilParameters(){
 
-  veg.set_state_from_restartdata(this->restartdata);
-  ground.set_state_from_restartdata(&cd.d_snow, &cd.m_soil, this->restartdata);
-  snowenv.set_state_from_restartdata(this->restartdata);
-  soilenv.set_state_from_restartdata(this->restartdata);
-  solprntenv.set_state_from_restartdata(this->restartdata);
-  soilbgc.set_state_from_restartdata(this->restartdata);
-  fire.set_state_from_restartdata(this->restartdata);
+  int tsd = this->cd.yrsdist;
 
-  for(int ii=0; ii<NUM_PFT; ii++){
-    vegbgc[ii].set_state_from_restartdata(this->restartdata);
-    vegenv[ii].set_state_from_restartdata(this->restartdata);
-  }
+  this->soilbgc.calpar.kdcrawc = interpolate(chtlu->archive.oldkdc, chtlu->newkdc, tsd/n_years, 'linear');
 
-  //  FIX: how to handle ground?? and snwstate_dim?? Looks like it is
-  //  used for both m_snow and d_snow and y_snow??? which to update???
-  // add more here....
-  // Right now, ground is instantiated when the model is, and various
-  // parts of it are updated in different functions above. It is
-  // difficult to tell if all ground values are being updated correctly
-  // since so many values are out of bounds when written to the
-  // restart file. 
+  /*
 
+  kdc_int = (tsd / nyears) * kdc_new + (1 - tsd / nyears) * kdc_old;
+
+  */
 }
 
+  /** Synchronizes Cohort and CohortData's internal fields from the
+   * RestartData object.
+   */
+  void Cohort::set_state_from_restartdata()
+  {
+    BOOST_LOG_SEV(glg, info) << "Updating this Cohort and CohortData object with "
[…]
