+++ mads_calibration/calibration_files/plotting/utils.py
@@ -1,0 +1,978 @@
+import csv
+import os
+import numpy as np
+from sklearn.cluster import KMeans
+from sklearn.metrics import r2_score,mean_squared_error,mean_absolute_error
+from sklearn.metrics import mean_absolute_percentage_error
+from scipy.optimize import curve_fit
+from matplotlib import pyplot as plt
+import pandas as pd
+import statistics as stat
+import seaborn as sns
+import math
+from scipy.stats import linregress
+from sklearn.ensemble import RandomForestRegressor
+from sklearn.linear_model import LinearRegression
+from sklearn.neighbors import KNeighborsRegressor
+from sklearn.model_selection import train_test_split
+#-------------------------------FUNCTIONS TO LOAD CSV FILES----------------------------------------------------
+
+def spaghetti_match_plot(df_x,df_y,logy=False):
+    ''' plots the spaghetti plot of modeled v.s. observed values 
+        df_x: parameter dataframe
+        df_y: model output dataframe
+        logy: True enables the logplot option
+    '''
+    df_y.iloc[0:-1,:].transpose().plot(logy=logy,legend=False,alpha=0.5,figsize=(10,5))
+    nrange=range(len(df_y.columns))
+    ax = df_y.iloc[-1,:].plot(logy=logy,legend=True,style="o",color='red',xticks=nrange, rot=90, label="Targets");
+    ax.set_xticklabels(df_y.columns,fontsize=12)
+    # ax.set_xlabel("Parameters",fontsize=14)
+    # ax.set_ylabel("Targets",fontsize=14)
+    
+def spaghetti_match_plot_r2(df_x,df_y,r2lim=0.5):
+    ''' plots the spaghetti plot of restricted by R^2 value modeled v.s. observed values 
+        df_x: parameter dataframe
+        df_y: model output dataframe
+        r2lim: selects outputs > r2lim (-inf,1.0)
+    '''
+    xparams, ymodel = get_params(df_x,df_y,r2lim)
+    ymodel.iloc[0:-1,:].transpose().plot(legend=False,alpha=0.5,figsize=(10,5))
+    nrange=range(len(df_y.columns))
+    ax = df_y.iloc[-1,:].plot(legend=False,style="o",color='red',xticks=nrange, rot=90, label="Targets");
+    ax.set_xticklabels(df_y.columns,fontsize=12)
+    # ax.set_xlabel("Parameters",fontsize=14)
+    # ax.set_ylabel("Targets",fontsize=14)
+    
+    return
+    
+def plot_r2_rmse(df_y):
+    ''' plots R2 v.s. RMSE, with the condition that last row of the dataframe are targets
+        df_y: model output dataframe
+    '''
+    [n,m]=np.shape(df_y)
+    r2=[r2_score(df_y.iloc[i,:], df_y.iloc[-1,:]) for i in range(n-1)]
+    rmse=[mean_squared_error(df_y.iloc[i,:], df_y.iloc[-1,:]) for i in range(n-1)]
+    r2=np.asarray(r2)
+    rmse=np.asarray(rmse)
+    plt.plot(rmse,r2,'o'), plt.xlabel('RMSE'), plt.ylabel('$R^2$');
+    
+    return rmse
+   
+def get_params_r2_rmse(x,y,r2lim=0.95,n_top_runs=None):
+    '''
+    Inputs:
+    x: parameters dataframe 
+    y: model outputs dataframe
+    r2lim: the R square limit
+    n_top_runs (int): if specified, will output top n params and results based on combined accuracy (RMSE, MAPE, R2)
+    
[…]
