+++ mads_calibration/calibration_files/plotting/20_percent/check_all-run/info.txt
@@ -1,0 +1,1 @@
+sampling_method: lhc
