### 2025-05-06 84bd213
Changes to allow compilation
Further debugging and integrating appropriate thermokarst
processes are needed.
