+++ include/diagnostics.h
@@ -252,9 +252,9 @@ struct soidiag_fir {
 };
 
 struct soidiag_thermokarst {
-  double thermokarstthick;
+  double removal_thickness;
   double rolr;
-  soidiag_thermokarst() : thermokarstthick(UIN_D), rolr(UIN_D) {}
+  soidiag_thermokarst() : removal_thickness(UIN_D), rolr(UIN_D) {}
 };
 
 #endif /*DIAGNOSTICS_H_*/
