### 2025-05-06 53a4aba
Changes to allow compilation
Further debugging and integrating appropriate thermokarst
processes are needed.
