+++ include/Soil_Env.h
@@ -8,6 +8,7 @@
 #include "CohortData.h"
 #include "EnvData.h"
 #include "FireData.h"
+#include "ThermokarstData.h"
 #include "RestartData.h"
 
 #include "errorcode.h"
