+++ src/TEM.cpp
@@ -909,6 +909,7 @@ void advance_model(const int rowidx, const int colidx,
     runner.cohort.load_proj_climate(modeldata.proj_climate_file);
     runner.cohort.load_proj_co2(modeldata.proj_co2_file);
     runner.cohort.load_proj_explicit_fire(modeldata.proj_exp_fire_file);
+    runner.cohort.load_proj_explicit_thermokarst(modeldata.proj_exp_thermokarst_file);
 
     BOOST_LOG_SEV(glg, warn) << "MAKE SURE YOUR FIRE INPUTS ARE SETUP CORRECTLY!";
 
