### 2025-05-07 8c551cd
Initial soil range carbon output
Only works starting from soil surface currently.
