### 2025-05-07 b4c7898
Initial soil range carbon output
Only works starting from soil surface currently.
