+++ src/Runner.cpp
@@ -4885,6 +4885,30 @@ void Runner::output_netCDF(std::map<std::string, OutputSpec> &netcdf_outputs, in
   map_itr = netcdf_outputs.end();
 
 
+  //SOC from 0cm to 30cm
+  map_itr = netcdf_outputs.find("SOC0_30");
+  if (map_itr != netcdf_outputs.end()) {
+    BOOST_LOG_SEV(glg, debug) << "NetCDF output: SOC0_30";
+    curr_spec = map_itr->second;
+
+    #pragma omp critical(outputSOC0_30)
+    {
+      double m_soc_0_30 = cohort.ground.getCarbonForDepthRange(0.0, 0.3);
+
+      if(curr_spec.monthly)
+
+        outhold.soc_0_30_for_output.push_back(m_soc_0_30);
+
+        if (output_this_timestep) {
+          output_nc_3dim(&curr_spec, file_stage_suffix, &outhold.soc_0_30_for_output[0], 1, month_start_idx, months_to_output);
+          outhold.soc_0_30_for_output.clear();
+        }
+
+    } //end critical(outputSOC0_30)
+  } //end SOC0_30
+  map_itr = netcdf_outputs.end();
+
+
   //SOCFROZEN
   map_itr = netcdf_outputs.find("SOCFROZEN");
   if (map_itr != netcdf_outputs.end()) {
