+++ src/CohortLookup.cpp
@@ -158,7 +158,7 @@ std::string CohortLookup::calparbgc2str() {
   s << this->kdcsomcr << "// kdcsomcr:\n";
   s << this->max_ponding_s << "// summermaxponding: \n:";
   s << this->max_ponding_w << "// wintermaxponding: \n:";
-  s << this->max_ponding_w << "// inflowfactor: \n:";
+  s << this->inflow_factor << "// inflowfactor: \n:";
   s << this->rhq10 << "// rhq10:\n";
   s << this->s2dfraction << "// s2dfraction \n:";
   s << this->d2mfraction << "// d2mfraction \n:";
@@ -174,6 +174,7 @@ std::string CohortLookup::calparbgc2str() {
   s << this->oxidTref_ch4 << "// oxidTref_ch4:\n";
   s << this->oxidkm_ch4 << "// oxidkm_ch4:\n";
   s << this->oxidVmax_ch4 << "// oxidVmax_ch4:\n";
+  s << this->rhmoist << "// rhmoist:\n";
   return s.str();
 }
 
@@ -183,7 +184,7 @@ void CohortLookup::assignBgcCalpar(std::string & dircmt) {
 
   // get a list of data for the cmt number
   std::list<std::string> l = temutil::parse_parameter_file(
-      dircmt + "cmt_calparbgc.txt", temutil::cmtcode2num(this->cmtcode), 36
+      dircmt + "cmt_calparbgc.txt", temutil::cmtcode2num(this->cmtcode), 37
   );
 
   // pop each line off the front of the list
@@ -230,6 +231,8 @@ void CohortLookup::assignBgcCalpar(std::string & dircmt) {
 
   temutil::pfll2data(l, oxidkm_ch4);
   temutil::pfll2data(l, oxidVmax_ch4);
+
+  temutil::pfll2data(l, rhmoist);
 }
 
 /** Assign "veg dimension?" from parameter file. */
