### 2025-05-12 597251a
adjusting util scripts for thermokarst input
