### 2025-05-12 4313d88
adjusting util scripts for thermokarst input
