### 2025-05-13 uaf-arctic-eco-modeling/dvm-dos-tem#782 @mengqi-z
JOSS Paper Proofread
Hi @tobeycarman,

This is for the paper submitted to JOSS https://github.com/openjournals/joss-reviews/issues/7667

Could you please address the following issues in the paper?

1. Add the missing DOI for the reference "Taking off the training wheels: The properties of a dynamic vegetation model without climate envelopes". The correct DOI is: 10.5194/gmdd-8-3293-2015
2. Add the missing DOI for the reference "Design, development, and application of LANDIS-II, a spatial landscape simulation model with flexible temporal and spatial resolution". The DOI is: 10.1016/j.ecolmodel.2006.10.009
3. Fix the DOI for the reference "An individual-based process model to simulate landscape-scale forest ecosystem dynamics" by removing the duplicated "https://doi.org/"
4. Define the acronym PFTs (Plant Functional Types) at its first mention on line 75.

In addition, please make sure the author names and institutions are correct. Thanks!
