### 2025-05-14 e4756f6
fixing oxidation bug in CH4 code
