### 2025-05-15 6d6981a
Address issue #782 for joss paper: fix DOIs, define acronym
