+++ joss-paper/joss-paper.bib
@@ -39,6 +39,7 @@
 @article{Fisher2015,
   title={Taking off the training wheels: the properties of a dynamic vegetation model without climate envelopes},
   author={Fisher, RA and Muszala, S and Verteinstein, M and Lawrence, P and Xu, C and McDowell, NG and Knox, RG and Koven, C and Holm, J and Rogers, BM and others},
+  doi = {10.5194/gmdd-8-3293-2015},
   journal={Geoscientific Model Development Discussions (Online)},
   volume={8},
   number={4},
@@ -178,6 +179,7 @@
 @article{Scheller2007,
   title={Design, development, and application of LANDIS-II, a spatial landscape simulation model with flexible temporal and spatial resolution},
   author={Scheller, Robert M and Domingo, James B and Sturtevant, Brian R and Williams, Jeremy S and Rudy, Arnold and Gustafson, Eric J and Mladenoff, David J},
+  doi = {10.1016/j.ecolmodel.2006.10.009},
   journal={ecological modelling},
   volume={201},
   number={3-4},
@@ -217,7 +219,7 @@
    pages = {87-100},
    year = {2012},
    issn = {0304-3800},
-   doi = {https://doi.org/10.1016/j.ecolmodel.2012.02.015},
+   doi = {10.1016/j.ecolmodel.2012.02.015},
    url = {https://www.sciencedirect.com/science/article/pii/S0304380012000919},
    author = {Rupert Seidl and Werner Rammer and Robert M. Scheller and Thomas A. Spies},
    keywords = {Forest ecosystem dynamics, Complex adaptive systems, Individual-based modeling, Ecological field theory, Hierarchical multi-scale modeling, Forest structure and functioning},
