+++ include/Cohort.h
@@ -150,6 +150,7 @@ private:
   void assignSoilBd2pfts_monthly();
   void getBd4allveg_monthly();
   void cmtChange(const int &currmind);
-
+  void interpolateSoilParameters();
+  
   };
 #endif /*COHORT_H_*/
