+++ include/CohortLookup.h
@@ -34,6 +34,17 @@ public:
   // Loading soil-specific parameters
   void loadSoilParams();
 
+  // archived parameters
+  soipar_cal archive_soical_params;
+  soipar_dim archive_soidim_params;
+  soipar_env archive_soienv_params;
+  soipar_bgc archive_soibgc_params;
+  snwpar_dim archive_snwdim_params;
+  snwpar_env archive_snwenv_params;
+
+  // method for archiving soil params
+  void archiveSoilParams();
+
   //calibration related
   //vegetation
   double cmax[NUM_PFT];
