### 2025-05-21 8c7d17a
Adds support for arbitrary points in the soil column
