### 2025-05-21 c8352af
Adds support for arbitrary points in the soil column
