### 2025-05-21 90db0e7
Adds VWC output for arbitrary sections of soil
