+++ src/Runner.cpp
@@ -5985,6 +5985,43 @@ void Runner::output_netCDF(std::map<std::string, OutputSpec> &netcdf_outputs, in
   map_itr = netcdf_outputs.end();
 
 
+  //VWC from 0cm to 30cm
+  map_itr = netcdf_outputs.find("VWC0_30cm");
+  if (map_itr != netcdf_outputs.end()) {
+    BOOST_LOG_SEV(glg, debug) << "NetCDF output: VWC0_30cm";
+    curr_spec = map_itr->second;
+
+    #pragma omp critical(outputVWC0_30cm)
+    {
+
+      //Monthly
+      if(curr_spec.monthly){
+
+        double m_vwc_0_30 = cohort.ground.getVWCForDepthRange(0.0, 0.3);
+        outhold.vwc_0_30_for_output.push_back(m_vwc_0_30);
+
+        if (output_this_timestep) {
+          output_nc_3dim(&curr_spec, file_stage_suffix, &outhold.vwc_0_30_for_output[0], 1, month_start_idx, months_to_output);
+          outhold.vwc_0_30_for_output.clear();
+        }
+      }
+      //Yearly
+      else if(curr_spec.yearly){
+
+        double y_vwc_0_30 = cohort.ground.getVWCForDepthRange(0.0, 0.3);
+        outhold.vwc_0_30_for_output.push_back(y_vwc_0_30);
+
+        if (output_this_timestep) {
+          output_nc_3dim(&curr_spec, file_stage_suffix, &outhold.vwc_0_30_for_output[0], 1, year_start_idx, years_to_output);
+          outhold.vwc_0_30_for_output.clear();
+        }
+
+      }
+    } //end critical(outputVWC0_30cm)
+  } //end VWC0_30cm
+  map_itr = netcdf_outputs.end();
+
+
   //VWCDEEP
   map_itr = netcdf_outputs.find("VWCDEEP");
   if(map_itr != netcdf_outputs.end()){
