### 2025-05-21 5e0dac4
Adds VWC output for arbitrary sections of soil
