### 2025-05-21 4a9d74a
Roughs out synthetic depth temperature output
This depends on the temutil::interpolate() method
which only exists on the cmt_change branch. A future
patch will add it here, at which point the code should
function again.
