### 2025-05-21 3b7e64e
Roughs out synthetic depth temperature output
This depends on the temutil::interpolate() method
which only exists on the cmt_change branch. A future
patch will add it here, at which point the code should
function again.
