+++ include/OutputHolder.h
@@ -81,6 +81,8 @@ public:
   std::vector<std::array<double, MAX_SOI_LAY>> layerdz_for_output;
   std::vector<std::array<int, MAX_SOI_LAY>> layertype_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> lwclayer_for_output;
+
+  std::vector<double> tsoil_30_for_output;
   std::vector<std::array<double, MAX_SOI_LAY>> tlayer_for_output;
 
   std::vector<double> vwc_0_30_for_output;
