### 2025-05-23 8a76520
adding in catches for very low alpha values
