### 2025-05-23 e95b5bd
adding all archive parameters
