### 2025-05-29 #7667 @mengqi-z
@tobeycarman - Just checking to see if you have been able to complete all the items listed. Once you are ready, we can move forward to the next stage to get your paper accepted and published.

> [@tobeycarman](https://github.com/tobeycarman) - Thank you!
>
> For the next steps, could you:
>
> * [ ]  Double check authors and affiliations (including ORCIDs).[ ]  Make a release of the software with the latest changes from the review and **post the version number here**. This is the version that will be used in the JOSS paper.[ ]  Archive the released software on Zenodo or a similar service (e.g., figshare, an institutional repository).[ ]  Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.[ ]  Make sure that the license listed for the archive is the same as the software license.[ ]  Please **post the DOI of the archived version here**.
>
> I can then move forward with accepting the submission.
