### 2025-05-30 750e995
Add custom exception class and handling to allow us to pass over this issue with define mode.
Previoulsly there was not an issue ending define mode. Now with more
recent netcdf, it seems like it returns a negative status code if you
are already out of define mode and you try to leave it. So here
we simply handle that exception case and move on.
