### 2025-05-31 bf169d8
Fix order in new TEMUtilityFuncitons netcdf handling functions.
previous order resulted in unwanted std out in cases of handled errors.
