### 2025-06-04 7f48280
Introduces interpolation() from the cmt_change branch
This method was implemented as part of the cmt change
development and is patched in here as it will not be
merged to the main branch for a while yet.
