### 2025-06-09 ad9c313
Bump version number in Sphinx docs.
