### 2025-06-09 #7667 @tobeycarman
@mengqi-z I believe I have completed all the checklist items.

- [x] Double check authors and affiliations (including ORCIDs).
- [x] [Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
- [x] [Archive the released software on Zenodo or a similar service (e.g., figshare, an institutional repository).
- [x] Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.
- [x] Please post the DOI of the archived version here.

Released DVMDOSTEM v0.8.2, June 9, 2025
Github Release: https://github.com/uaf-arctic-eco-modeling/dvm-dos-tem/releases/tag/v0.8.2
Zenodo archive: https://zenodo.org/records/15625536
DOI: https://doi.org/10.5281/zenodo.15625536
