### 2025-06-10 #7667 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
