### 2025-06-13 c66b834
Quick fix to allow (small) regional DA
