### 2025-06-19 50453a8
Archiving and interpolating between old and new cmt
Some code is commented for testing, specifically which parameters
are reloaded and archived, as we want to use mostly the same as
the initial CMT but may need revisiting.
