+++ src/CohortLookup.cpp
@@ -62,9 +62,9 @@ void CohortLookup::loadVegetationParams() {
 void CohortLookup::loadSoilParams() {
   BOOST_LOG_SEV(glg, info) << "Load soil parameters from file...";
   assignBgcCalparSoil(dir);
-  assignGroundDimension(dir);
-  assignEnv4Ground(dir);
-  assignBgc4Ground(dir);
+  // assignGroundDimension(dir);
+  // assignEnv4Ground(dir);
+  // assignBgc4Ground(dir);
 }
 
 /*
@@ -79,64 +79,64 @@ void CohortLookup::archiveSoilParams(){
   archive_soical_params.kdcsompr = this->kdcsompr;
   archive_soical_params.kdcsomcr = this->kdcsomcr;
 
-  archive_snwdim_params.snwdenmax = this->snwdenmax;
-  archive_snwdim_params.snwdennew = this->snwdennew;
-  archive_snwdim_params.initsnwthick = this->initsnwthick;
-  archive_snwdim_params.initsnwdense = this->initsnwdense;
+  // archive_snwdim_params.snwdenmax = this->snwdenmax;
+  // archive_snwdim_params.snwdennew = this->snwdennew;
+  // archive_snwdim_params.initsnwthick = this->initsnwthick;
+  // archive_snwdim_params.initsnwdense = this->initsnwdense;
   
-  archive_soidim_params.maxdmossthick = this->maxdmossthick;
-  archive_soidim_params.initdmossthick = this->initdmossthick;
-  archive_soidim_params.mosstype = this->mosstype;
-  archive_soidim_params.initfibthick = this->initfibthick;
-  archive_soidim_params.inithumthick = this->inithumthick;
-  archive_soidim_params.coefshlwa = this->coefshlwa;
-  archive_soidim_params.coefshlwb = this->coefshlwb;
-  archive_soidim_params.coefdeepa = this->coefdeepa;
-  archive_soidim_params.coefdeepb = this->coefdeepb;
-  archive_soidim_params.coefminea = this->coefminea;
-  archive_soidim_params.coefmineb = this->coefmineb;
+  // archive_soidim_params.maxdmossthick = this->maxdmossthick;
+  // archive_soidim_params.initdmossthick = this->initdmossthick;
+  // archive_soidim_params.mosstype = this->mosstype;
+  // archive_soidim_params.initfibthick = this->initfibthick;
+  // archive_soidim_params.inithumthick = this->inithumthick;
+  // archive_soidim_params.coefshlwa = this->coefshlwa;
+  // archive_soidim_params.coefshlwb = this->coefshlwb;
+  // archive_soidim_params.coefdeepa = this->coefdeepa;
+  // archive_soidim_params.coefdeepb = this->coefdeepb;
+  // archive_soidim_params.coefminea = this->coefminea;
+  // archive_soidim_params.coefmineb = this->coefmineb;
 
-  archive_snwenv_params.snwalbmax = this->snwalbmax;
-  archive_snwenv_params.snwalbmin = this->snwalbmin;
-  archive_snwenv_params.initsnwtem = this->initsnwtem;
+  // archive_snwenv_params.snwalbmax = this->snwalbmax;
+  // archive_snwenv_params.snwalbmin = this->snwalbmin;
+  // archive_snwenv_params.initsnwtem = this->initsnwtem;
 
-  archive_soienv_params.psimax = this->psimax;
-  archive_soienv_params.evapmin = this->evapmin;
-  archive_soienv_params.drainmax = this->drainmax;
-  archive_soienv_params.rtdp4gdd = this->rtdp4gdd;
[…]
