### 2025-06-19 6df4f11
Implement changes to initiate thermokarst in cohort
