+++ config/config.js
@@ -100,7 +100,7 @@
       "nfeed": true,
       "avlnflg": true,
       "baseline": true,
-      "dsb": false,
+      "dsb": true,
       "dsl": true,
       "dyn_lai": true
     },
