### 2025-06-19 e0d1e18
Implement changes to initiate thermokarst in cohort
