+++ src/Cohort.cpp
@@ -219,10 +219,11 @@ void Cohort::setModelData(ModelData* mdp) {
   md = mdp;
 };
 
-void Cohort::setProcessData(EnvData * alledp, BgcData * allbdp, FirData *fdp) {
+void Cohort::setProcessData(EnvData * alledp, BgcData * allbdp, FirData *fdp, ThermokarstData * tdp) {
   edall = alledp;
   bdall = allbdp;
   fd = fdp;
+  td = tdp;
   bdall->cd = &cd;
   edall->cd = &cd;
 
@@ -316,6 +317,11 @@ void Cohort::initialize_state_parameters() {
   fire.initializeState();
 
   fire.initializeParameter();
+
+  // thermokarst processes - these do not do anything but following convention
+  td->init();
+
+  thermokarst.initializeState();
   
   BOOST_LOG_SEV(glg, debug) << "Done with Cohort::initStatepar()!  " << ground.layer_report_string("depth thermal ptr desc");
 }
@@ -764,7 +770,7 @@ void Cohort::updateMonthly_Dsb(const int & yrind, const int & currmind, std::str
   BOOST_LOG_NAMED_SCOPE("dsb");
 
   updateMonthly_Fir(yrind, currmind, stage);
-  // updateMonthly_Thermokarst(yrind, currmind, stage);
+  updateMonthly_Thermokarst(yrind, currmind, stage);
 
   //updateMonthly_Flood(...)
 }
@@ -842,73 +848,71 @@ void Cohort::updateMonthly_Fir(const int & year, const int & midx, std::string s
 
 /** Thermokarst Disturbance module. */
 void Cohort::updateMonthly_Thermokarst(const int & year, const int & midx, std::string stage) {
-  // BOOST_LOG_NAMED_SCOPE("fire")
+  BOOST_LOG_NAMED_SCOPE("thermokarst")
 
-  // // FIX ?? not sure this may no longer be necessary??
-  // // FIX? should this get moved into the "if fire" block?, or do we always zero out the FirData values?
-  // if(cd.mthsdist >= fire.getFRI()*12){
-  //   fd->fire_a2soi.orgn = 0.0;
-  // }
-  // if (midx == 0) {
-  //   fd->beginOfYear();
-  // }
+  // This does not current 
+  if (midx == 0) {
+    td->beginOfYear();
+  }
 
-  // // see if it is an appropriate time to burn
-  // if ( fire.should_initiate(year, midx, stage) ) {
+  // see if it is an appropriate time to initiate
+  if ( thermokarst.should_initiate(year, midx, stage) ) {
 
-  //   BOOST_LOG_SEV(glg, debug) << "Right before fire.burn(..)  " << ground.layer_report_string();
+    BOOST_LOG_SEV(glg, debug) << "Right before thermokarst.initiate(..)  " << ground.layer_report_string();
 
-  //   // Fire!
-  //   //  - Update C/N pools for each pft through 'bd', but not soil structure.
-  //   //  - Soil root fraction also updated through 'cd'.
-  //   fire.burn(year);
+    // Thermokarst!
+    //  - Update C/N pools for each pft through 'bd', but not soil structure.
+    //  - Soil root fraction also updated through 'cd'.
+    thermokarst.initiate(year);
     
-  //   BOOST_LOG_SEV(glg, debug) << "Right after fire.burn(..)  " << ground.layer_report_string();
[…]
