### 2025-06-20 bdd2e4c
Formatting tweaks for consistency
