+++ src/Soil_Bgc.cpp
@@ -365,24 +365,24 @@ void Soil_Bgc::set_state_from_restartdata(const RestartData & rdata) {
 
 void Soil_Bgc::initializeParameter() {
   BOOST_LOG_SEV(glg, info) << "Initializing parameters in Soil_Bgc from chtlu (CohortLookup) values.";
-  calpar.micbnup    = chtlu->micbnup;
-  calpar.kdcrawc    = chtlu->kdcrawc;
-  calpar.kdcsoma    = chtlu->kdcsoma;
-  calpar.kdcsompr   = chtlu->kdcsompr;
-  calpar.kdcsomcr   = chtlu->kdcsomcr;
-  bgcpar.rhq10      = chtlu->rhq10;
+  calpar.micbnup = chtlu->micbnup;
+  calpar.kdcrawc = chtlu->kdcrawc;
+  calpar.kdcsoma = chtlu->kdcsoma;
+  calpar.kdcsompr = chtlu->kdcsompr;
+  calpar.kdcsomcr = chtlu->kdcsomcr;
+  bgcpar.rhq10 = chtlu->rhq10;
   bgcpar.rhmoistfrozen = chtlu->rhmoistfrozen;
-  bgcpar.moistmin   = chtlu->moistmin;
-  bgcpar.moistmax   = chtlu->moistmax;
-  bgcpar.moistopt   = chtlu->moistopt;
-  bgcpar.fsoma      = chtlu->fsoma;
-  bgcpar.fsompr     = chtlu->fsompr;
-  bgcpar.fsomcr     = chtlu->fsomcr;
-  bgcpar.som2co2    = chtlu->som2co2;
-  bgcpar.lcclnc     = chtlu->lcclnc;
-  bgcpar.kn2        = chtlu->kn2;
-  bgcpar.propftos   = chtlu->propftos;
-  bgcpar.fnloss     = chtlu->fnloss;
+  bgcpar.moistmin = chtlu->moistmin;
+  bgcpar.moistmax = chtlu->moistmax;
+  bgcpar.moistopt = chtlu->moistopt;
+  bgcpar.fsoma = chtlu->fsoma;
+  bgcpar.fsompr = chtlu->fsompr;
+  bgcpar.fsomcr = chtlu->fsomcr;
+  bgcpar.som2co2 = chtlu->som2co2;
+  bgcpar.lcclnc = chtlu->lcclnc;
+  bgcpar.kn2 = chtlu->kn2;
+  bgcpar.propftos = chtlu->propftos;
+  bgcpar.fnloss = chtlu->fnloss;
   bgcpar.nmincnsoil = chtlu->nmincnsoil;
 
   BOOST_LOG_SEV(glg, info) << "Calculating parameter in Soil_Bgc from Jenkinson and Rayner (1977).";
@@ -404,10 +404,10 @@ void Soil_Bgc::initializeParameter() {
 void Soil_Bgc::initSoilCarbon(double & initshlwc, double & initdeepc,
                               double & initminec) {
   for(int il =0; il <MAX_SOI_LAY ; il++) {
-    bd->m_sois.rawc[il]  = 0.;
-    bd->m_sois.soma[il]  = 0.;
-    bd->m_sois.sompr[il] = 0.;
-    bd->m_sois.somcr[il] = 0.;
+    bd->m_sois.rawc[il] = 0.0;
+    bd->m_sois.soma[il] = 0.0;
+    bd->m_sois.sompr[il] = 0.0;
+    bd->m_sois.somcr[il] = 0.0;
   }
 
   initOslayerCarbon(initshlwc, initdeepc);
@@ -589,7 +589,7 @@ void Soil_Bgc::deltac() {
   for (int il =0; il<cd->m_soil.numsl; il++) {
[…]
