### 2025-06-23 uaf-arctic-eco-modeling/dvm-dos-tem#791 @tobeycarman
Final adjustments for JOSS paper
Wrapping up a few loose ends, from here:
https://github.com/openjournals/joss-reviews/issues/7667#issuecomment-2988811578

Namely:

- [ ] Fix capitalization in references section
- [ ] remove extraneous comma
- [ ] align title in main repo `.cff` file with the JOSS title
- [ ] add note to `.cff` file for citing JOSS paper
