### 2025-06-23 ecaa815
fix capitalization in joss-paer.bib
