+++ joss-paper/joss-paper.bib
@@ -19,7 +19,7 @@
    issue = {2},
    journal = {Ecological Applications},
    month = {3},
-   title = {Assessing dynamic vegetation model parameter uncertainty across Alaskan arctic tundra plant communities},
+   title = {Assessing dynamic vegetation model parameter uncertainty across {A}laskan arctic tundra plant communities},
    volume = {32},
    year = {2022},
 }
@@ -32,7 +32,7 @@
    journal = {Ecological Applications},
    month = {6},
    pages = {1022-1043},
-   title = {Changes in vegetation in northern Alaska under scenarios of climate change, 2003–2100: implications for climate feedbacks},
+   title = {Changes in vegetation in northern {A}laska under scenarios of climate change, 2003–2100: implications for climate feedbacks},
    volume = {19},
    year = {2009},
 }
@@ -55,7 +55,7 @@
    journal = {Ecological Applications},
    month = {1},
    pages = {5-27},
-   title = {The role of driving factors in historical and projected carbon dynamics of upland ecosystems in Alaska},
+   title = {The role of driving factors in historical and projected carbon dynamics of upland ecosystems in {A}laska},
    volume = {28},
    year = {2018},
 }
@@ -67,7 +67,7 @@
    journal = {Environmental Research Letters},
    month = {12},
    pages = {045016},
-   title = {Modeling the effects of fire severity and climate warming on active layer thickness and soil carbon storage of black spruce forests across the landscape in interior Alaska},
+   title = {Modeling the effects of fire severity and climate warming on active layer thickness and soil carbon storage of black spruce forests across the landscape in interior {A}laska},
    volume = {8},
    year = {2013},
 }
@@ -80,7 +80,7 @@
    journal = {Environmental Research Letters},
    month = {9},
    pages = {093003},
-   title = {Comparing assumptions and applications of dynamic vegetation models used in the Arctic-Boreal zone of Alaska and Canada},
+   title = {Comparing assumptions and applications of dynamic vegetation models used in the {A}rctic-{B}oreal zone of {A}laska and {C}anada},
    volume = {19},
    year = {2024}
 }
@@ -104,7 +104,7 @@
    issn = {17489326},
    issue = {3},
    journal = {Environmental Research Letters},
-   title = {The effects of fire on the thermal stability of permafrost in lowland and upland black spruce forests of interior Alaska in a changing climate},
+   title = {The effects of fire on the thermal stability of permafrost in lowland and upland black spruce forests of interior {A}laska in a changing climate},
    volume = {8},
    year = {2013},
 }
@@ -128,7 +128,7 @@
    journal = {Journal of Advances in Modeling Earth Systems},
    month = {12},
    pages = {4245-4287},
-   title = {The Community Land Model Version 5: Description of New Features, Benchmarking, and Impact of Forcing Uncertainty},
+   title = {The {C}ommunity {L}and {M}odel {V}ersion 5: {D}escription of {N}ew {F}eatures, {B}enchmarking, and {I}mpact of {F}orcing {U}ncertainty},
    volume = {11},
    year = {2019},
 }
@@ -139,7 +139,7 @@
    issn = {19449224},
    issue = {2},
    journal = {Global Biogeochemical Cycles},
-   title = {Interactions between carbon and nitrogen dynamics in estimating net primary productivity for potential vegetation in North America},
+   title = {Interactions between carbon and nitrogen dynamics in estimating net primary productivity for potential vegetation in {N}orth {A}merica},
    volume = {6},
    year = {1992},
 }
@@ -150,7 +150,7 @@
    issn = {10510761},
    issue = {4},
    journal = {Ecological Applications},
[…]
