### 2025-06-23 8f79ecb
remove extra comma in joss paper summary
