### 2025-06-24 24a278c
re-arrange citation file, add comments
