+++ CITATION.cff
@@ -1,9 +1,7 @@
 # cff-version: 1.2.0
-title: "DVMDOSTEM"
+title: "DVMDOSTEM v0.8.3: a terrestrial ecosystem model designed to represent arctic, boreal and permafrost ecosystem dynamics"
 
-message: >-
-  If you use this software, please cite it using the
-  metadata from this file.
+message: "If you use this software, please cite using the metadata in this file."
 type: software
 identifiers:
   - type: doi
@@ -30,55 +28,53 @@ keywords:
   - Tundra
 license: MIT
 authors:
-  - family-names: "Euskirchen"
-    given-names: "Eugénie"
-    orcid: "https://orcid.org/0000-0002-0848-4295"
-    affiliation: "Institute of Arctic Biology, University of Alaska Fairbanks, Fairbanks 99775, Alaska, USA"
-
-  - family-names: "Genet"
+  - family-names: "Genet" # Wildfire, scientific direction, 
     given-names: "Hélène"
     orcid: "https://orcid.org/0000-0003-4537-9563"
     affiliation: "Institute of Arctic Biology, University of Alaska Fairbanks, Fairbanks 99775, Alaska, USA"
 
-  - family-names: "McGuire"
+  - family-names: "Euskirchen" # DVM, scientific direction
+    given-names: "Eugénie"
+    orcid: "https://orcid.org/0000-0002-0848-4295"
+    affiliation: "Institute of Arctic Biology, University of Alaska Fairbanks, Fairbanks 99775, Alaska, USA"
+
+  - family-names: "McGuire" # TEM, scientific direction
     given-names: "A. D."
     orcid: "https://orcid.org/0000-0003-4646-0750"
     affiliation: "Institute of Arctic Biology, University of Alaska Fairbanks, Fairbanks 99775, Alaska, USA"
 
-  - family-names: "Carman"
+  - family-names: "Carman" # Software development and maintenance
     given-names: "Tobey B"
     orcid: "https://orcid.org/0000-0003-4617-4674"
     affiliation: "Institute of Arctic Biology, University of Alaska Fairbanks, Fairbanks 99775, Alaska, USA"
 
-  - family-names: "Rutter"
+  - family-names: "Rutter" # Software development and maintenance
     given-names: "Ruth A"
     orcid: "https://orcid.org/0009-0009-7043-6081"
     affiliation: "Institute of Arctic Biology, University of Alaska Fairbanks, Fairbanks 99775, Alaska, USA"
 
-  - family-names: "Maglio"
-    given-names: "Ben"
-    orcid: "https://orcid.org/0000-0002-1948-0177"
-    affiliation: "Institute of Arctic Biology, University of Alaska Fairbanks, Fairbanks 99775, Alaska, USA"
-
-  - family-names: "Clein"
-    given-names: "Joy"
+  - family-names: "Clein" # Calibration
     orcid: "https://orcid.org/0000-0002-2816-5312"
     affiliation: "Institute of Arctic Biology, University of Alaska Fairbanks, Fairbanks 99775, Alaska, USA"
 
-  - family-names: "Yuan"
+  - family-names: "Yuan" # Merging DVM, DOS, and TEM
     given-names: "Fengming"
     orcid: "https://orcid.org/0000-0003-0910-5231"
     affiliation: "Environmental Science Division and Climate Change Science Institute, Oak Ridge National Laboratory, Oak Ridge, TN 37831"
 
-  - family-names: "Yi"
+  - family-names: "Jafarov" # Calibration
+    given-names: "Elchin"
+    affiliation: "Woodwell Climate Research Center, Falmouth 02540, Massachusetts, USA"
[…]
