### 2025-06-24 4fb4e93
Update repo .cff file to match JOSS paper author list and title
