+++ CITATION.cff
@@ -1,5 +1,5 @@
 # cff-version: 1.2.0
-title: "DVMDOSTEM"
+title: "DVMDOSTEM v0.8.3: a terrestrial ecosystem model designed to represent arctic, boreal and permafrost ecosystem dynamics"
 
 message: >-
   If you use this software, please cite it using the
@@ -80,5 +80,58 @@ authors:
     orcid: "https://orcid.org/0000-0002-4536-9851"
     affiliation: "Department of Earth, Atmospheric, and Planetary Science, Purdue University, West Lafayette IN 47907 USA"
 
+  - family-names: "Jafarov"
+    given-names: "Elchin"
+    affiliation: "Woodwell Climate Research Center, Falmouth 02540, Massachusetts, USA"
+    orcid: "https://orcid.org/0000-0002-8310-3261"
+
+  - family-names: "Briones"
+    given-names: "Valeria"
+    affiliation: "Woodwell Climate Research Center, Falmouth 02540, Massachusetts, USA"
+    orcid: "https://orcid.org/0000-0002-5649-851X"
+
+  - given-names: "Andrew"
+    family-names: "L Mullen"
+    affiliation: "Woodwell Climate Research Center, Falmouth 02540, Massachusetts, USA"
+    orcid: "https://orcid.org/0000-0002-9127-9996"
+
+  - given-names: "Heather"
+    family-names: "Greaves"
+    affiliation: "Institute of Arctic Biology, University of Alaska Fairbanks, Fairbanks 99775, Alaska, USA"
+    orcid: "https://orcid.org/0000-0002-8800-019X"
+
+  - given-names: "Chu-Chun"
+    family-names: "Chang"
+    affiliation: "Woodwell Climate Research Center, Falmouth 02540, Massachusetts, USA"
+    orcid: "https://orcid.org/0000-0002-2971-2349"
+
+  - given-names: "Joshua M."
+    family-names: "Rady"
+    affiliation: "Woodwell Climate Research Center, Falmouth 02540, Massachusetts, USA"
+    orcid: "https://orcid.org/0000-0002-7806-136X"
+
+  - given-names: "Doğukan"
+    family-names: "Teber"
+    affiliation: "Woodwell Climate Research Center, Falmouth 02540, Massachusetts, USA"
+    orcid: "https://orcid.org/0009-0000-1547-950X"
+
+  - given-names: "Brendan M."
+    family-names: "Rogers"
+    affiliation: "Woodwell Climate Research Center, Falmouth 02540, Massachusetts, USA"
+    orcid: "https://orcid.org/0000-0001-6711-8466"
+
+  - given-names: "Trevor"
+    family-names: "Smith"
+    affiliation: "Woodwell Climate Research Center, Falmouth 02540, Massachusetts, USA"
+
+  - given-names: "Hannah"
+    family-names: "Mevenkamp"
+    affiliation: "Institute of Arctic Biology, University of Alaska Fairbanks, Fairbanks 99775, Alaska, USA"
+    orcid: "https://orcid.org/0000-0002-7241-5374"
+
+  - given-names: "Mark J."
+    family-names: "Lara"
+    affiliation: "Department(s) of Plant Biology and Geography, University of Illinois Urbana-Champaign, Urbana, IL 61801"
+    orcid: "https://orcid.org/0000-0002-4670-7031"
 
 
