### 2025-06-24 a12d2a7
version bump docs, joss paper
