### 2025-06-24 3b29c05
Adjust cff file to use identifiers in a...
...minimally difficult way with respect to updates and correctness.

Ideally we would include the release DOI in the .cff file, but we can't
know this value before the release is made, so we are instead using the
concept DOI here and the GitHub URL to the release, which we can know in
advance. Trying to minimize the manual changes that have to be
remembered when creating a new release.

Updated the release instructions accordingly.
Add cff-version specifier in order to pass cffconvert --validate
