### 2025-06-25 59b0626
remove code formatting from .cff yaml file; not heeded most places
