### 2025-06-25 5581c19
adjust cff author list (name formatting)
