### 2025-06-26 80babd6
Correctly setting thermokarst pointers
Required for thermokarst data and others. This version is
removing the soil organic layers at a hard coded proportion
of ~90%
