### 2025-06-26 #7667 @tobeycarman
@kthyng OK, I fixed the capitalization in the references and the errant comma. I also improved the repo's citation file (which is what feeds into the Zenodo metadata). We decided that we are ok with not having the author and metadata for the repo and the JOSS paper match exactly.

I made another release (v0.8.3) to capture these changes.
