### 2025-06-30 96564ab
Remaining changes following pointer correction
