+++ src/Thermokarst.cpp
@@ -110,12 +110,12 @@ void Thermokarst::initializeState() {
   // this is set from wildfire changes, what state do we
   // need to initialize in the case of a thermokarst
   // disturbance occurrence?
-  td->thermokarst_a2soi.orgn = 0.0;
+  tkdata->thermokarst_a2soi.orgn = 0.0;
 };
 
 // Looks like this is just used when setting up a Cohort from a Restart file...
 void Thermokarst::set_state_from_restartdata(const RestartData & rdata) {
-  td->thermokarst_a2soi.orgn = rdata.thermokarsta2sorgn;
+  tkdata->thermokarst_a2soi.orgn = rdata.thermokarsta2sorgn;
 }
 
 /** Figure out whether or not there should be thermokarst, based on stage, yr, month.
@@ -161,11 +161,11 @@ void Thermokarst::initiate(int year) {
   BOOST_LOG_NAMED_SCOPE("Thermokarsting");
   BOOST_LOG_SEV(glg, info) << "HELP!! - THERMOKARST!! RUN FOR YOUR LIFE!";
 
-  BOOST_LOG_SEV(glg, debug) << td->report_to_string("Before Thermokarst::initiate(..)");
+  BOOST_LOG_SEV(glg, debug) << tkdata->report_to_string("Before Thermokarst::initiate(..)");
   BOOST_LOG_SEV(glg, info) << "Clearing the ThermokarstData object...";
-  td->clearing();
-  BOOST_LOG_SEV(glg, debug) << td->report_to_string("After ThermokarstData::clearing(..)");
-  
+  tkdata->clearing();
+  BOOST_LOG_SEV(glg, debug) << tkdata->report_to_string("After ThermokarstData::clearing(..)");
+
   // What do we want this to be called / do?
   // changing burn to thermokarst, referring to OL soil
   // lost through thermokarst disturbance (either slump or
@@ -175,7 +175,7 @@ void Thermokarst::initiate(int year) {
   // but also how to handle different thermokarst types in the
   // future
 
-  BOOST_LOG_SEV(glg, debug) << td->report_to_string("After Thermokarst::getThermokarstOrgSoilthick(..)");
+  BOOST_LOG_SEV(glg, debug) << tkdata->report_to_string("After Thermokarst::getThermokarstOrgSoilthick(..)");
 
   BOOST_LOG_SEV(glg, info) << "Setup some temporary pools for tracking various thermokarst related attributes (depths, C, N)";
   // Do we need other variables for tracking? ALSO RENAME appropriately
@@ -265,14 +265,14 @@ void Thermokarst::initiate(int year) {
       BOOST_LOG_SEV(glg, info) << "Not much to do here. Can't really thermokarst non-organic layers. but maybe we will in the future";
 
       if(totbotdepth <= thermokarstdepth) { //may not be needed, but just in case
-        BOOST_LOG_SEV(glg, info) << "For some reason totbotdepth <= thermokarstdepth, so we are setting td->thermokarst_soid.removal_thickness = totbotdepth??";
-        td->thermokarst_soid.removal_thickness = totbotdepth;
+        BOOST_LOG_SEV(glg, info) << "For some reason totbotdepth <= thermokarstdepth, so we are setting tkdata->thermokarst_soid.removal_thickness = totbotdepth??";
+        tkdata->thermokarst_soid.removal_thickness = totbotdepth;
       }
     }
   } // end soil layer loop
 
   //Setting relative organic layer removed (rolr) value
[…]
