### 2025-06-30 1d2b671
Remaining changes following pointer correction
