+++ include/Cohort.h
@@ -89,8 +89,8 @@ public :
 
   FirData year_fd[12]; //Monthly fire data, for all PFTs and soil
   FirData* fd;   //Fire data for an individual month
-  ThermokarstData year_td[12]; // Monthly fire data, for all PFTs and soil
-  ThermokarstData *td;         // Fire data for an individual month
+  ThermokarstData year_tkdata[12]; // Monthly fire data, for all PFTs and soil
+  ThermokarstData *tkdata;     // Fire data for an individual month
 
   ModelData* md;
 
