### 2025-07-04 dc18bdd
Adding drainage change and CMT6 hard coding
