### 2025-07-08 #7667 @kthyng
@tobeycarman Are you seeing the capitalization corrected in this proof? I still see "alaska" in Yi et al 2009.
