### 2025-07-08 #7667 @tobeycarman
Also FWIW a co-author would like to add a reference to a recently published paper, so I could go ahead and make yet another release and see if that wakes up the editorial bot???
