### 2025-07-08 #7667 @tobeycarman
@kthyng ahh, ok, I bet the branch is the problem: we had developed the paper in the `joss-paper-2` branch, but that got merged to `master`. We planned to leave the paper in the repo in the "joss-paper" directory - perhaps this is not advisable for some reason? I will admit I am a little confused on the best way to keep the paper and code perfectly related and in sync with the git version. For example if I now make an additional commit to `master` that updates the paper with the added reference, then that commit will be ahead of v0.8.3 which is the archive version...maybe this is simply too small of a detail to be concerned about, but it makes me think I have approached this wrong...
