### 2025-07-08 3faf12e
Add reference to JOSS paper.
