+++ joss-paper/joss-paper.bib
@@ -108,6 +108,17 @@
    volume = {8},
    year = {2013},
 }
+@article{Jafarov2025,
+   author = {Jafarov, E. E. and Genet, H. and Vesselinov, V. V. and Briones, V. and Kabeer, A. and Mullen, A. L. and Maglio, B. and Carman, T. and Rutter, R. and Clein, J. and Chang, C.-C. and Teber, D. and Smith, T. and Rady, J. M. and Sch\"adel, C. and Watts, J. D. and Rogers, B. M. and Natali, S. M.},
+   doi = {10.5194/gmd-18-3857-2025},
+   journal = {Geoscientific Model Development},
+   number = {12},
+   pages = {3857--3875},
+   title = {Estimation of above- and below-ground ecosystem parameters for DVM-DOS-TEM v0.7.0 using MADS v1.7.3},
+   url = {https://gmd.copernicus.org/articles/18/3857/2025/},
+   volume = {18},
+   year = {2025},
+}
 @article{Koven2011,
    abstract = {Permafrost soils contain enormous amounts of organic carbon, which could act as a positive feedback to global climate change due to enhanced respiration rates with warming. We have used a terrestrial ecosystem model that includes permafrost carbon dynamics, inhibition of respiration in frozen soil layers, vertical mixing of soil carbon from surface to permafrost layers, and CH 4 emissions from flooded areas, and which better matches new circumpolar inventories of soil carbon stocks, to explore the potential for carbon-climate feedbacks at high latitudes. Contrary to model results for the Intergovernmental Panel on Climate Change Fourth Assessment Report (IPCC AR4), when permafrost processes are included, terrestrial ecosystems north of 60° N could shift from being a sink to a source of CO 2 by the end of the 21st century when forced by a Special Report on Emissions Scenarios (SRES) A2 climate change scenario. Between 1860 and 2100, the model response to combined CO 2 fertilization and climate change changes from a sink of 68 Pg to a 27 + -7 Pg sink to 4 + -18 Pg source, depending on the processes and parameter values used. The integrated change in carbon due to climate change shifts from near zero, which is within the range of previous model estimates, to a climate-induced loss of carbon by ecosystems in the range of 25 + -3 to 85+ -16 Pg C, depending on processes included in the model, with a best estimate of a 62 + -7 Pg C loss. Methane emissions from high-latitude regions are calculated to increase from 34 Tg CH 4/y to 41-70 TgCH 4/y, with increases due to CO 2 fertilization, permafrost thaw, and warming-induced increased CH 4 flux densities partially offset by a reduction in wetland extent.},
    author = {Charles D. Koven and Bruno Ringeval and Pierre Friedlingstein and Philippe Ciais and Patricia Cadule and Dmitry Khvorostyanov and Gerhard Krinner and Charles Tarnocai},
